### 2020-07-15 #26 @snacktavish
Physcraper: Automated phylogenetic updating
Submitting Author: Emily Jane McTavish @snacktavish
Package Name: Physcraper
One-Line Description of Package: Automated phylogenetic updating
Repository Link:  https://github.com/McTavishLab/physcraper
Version submitted:   v0.55
Archive: https://zenodo.org/record/5512990
Version accepted: 1.0

---

Physcraper is an open‐source, command-line Python program that automatizes the update of published
phylogenies by enriching underlying gene alignments with public DNA sequence data, and linking taxonomic information across databases.

- Please indicate which [category or categories](https://www.pyopensci.org/dev_guide/peer_review/aims_scope.html) this package falls under:

\* Please fill out a pre-submission inquiry before submitting a data visualization package. For more info, see [this section](https://www.pyopensci.org/dev_guide/peer_review/aims_scope.html#notes-on-categories) of our guidebook.

- Explain how the and why the package falls under these categories (briefly, 1-2 sentences):

This package retrieves data from GenBank (https://www.ncbi.nlm.nih.gov/genbank/), the OpenTreeofLife (www.opentreeoflife.org), and TreeBase (www.treebase.org/), and links these datasets together to incorporate new sequence data into evolutionary estimates. Physcraper also provides a framework for comparison of published phylogenies with their updated versions, by using the conflict API of the Open Tree of Life project.

Physcraper can be used by the nonspecialist, as a tool to generate phylogenetic hypotheses based on already available expert phylogenetic knowledge.
Phylogeneticists and taxonomic group specialists will find it useful as a tool to facilitate molecular dataset gathering and comparison of alternative phylogenetic hypotheses.

The closest analogue in python is PyPhlawd https://fephyfofum.github.io/PyPHLAWD/.
This approach differs from other tools (PyPhlawd, Supersmart) automatize the assembly of DNA alignments from the GenBank database for phylogenetic reconstruction,  in that Physcraper leverages existing homology statements in the form of alignments and existing phylogenetic trees to select appropriate loci and improve inference.
Another key feature of Physcraper is linking of taxonomic identifiers across databases including NCBI, GBIF and OpenTree of Life which provides easy interoperability with those resources, and the ability to easily compare relationships between new inferences and existing taxonomy and previously published studies.

No pre-submission inquiry

For details about the pyOpenSci packaging requirements, see our [packaging guide](https://www.pyopensci.org/dev_guide/packaging/packaging_guide.html). Confirm each of the following by checking the box.  This package:

- [x] has an [OSI approved license](https://opensource.org/licenses)

- [ ] The package has an **obvious research application** according to JOSS's definition in their [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements). Be aware that completing the pyOpenSci review process **does not** guarantee acceptance to JOSS. Be sure to read their submission requirements (linked above) if you are interested in submitting to JOSS.
- [ ] The package is not a "minor utility" as defined by JOSS's [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements): "Minor ‘utility’ packages, including ‘thin’ API clients, are not acceptable." pyOpenSci welcomes these packages under "Data Retrieval", but JOSS has slightly different criteria.

- [x] I agree to abide by [pyOpenSci's Code of Conduct](https://www.pyopensci.org/contributing-guide/peer_review/coc.html) during the review process and in maintaining my package should it be accepted.

**P.S.** *Have feedback/comments about our review process? Leave a comment [here](https://github.com/pyOpenSci/governance/issues/8)*

[Editor and review templates can be found here](https://www.pyopensci.org/contributing-guide/appendices/templates.html)
