+++ README.md
@@ -1,7 +1,7 @@
 
 <!-- README.md is generated from README.Rmd; please edit the .Rmd file -->
 
-<img align="left" width="260" src="https://raw.githubusercontent.com/McTavishLab/physcraper/main/docs/physcraper.svg">
+<img align="left" width="250" src="https://raw.githubusercontent.com/McTavishLab/physcraper/main/docs/physcraper.svg">
 
 # Physcraper
 
@@ -32,17 +32,10 @@ and tools for developers, please refer to Physcraper’s [documentation
 website](https://physcraper.readthedocs.io/en/latest/) for more
 details\!
 
-Installation instuctions at [INSTALL](docs/mds/INSTALL.md)
-
 :hamster: :palm\_tree: :frog: :ear\_of\_rice: :panda\_face: :tulip:
 :octopus: :blossom: :whale: :mushroom: :ant: :cactus: :fish:
 :maple\_leaf: :water\_buffalo: 🦠 :shell: :bug: :octocat:
 
-`Physcraper` relies on:
-
-Dendropy: *Sukumaran, J and MT Holder. 2010.
+Physcraper relies on Dendropy: *Sukumaran, J and MT Holder. 2010.
 DendroPy: a Python library for phylogenetic computing. Bioinformatics
 26: 1569-1571*.
-
-[The Open Tree of Life Project](oentreeoflife.org):
-*Open Tree of Life, Benjamin Redelings, Luna Luisa Sanchez Reyes, Karen A. Cranston, Jim Allman, Mark T. Holder, & Emily Jane McTavish. (2019). Open Tree of Life Synthetic Tree (Version 12.3). Zenodo. [https://doi.org/10.5281/zenodo.3937741](https://doi.org/10.5281/zenodo.3937741)*
