+++ docs/mds/intro-dendropy.md
@@ -1,1 +1,7 @@
-Physcraper relies on Dendropy: *Sukumaran, J and MT Holder. 2010. DendroPy: a Python library for phylogenetic computing. Bioinformatics 26: 1569-1571*.
+Physcraper relies on
+
+- Dendropy:
+ *Sukumaran, J and MT Holder. 2010. DendroPy: a Python library for phylogenetic computing. Bioinformatics 26: 1569-1571*.
+
+- [The Open Tree of Life Project](oentreeoflife.org):
+ *Open Tree of Life, Benjamin Redelings, Luna Luisa Sanchez Reyes, Karen A. Cranston, Jim Allman, Mark T. Holder, & Emily Jane McTavish. (2019). Open Tree of Life Synthetic Tree (Version 12.3). Zenodo. [https://doi.org/10.5281/zenodo.3937741](https://doi.org/10.5281/zenodo.3937741)*
