+++ README.Rmd
@@ -35,8 +35,13 @@ IS_README <- TRUE
 
 
 This is the code repository.
-For an introduction to the tool, installation and function usage instructions, tutorials and examples, and tools for developers,
-please refer to Physcraper's [documentation website](https://physcraper.readthedocs.io/en/latest/) for more details!
+Please refer to Physcraper's [documentation website](https://physcraper.readthedocs.io/en/latest/) for more details
+on [installation instructions](https://physcraper.readthedocs.io/en/latest/install.html),
+[function usage](https://physcraper.readthedocs.io/en/latest/apidocs.html),
+[tutorials](https://physcraper.readthedocs.io/en/latest/physcraper_run.html),
+[examples](https://physcraper.readthedocs.io/en/latest/data_exploration.html),
+and [tools for developers](https://physcraper.readthedocs.io/en/latest/CONTRIBUTING.html).
+
 
 
 :hamster: :palm_tree: :frog: :ear_of_rice: :panda_face: :tulip: :octopus: :blossom: :whale: :mushroom: :ant: :cactus: :fish: :maple_leaf: :water_buffalo: 🦠 :shell: :bug: :octocat:
