+++ docs/mds/Tutorial.md
@@ -1,6 +1,12 @@
+<br/>
+
 ## The Physcraper framework
 
+<br/>
+
 ![](../img/schematic.svg)
+
+<br/>
 
 If you have access to a single gene alignment, and a tree, you can automate adding homologous data into your tree by searching GenBank.
 
@@ -14,12 +20,14 @@ The sequences added in the search are limited to a user specified taxon or monop
 
 These automated tree can provide a quick inference or potential relationships, of problems in the taxonomic assignments of sequences, and flag areas of potential systematic interest.
 
+<br/>
+
 
 ## The Open Tree of Life
 
 The Open Tree of Life (https://opentreeoflife.github.io/) is a project that unites phylogenetic inferences and taxonomy to provide a synthetic estimate of species relationships across the entire tree of life.
-![](../img/otol_logo.png)
 
+![](../img/otol_logo.png)
 
 Open Tree of Life aims to construct a comprehensive, dynamic and digitally-available tree of life by synthesizing published phylogenetic trees along with taxonomic data.
 Currently the tree comprises 2.3 million tips.
@@ -27,9 +35,10 @@ However, only around 90,000 of those taxa are represented by phylogenetic estima
 
 ![](../img/synthtreeleg.svg)
 
-(figure from [Hinchliff et al. 2015](https://www.pnas.org/content/112/41/12764.short))
-For more information on OpenTree see https://opentreeoflife.github.io
+Figure from [Hinchliff et al. 2015](https://www.pnas.org/content/112/41/12764.short).
+For more information on the Open Tree of Life project see https://opentreeoflife.github.io
 
+<br/>
 
 ## Updating a tree from Open Tree of Life
 
@@ -39,33 +48,32 @@ This is a great place to start of finding existing estimates of phylogenetic rel
 and assessing regions of the tree of life which are lacking available phylogenetic estimates.
 There is a lot of sequence data available that has never been incorporated into any phylogenetic estimates.
 
-
 ### *Find a starting tree with your taxon of interest*
 
+For this example we'll use a tree that is already in the Open Tree of Life database. You can find more details about finding a tree to update at the section [Finding Data](https://physcraper.readthedocs.io/en/latest/find_trees.html).
 
-For this example we'll use find a tree that is already in the OpenTree of Life database. (More search information at [Finding Data](https://physcraper.readthedocs.io/en/latest/find_trees.html).
-Search on OpenTree of life using your taxon of interest, e.g. 'Malvaceae'
+To find trees containing your taxon of interest (e.g. 'Malvaceae') on OpenTree of life use:
 
     $ find_trees.py --taxon_name "Malvaceae"
 
-This prints a bunch of studies out to the screen. We will need an alignment to update (which OpenTree doesn't store), so lets just look at trees that have data stored in tree base.
+This prints a bunch of studies out to the screen. We will need an alignment to update (which OpenTree doesn't store), so lets just look at trees that have data stored in TreeBase.
 
     $ find_trees.py --taxon_name "Malvaceae" --treebase
 
 There are a bunch of options!
 
-Lets update Wilkie et al (2006).
-You can view the study on the OpenTree database: [Wilkie2006](https://tree.opentreeoflife.org/curator/study/view/pg_55)
+Lets update the Wilkie et al. ([2006](https://doi.org/10.1600/036364406775971714)) study.
+You can view the study on the OpenTree database at [Wilkie, 2006](https://tree.opentreeoflife.org/curator/study/view/pg_55)
 
-While this study was focussed on the family "Sterculiacea",
+While this study was focussed on the family Sterculiacea,
 phylogenetic inference have suggested that this family is not [monophyletic]((https://tree.opentreeoflife.org/opentree/argus/ottol@996482))
 
[…]
