### 2020-07-21 055c0ff
fix write info bug
