+++ physcraper/scrape.py
@@ -126,7 +126,7 @@ class PhyscraperScrape(object):
         self.repeat = 1  # used to determine if we continue updating the tree
         self.newseqs_acc = []  # all ever added Genbank accession numbers during any PhyScraper run, used to speed up adding process
         self.seq_filter = ['deleted', 'subsequence,', 'not', "removed", "deleted,",
-                           "local"] 
+                           "local"]
         self.reset_markers()
         self.gb_not_added = []  # list of blast seqs not added
         self.del_superseq = set()  # items that were deleted bc they are superseqs, needed for assert statement
@@ -246,7 +246,7 @@ class PhyscraperScrape(object):
         result_handle.close()
         save_file.close()
 
-    def run_blast_wrapper(self): 
+    def run_blast_wrapper(self):
         """generates the blast queries and saves them depending on the blasting method to different file formats
 
         It runs blast if the sequences was not blasted since the user defined threshold in the config file (delay).
@@ -330,7 +330,7 @@ class PhyscraperScrape(object):
                 if gb_acc == None:
                     continue
                 gi_id = get_gi_from_blast(sseqid)
-                sseq = sseq.replace("-", "") 
+                sseq = sseq.replace("-", "")
                 taxname = sscinames.replace(" ", "_").replace("/", "_")
                 pident = float(pident)
                 evalue = float(evalue)
@@ -399,7 +399,7 @@ class PhyscraperScrape(object):
                                         "-entry", gb_id,
                                         "-outfmt", "%f",
                                         "-out", seq_path])
-        
+
             except subprocess.CalledProcessError as grepexc:
                 sys.stderr.write("error code {}, {}".format(grepexc.returncode, grepexc.output))
                 sys.exit()
@@ -878,7 +878,7 @@ class PhyscraperScrape(object):
                    taxon_namespace = self.data.aln.taxon_namespace)
         try:
             rooted_tre = root_tree_from_synth(newtre, self.data.otu_dict)
-        except: 
+        except:
             sys.stderr.write("Tree not rooted, root before running conflict analyses.\n")
         self.data.tre = newtre
 
@@ -1017,7 +1017,7 @@ class PhyscraperScrape(object):
         sumtree_file = open(summarized_tree_path, 'w')
         subprocess.check_call(["sumtrees.py",
                                 "-t", besttreepath,
-                                "-f", str(min_clade_freq), 
+                                "-f", str(min_clade_freq),
                                 "-d0",
                                 bootpath], stdout=sumtree_file)
         sumtree_file.close()
