+++ physcraper/ncbi_data_parser.py
@@ -16,12 +16,8 @@ _DEBUG = 0
 def get_acc_from_blast(query_string):
     """
     Get the accession number from a blast query.
-    
-    Get acc is more difficult now, as new seqs not always have gi number, then query changes.
-    
     :param query_string: string that contains acc and gi from local blast query result
-    :return: gb_acc 
-
+    :return: gb_acc
     """
     if len(query_string.split("|")) >= 3:
         gb_acc = query_string.split("|")[3]
@@ -35,14 +31,13 @@ def get_acc_from_blast(query_string):
 
 def get_gi_from_blast(query_string):
     """
-    Get the gi number from a blast query. 
+    Get the gi number from a blast query.
     Get acc is more difficult now, as new seqs not always have gi number, then query changes.
-
-    If not available return None. 
+    If not available return None.
 
     :param query_string: string that contains acc and gi from local blast query result
     :return: gb_id if available
-    """      
+    """
     if len(query_string.split("|")) >= 3:
         gb_id = query_string.split("|")[1]
     else:
@@ -51,33 +46,10 @@ def get_gi_from_blast(query_string):
     assert gb_id.isdigit() is True
 
 def get_tax_info_from_acc(gb_id, data_obj, ids_obj):
-    '''takes an accessionumber and returns the ncabi_id and the taxon name'''
+    '''takes an accession number and returns the ncbi_id and the taxon name'''
 #    debug("Getting tax info from acc {}".format(gb_id))
-    ncbi_id = None
-    tax_name = None
-    if gb_id[:6] == "unpubl":  # There may not be ncbi id, because they aren't published
-            tax_name = data_obj.gb_dict[gb_id]["^ot:ottTaxonName"]
-            ncbi_id = data_obj.gb_dict[gb_id]["^ncbi:taxon"]
-            ott_id = data_obj.gb_dict[gb_id]["^ot:ottId"]
-            if tax_name is None:
-                tax_name = data_obj.gb_dict[gb_id][u'^user:TaxonName']
-            if ncbi_id is None: 
-                # debug(tax_name.split(" ")[0])
-                tax_lin_name = tax_name.split(" ")[0]
-                tax_lin_name = tax_lin_name.split("_")[0]
-                # debug(tax_lin_name)
-                ncbi_id = ids_obj.ncbi_parser.get_id_from_name(tax_lin_name) #TODO What should happen here if the unpublished sequence doesn't have a name that is found?
-  #  elif len(gb_id.split(".")) >= 2:  # used to figure out if gb_id is from Genbank
-  #          if gb_id in data_obj.gb_dict.keys() and "staxids" in data_obj.gb_dict[gb_id].keys():
-  #              tax_name = data_obj.gb_dict[gb_id]["sscinames"]
-  #              ncbi_id = data_obj.gb_dict[gb_id]["staxids"]
-  #          else:  # all web blast results
-  #              if tax_name is None:
-  #                  sys.stderr.write("no species name returned for {}\n".format(gb_id))
-  #              ncbi_id = ids_obj.get_ncbiid_from_acc(gb_id)
-    else:
-        ncbi_id = ids_obj.get_ncbiid_from_acc(gb_id)
-        tax_name = ids_obj.ncbiid_to_spn.get(ncbi_id)
+    ncbi_id = ids_obj.get_ncbiid_from_acc(gb_id)
+    tax_name = ids_obj.ncbiid_to_spn.get(ncbi_id)
     if ncbi_id == None:
         sys.stderr.write("Failed to get information for sequence with accession number {}".format(gb_id))
     return ncbi_id, tax_name
@@ -103,7 +75,7 @@ def get_ncbi_tax_id(handle):
 
[…]
