+++ physcraper/configobj.py
@@ -1,3 +1,4 @@
+""" Physcraper run Configuration object generator"""
 import sys
 import os
 import datetime
@@ -5,45 +6,40 @@ import configparser
 import shutil
 import wget
 
-from physcraper import ncbi_data_parser  # is the ncbi data parser class and associated functions
-
-from physcraper.helpers import cd, debug
 
 _DEBUG = 0
 
 
 
 
-physcraper_dir = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
-#sys.stdout.write(physcraper_dir)
+PHYSCRAPER_DIR = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
+#sys.stdout.write(PHYSCRAPER_DIR)
 
 
-def is_number(s):
+def is_number(inputstr):
     """test if string can be coerced to float"""
     try:
-        float(s)
+        float(inputstr)
         return True
     except ValueError:
         return False
 
 
 
-class ConfigObj(object):
+class ConfigObj():
     """
     To build the class the following is needed:
 
       * **configfi**: a configuration file in a specific format, e.g. to read in self.e_value_thresh.
-                
-        The file needs to have a heading of the format: [blast] and then somewhere below that heading a string e_value_thresh = value
-
-      * **interactive**: defaults to True, is used to interactively update the local blast databases
 
     During the initializing process the following self objects are generated:
 
-      * **self.e_value_thresh**: the defined threshold for the e-value during Blast searches, check out: https://blast.ncbi.nlm.nih.gov/Blast.cgi?CMD=Web&PAGE_TYPE=BlastDocs&DOC_TYPE=FAQ
+      * **self.e_value_thresh**: the defined threshold for the e-value during Blast searches,
+            check out: https://blast.ncbi.nlm.nih.gov/Blast.cgi?CMD=Web&PAGE_TYPE=BlastDocs&DOC_TYPE=FAQ
       * **self.hitlist_size**: the maximum number of sequences retrieved by a single blast search
       * **self.minlen**: value from 0 to 1. Defines how much shorter new seq can be compared to input
-      * **self.trim_perc**: value that determines how many seq need to be present before the beginning and end of alignment will be trimmed
+      * **self.trim_perc**: value that determines how many seq need to be present before the beginning
+            and end of alignment will be trimmed
       * **self.maxlen**: max length for values to add to aln
       * **self.get_ncbi_taxonomy**: Path to sh file doing something...
       * **self.ott_ncbi**: file containing OTT id, ncbi and taxon name (??)
@@ -53,7 +49,7 @@ class ConfigObj(object):
           * either web-query (=remote)
           * from a local blast database (=local)
       * **self.num_threads**: number of cores to be used during a run
-      * **self.url_base**: 
+      * **self.url_base**:
 
           * if blastloc == remote: it defines the url for the blast queries.
           * if blastloc == local: url_base = None
@@ -67,17 +63,19 @@ class ConfigObj(object):
               * self.ncbi_names: path to 'names.dmp' file, that contains the different ID's
     """
 
-    def __init__(self, configfile = None, interactive=False):
+    def __init__(self, configfile=None):
        # debug(configfi)
         if _DEBUG:
             sys.stdout.write("Building config object\n")
         if configfile:
             self.set_defaults()
-            self.read_config(configfile, interactive)
+            self.read_config(configfile)
         else:
             sys.stdout.write("No config file, using defaults\n")
             self.set_defaults()
     def set_defaults(self):
+        """ In the absence of an input configuration file, sets default values.
+        """
         self.email = None
         self.e_value_thresh = 0.00001
         self.hitlist_size = 10
@@ -93,7 +91,10 @@ class ConfigObj(object):
[…]
