+++ physcraper/aligntreetax.py
@@ -1,3 +1,4 @@
+"""AlignTreeTax: The core data object for physcraper. Holds and links name stpaces for a tree, an alignment and the taxa adn their metadata"""
 import sys
 import re
 import os
@@ -13,6 +14,7 @@ from physcraper.helpers import standardize_label, to_string, debug
 _VERBOSE = 0
 
 def set_verbose():
+    """Set verbosity of outputs"""
     global _VERBOSE
     _VERBOSE = 1
 
@@ -37,9 +39,11 @@ def generate_ATT_from_files(workdir,
     :param workdir: path to working directory
     :param config_obj: config class including the settings
     :param treefile: path to phylogeny
-    :param otu_json: path to json file containing the translation of tip names to taxon names, generated with OtuJsonDict()
+    :param otu_json: path to json file containing the translation of tip names to taxon names,
+        or to an otu_dictionary
     :param tree_schema: a string defining the format of the input phylogeny
-    :param search_taxon: optional - OToL ID of the mrca of the clade of interest. If no search mrca ott_id is provided, will use all taxa in tree to calc mrca.
+    :param search_taxon: optional - OToL ID of the mrca of the clade of interest.
+        If no search mrca ott_id is provided, will use all taxa in tree to calc mrca.
 
     :return: object of class ATT
     """
@@ -50,7 +54,7 @@ def generate_ATT_from_files(workdir,
         os.makedirs(workdir)
     # use replaced aln as input
     if isinstance(otu_json, dict):
-            otu_dict = otu_json
+        otu_dict = otu_json
     elif isinstance(otu_json, str):
         assert os.path.exists(otu_json)
         with open(otu_json) as data_file:
@@ -67,7 +71,7 @@ def generate_ATT_from_files(workdir,
         ott_ids = set(ott_ids)
         mrca_ott = get_mrca_ott(ott_ids)
     return AlignTreeTax(treefile, otu_dict, alnfile, search_taxon=mrca_ott, workdir=workdir,
-                        configfile=configfile, tree_schema=tree_schema)
+                        configfile=configfile, aln_schema=aln_schema, tree_schema=tree_schema)
 
 def generate_ATT_from_run(workdir, start_files='output', tag=None, configfile=None):
     """Build an ATT object without phylesystem, use your own files instead.
@@ -81,35 +85,46 @@ def generate_ATT_from_run(workdir, start_files='output', tag=None, configfile=No
     rundir = "{}/run_{}".format(workdir, tag)
     outputsdir = "{}/outputs_{}".format(workdir, tag)
     inputsdir = "{}/inputs_{}".format(workdir, tag)
-    if configfile == None:
-       configfile = "{}/run.config".format(rundir)
+    if configfile is None:
+        configfile = "{}/run.config".format(rundir)
+    if start_files == 'output':
+        sdir = outputsdir
+    if start_files == 'input':
+        sdir = inputsdir
     try:
-        alnfi = "{}/physcraper_{}.fas".format(outputsdir, tag)
-        treefile = "{}/physcraper_{}.tre".format(outputsdir, tag)
+        alnfi = "{}/physcraper_{}.fas".format(sdir, tag)
+        treefile = "{}/physcraper_{}.tre".format(sdir, tag)
         otu_json = "{}/otu_info_{}.json".format(rundir, tag)
-        assert(os.path.exists(alnfi))
-        assert(os.path.exists(treefile))
-        assert(os.path.exists(otu_json))
-        otu_dict = json.load(open(otu_json, "r"))
-        mrca_ott = mrca_ott = int(open("{}/mrca.txt".format(inputsdir)).readline().split()[-1])
[…]
