+++ tests/test_multi.py
@@ -279,7 +279,7 @@ def test_write_labelled():
     os.remove(alnpath_ottid)
 
     data_obj.write_otus(schema='table', direc='tests/data/tmp/')
-    data_obj.write_otus(schema='json', direc='tests/data/tmp/') 
+    data_obj.write_otus(schema='json', direc='tests/data/tmp/')
     assert os.path.isfile('tests/data/tmp/otu_info_test.json')
     assert os.path.isfile('tests/data/tmp/otu_info_test.csv')
     os.remove('tests/data/tmp/otu_info_test.json')
