+++ docs/mds/PhyscraperRun.md
@@ -12,14 +12,22 @@ e.g.
 
 
 ```
-physcraper_run.py -s pg_55 -t tree5864 -tb -o pg55_treebase
+physcraper_run.py -s pg_55 -t tree5864 -tb -o pg_55_web
 ```
+
+The output files generated by this run are stored in docs/examples/pg_55_web
 
 Alternately, you can provide the gene alignment that you want to be updated using the `-a` command:
 
 ```
 physcraper_run_py -s OPENTREE_STUDY_ID -t OPENTREE_TREE_ID -o OUTPUT/DIRECTORY/NAME/AND/OR/PATH -a PATH/TO/GENE/ALIGNMENT/NAME -as ALIGNMENT_SCHEMA
 
+```
+
+The outputs from this command are avialable
+
+```
+physcraper_run.py -s pg_55 -t tree5864 -tb -o pg55_treebase
 ```
 
 
@@ -189,12 +197,22 @@ You can use your own blast database, for example set up on an AWS server.
 
 The simplest (but slowest) run is to choose a tree from opentree, and `physcraper` gets the alignment for you from treebase (argument `-tb`), using web blast:
 
-    physcraper_run.py -s pg_55 -t tree5864 -tb -o pg55_treebase
+```
+  physcraper_run.py -s pg_55 -t tree5864 -tb --email yourname@email.com -o pg_55_web
+```
+
+The output files generated by this run are stored in docs/examples/pg_55_web
 
 
 The fastest way is to choose a tree from opentree, give the path to the corresponding downloaded alignment (argument `-a`) and a local blast directory (argument `-db`). To set up the local blast DB see [Local DB](setting_up_local_database):
 
     physcraper_run.py -s ot_350 -t Tr53297 -a docs/examples/inputdata/ot_350Tr53297.aln -as "nexus" -db ~/ncbi/localblastdb/ -o ot_350
+
+or
+
+    physcraper_run.py -s pg_55 -t tree5864 -a docs/examples/inputdata/pg_55tree5864_ndhf.aln -as nexus --email yourname@email.com -db ~/ncbi/localblastdb -o pg_55_local
+
+The full outputs from this command are stored in docs/examples/pg_55_local
 
 
 To check tree download and the matching of names across tree and alignment without running the blast and tree estimation steps, use the flag (-no_est):
