+++ tests/test_fromfile.py
@@ -1,6 +1,6 @@
 
 import sys
-from physcraper import generate_ATT_from_files, generate_ATT_from_run
+from physcraper import IdDicts, PhyscraperScrape, generate_ATT_from_files, generate_ATT_from_run
 
 
 def test_generate_ATT_from_files():
@@ -29,6 +29,16 @@ def test_generate_ATT_from_files():
 def test_generate_ATT_from_run():
     workdir="tests/data/precooked/output"
 
-    sys.stdout.write("\nTesting 'generate_ATT_from_run (fromfile.py)'\n")
+    sys.stdout.write("\nTesting 'generate_ATT_from_run '\n")
     data_obj = generate_ATT_from_run(workdir=workdir)
       
+
+def test_example():
+    indir="docs/examples/pg_55_web"
+    workdir = "tests/tmp/example_test"
+    data_obj = generate_ATT_from_run(workdir=indir, start_files = "input")
+    data_obj.workdir = workdir
+    ids = IdDicts(data_obj.config)
+    scraper = PhyscraperScrape(data_obj, ids)
+    scraper.read_blast_wrapper("docs/examples/pg_55_web/blast_run_pg_55tree5864/")
+    assert(len(scraper.new_seqs_otu_id) == 25)
