+++ physcraper/aligntreetax.py
@@ -106,8 +106,8 @@ def generate_ATT_from_run(workdir, start_files='output', tag=None, configfile=No
             try:
                 sys.stdout.write("No output files found in {}, loading files from {}\n".format(outputsdir, inputsdir))
                 alnfi = "{}/physcraper_{}.fas".format(inputsdir, tag)
-                treefile = "{}/physcraper_{}.tre".format(sdir, tag)
-                otu_json = "{}/otu_info_{}.json".format(rundir, tag)
+                treefile = "{}/physcraper_{}.tre".format(inputsdir, tag)
+                otu_json = "{}/otu_info_{}.json".format(inputsdir, tag)
                 assert os.path.exists(alnfi)
                 assert os.path.exists(treefile)
                 assert os.path.exists(otu_json)
