+++ physcraper/aligntreetax.py
@@ -1,15 +1,17 @@
-"""AlignTreeTax: The core data object for physcraper. Holds and links name stpaces for a tree, an alignment and the taxa adn their metadata"""
+"""AlignTreeTax: The core data object for physcraper.
+Holds and links name spaces for a tree, an alignment and
+the taxa and their metadata"""
+
 import sys
 import re
 import os
 import json
-import pickle
 import shutil
 import dendropy
-from dendropy import Tree, DnaCharacterMatrix, DataSet, datamodel
+from dendropy import Tree, DnaCharacterMatrix, datamodel
 from physcraper import ncbi_data_parser, ConfigObj
 from physcraper.opentree_helpers import get_mrca_ott, bulk_tnrs_load
-from physcraper.helpers import standardize_label, to_string, debug
+from physcraper.helpers import to_string, debug
 
 _VERBOSE = 0
 
@@ -102,19 +104,17 @@ def generate_ATT_from_run(workdir, start_files='output', tag=None, configfile=No
         if start_files == 'input':
             sys.stdout.write("No run files found in {}. Data not loaded\n".format(inputsdir))
             return None
-        else:
-            try:
-                sys.stdout.write("No output files found in {}, loading files from {}\n".format(outputsdir, inputsdir))
-                alnfi = "{}/physcraper_{}.fas".format(inputsdir, tag)
-                treefile = "{}/physcraper_{}.tre".format(inputsdir, tag)
-                otu_json = "{}/otu_info_{}.json".format(inputsdir, tag)
-                assert os.path.exists(alnfi)
-                assert os.path.exists(treefile)
-                assert os.path.exists(otu_json)
-            except AssertionError:
-                sys.stdout.write("No run files found in {}. Data not loaded\n".format(inputsdir))
-                return None
-
+        try:
+            sys.stdout.write("No output files found in {}, loading files from {}\n".format(outputsdir, inputsdir))
+            alnfi = "{}/physcraper_{}.fas".format(inputsdir, tag)
+            treefile = "{}/physcraper_{}.tre".format(inputsdir, tag)
+            otu_json = "{}/otu_info_{}.json".format(inputsdir, tag)
+            assert os.path.exists(alnfi)
+            assert os.path.exists(treefile)
+            assert os.path.exists(otu_json)
+        except AssertionError:
+            sys.stdout.write("No run files found in {}. Data not loaded\n".format(inputsdir))
+            return None
     otu_dict = json.load(open(otu_json, "r"))
     mrca_ott = int(open("{}/mrca.txt".format(inputsdir)).readline().split()[-1])
     return AlignTreeTax(tree=treefile,
@@ -226,6 +226,7 @@ def write_labelled_aln(aligntreetax, label, filepath, schema="fasta", norepeats=
     tmp_aln = DnaCharacterMatrix.get(data=tmp_fasta,
                                      schema="fasta")
     for taxon in tmp_aln.taxon_namespace:
+        assert(taxon.label in aligntreetax.otu_dict), "{} not in {}".format(taxon.label, aligntreetax.otu_dict.keys())
         new_label = aligntreetax.otu_dict[taxon.label].get(label, None)
         if new_label is None:
             new_label = taxon.label
@@ -243,7 +244,7 @@ def write_labelled_aln(aligntreetax, label, filepath, schema="fasta", norepeats=
                       schema=schema)
 
 
-class AlignTreeTax(object):
+class AlignTreeTax():
     """wrap up the key parts together, requires OTT_id, and names must already match.
         Hypothetically, all the keys in the  otu_dict should be clean.
 
@@ -321,9 +322,9 @@ class AlignTreeTax(object):
     """
 
     def __init__(self, tree, otu_dict, alignment, search_taxon, workdir, configfile=None,
[…]
