### 2020-07-23 226b3d9
fix bug introducted in writeinfo bugfix
