### 2020-07-27 e2cf02b
fix no ingroup inputs
