### 2020-08-27 93aa064
extended methods
