+++ docs/source/methods_extended.rst
@@ -1,0 +1,2 @@
+
+.. mdinclude:: ../mds/methods_extended.md
