### 2020-08-28 abb82f1
fixing bullets
