### 2020-08-28 7cfcb5d
fixing bullets
