+++ docs/source/methods_extended.rst
@@ -1,2 +1,3 @@
-
+Framework
+===========
 .. mdinclude:: ../mds/methods_extended.md
