+++ docs/mds/methods_extended.md
@@ -1,102 +1,75 @@
-# The Physcraper framework
-
-```{r framework, echo=FALSE, fig.cap="The Physcraper framework consists of 4 steps (see text). The software is fully described on its documentation website at physcraper.readthedocs.io, along with installation instructions, function usage descriptions, examples and tutorials.", out.width = '85%', fig.retina= 2, fig.align = "center"}
-# knitr::include_graphics("https://raw.githubusercontent.com/McTavishLab/physcraper/pyopensci/docs/img/schematic.svg")
-knitr::include_graphics("docs/figs/schematic.pdf")
-```
-
-The general Physcraper framework is depicted in Figure \@ref(fig:framework). Briefly, it consists of 4 steps: 1) identifying and processing a phylogenetic tree to update and its underlying alignment; 2) performing a constrained BLAST search of sequences from the original alignment on the GenBank DNA database, and filtering of newly found sequences; 3) profile-aligning new sequences that passed the filtering to the original alignment; 4) performing a phylogenetic analysis and comparing the updated tree to previous phylogenetic estimates in the focus group. Next, we will describe technical details for each step.
-
-## The inputs: a tree and an alignment
-
-In order to take advantage of the OpenTree tools, it is reccommended that the input tree is either stored in the
-  OpenTree [Phylesystem](https://github.com/opentreeoflife/phylesystem), or submitted via OpenTree's curator [application](https://tree.opentreeoflife.org/curator) [@mctavish2015phylesystem].
-  Currently, only trees connected to a published study can be stored in the Phylesystem.
-  Users can choose from among the 2, 950 studies in OpenTree's Phylesystem that have alignments on TreeBASE.
-  If the user is not ready to make the input tree public, tree tip labels can be standardized to the unified OpenTree taxonomy using OpenTree's bulk Taxonomic Name Resolution Service [TNRS](https://tree.opentreeoflife.org/curator/tnrs/) tool.
-  This step is referred to as taxonomic name mapping and Phylesystem stored trees are processed in this way upon submission.
-  Physcraper saves a summary "csv" file with results from the taxon name standardization for advantage of the user, including the mappings to unique identifiers in the OpenTree and NCBI taxonomies.
-  If taxon names can't be mapped, their taxonomic information is not used in comparison analysis. These taxa will still be used in the sequence search and phylogenetic reconstruction steps.
-  Mapping tip names to OpenTree's unified taxonomy saves a set of user defined characteristics
-  that are essential for automatizing the phylogeny updating process. The most relevant of these is the standardized taxonomic names and the definition of ingroup and outgroup taxa, allowing to automatically set the root for the updated tree on the final steps of the pipeline.
-
-The input alignment should be a single locus alignment that was used in part or in whole, to generate the tree. Alignments are often stored in a public repository such as TreeBase [@piel2009treebase; @vos2012nexml],
-  DRYAD ([www.datadryad.org](http://datadryad.org/)), or a data repository associated with the journal where the tree was originally published.
-  If the alignment is stored in TreeBase, Physcraper can download it directly,
-  either from the TreeBASE website ([www.treebase.org](https://treebase.org/))
-  or through the TreeBASE GitHub repository (SuperTreeBASE; [github.com/TreeBASE/supertreebase](https://github.com/TreeBASE/supertreebase)).
-  If the alignment is on another repository, or constitutes personal data, a path to a local copy of the alignment has to be provided.
-
[…]
