+++ docs/source/examples.rst
@@ -1,0 +1,12 @@
+Examples
+===========
+
+|
+
+Interactive illustrated examples are available at package
+at `Physcraperex <https://mctavishlab.github.io/physcraperex/index.html>`_
+for different organisms:
+
+- the hollies
+- the family of chocolate
+- the fungi
