+++ physcraper/scrape.py
@@ -1,32 +1,30 @@
+"""The core blasting and new sequence ntegretaion module"""
 import sys
 import os
 import subprocess
 import datetime
 import glob
-import json
 import pickle
 import random
 
-from copy import deepcopy
 
-from dendropy import Tree, DnaCharacterMatrix, DataSet, datamodel
+from dendropy import Tree, DnaCharacterMatrix
 
 from Bio.Blast import NCBIXML
 from Bio.Seq import Seq
 
-from physcraper.configobj import ConfigObj
 from physcraper.ids import IdDicts
 from physcraper.aligntreetax import AlignTreeTax
 from physcraper.helpers import cd, get_raxml_ex
 from physcraper.ncbi_data_parser import get_gi_from_blast, get_acc_from_blast
 from physcraper.opentree_helpers import root_tree_from_synth
 
-from physcraper import ncbi_data_parser
 from physcraper import AWSWWW
 
 _VERBOSE = 0
 
 def set_verbose():
+    """Set output to verbose"""
     global _VERBOSE
     _VERBOSE = 1
 
@@ -59,7 +57,8 @@ class PhyscraperScrape():
 
             * key: gi id
             * value: corresponding seq
-          * **self.new_seqs_otu_id**: dictionary that contains the new sequences that passed the remove_identical_seq() step:
+          * **self.new_seqs_otu_id**: dictionary that contains
+          the new sequences that passed the remove_identical_seq() step:
 
             * key: otu_id
             * value: see otu_dict, is a subset of the otu_dict, all sequences that will be newly added to aln and tre
@@ -75,16 +74,22 @@ class PhyscraperScrape():
 
           * **self.newseqs_file**: filename of files that contains the sequences from self.new_seqs_otu_id
           * **self.date**: Date of the run - may lag behind real date!
-          * **self.repeat**: either 1 or 0, it is used to determine if we continue updating the tree, no new seqs found = 0
-          * **self.newseqs_acc**: list of all gi_ids that were passed into remove_identical_seq(). Used to speed up adding process
-          * **self.blocklist**: list of gi_id of sequences that shall not be added or need to be removed. Supplied by user.
-          * **self.seq_filter**: list of words that may occur in otu_dict.status and which shall not be used in the building of FilterBlast.sp_d (that's the main function), but it is also used as assert statement to make sure unwanted seqs are not added.
+          * **self.repeat**: either 1 or 0, it is used to determine if we continue updating the tree,
+          no new seqs found = 0
+          * **self.newseqs_acc**: list of all gi_ids that were passed into remove_identical_seq().
+          Used to speed up adding process
+          * **self.blocklist**: list of gi_id of sequences that shall not be added or need to be removed.
+          Supplied by user.
+          * **self.seq_filter**: list of words that may occur in otu_dict.status and which shall not be
+          used in the building of FilterBlast.sp_d
+           (that's the main function), but it is also used as assert statement to make sure unwanted seqs are not added.
           * **self.unpublished**: True/False. Used to look for local unpublished seq that shall be added if True.
           * **self.path_to_local_seq:** Usually False, contains path to unpublished sequences if option is used.
 
         Following functions are called during the init-process:
 
-            * **self.reset_markers()**: adds things to self: I think they are used to make sure certain function run, if program crashed and pickle file is read in.
+            * **self.reset_markers()**: adds things to self: I think they are used to make sure certain function run,
+            if program crashed and pickle file is read in.
                 * self._blasted: 0/1, if run_blast_wrapper() was called, it is set to 1 for the round.
                 * self._blast_read: 0/1, if read_blast_wrapper() was called, it is set to 1 for the round.
[…]
