### 2020-09-02 #26 @snacktavish
Edited to new release (0.55), with some minor updates.
