+++ requirements.txt
@@ -4,8 +4,6 @@ configparser
 coverage
 DendroPy
 DateTime
--e git+https://github.com/OpenTreeOfLife/python-opentree@get-tree#egg=opentree
-
 opentree
 pandas
 
