+++ docs/source/conf.py
@@ -13,7 +13,11 @@
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
 import physcraper
-# import sys
+import os
+import sys
+from datetime import date
+
+
 # sys.path.insert(0, os.path.abspath('.'))
 
 
@@ -39,7 +43,8 @@ release = u'0.1'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'm2r',
+    # 'recommonmark',
+    'm2r2',
     'sphinx.ext.autodoc',
     'sphinx.ext.intersphinx',
     'sphinx.ext.viewcode',
@@ -51,8 +56,8 @@ templates_path = ['_templates']
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
 #
-source_suffix = ['.rst', '.md']
-#source_suffix = '.rst'
+# source_suffix = ['.rst', '.md']
+source_suffix = '.rst'
 
 # The master toctree document.
 master_doc = 'index'
@@ -78,6 +83,7 @@ pygments_style = None
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
+
 html_theme = 'alabaster'
 
 # Theme options are theme-specific and customize the look and feel of a theme
