+++ docs/source/conf.py
@@ -19,6 +19,8 @@ from datetime import date
 
 
 # sys.path.insert(0, os.path.abspath('.'))
+import sphinx
+import sphinx_readable_theme
 
 
 # -- Project information -----------------------------------------------------
@@ -83,8 +85,9 @@ pygments_style = None
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
+html_theme_path = [sphinx_readable_theme.get_html_theme_path()]
 
-html_theme = 'alabaster'
+html_theme = 'readable'
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
