+++ docs/source/index.rst
@@ -11,19 +11,10 @@ Welcome to Physcraper's documentation!
 
 .. mdinclude:: ../mds/intro-dendropy.md
 
-|
-|
-|
-|
-|
-|
-|
-
-
-
 .. toctree::
    :maxdepth: 3
    :caption: Documentation topics:
+   :hidden:
 
    how_to_start
    methods_extended
