+++ docs/source/index.rst
@@ -16,7 +16,7 @@ Welcome to Physcraper's documentation!
    :caption: Documentation topics:
    :hidden:
 
-   how_to_start
+   Introduction <how_to_start>
    methods_extended
    examples
    install
