+++ docs/source/methods_extended.rst
@@ -1,3 +1,3 @@
-Framework
+How does Physcraper work
 ===========
 .. mdinclude:: ../mds/methods_extended.md
