+++ docs/source/install.rst
@@ -1,4 +1,4 @@
-Installation
+Installing Physcraper
 ============
 
 .. mdinclude:: ../mds/INSTALL.md
