+++ docs/source/conf.py
@@ -21,6 +21,8 @@ from datetime import date
 # sys.path.insert(0, os.path.abspath('.'))
 import sphinx
 import sphinx_readable_theme
+import sphinx_rtd_theme
+
 
 
 # -- Project information -----------------------------------------------------
@@ -50,6 +52,7 @@ extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.intersphinx',
     'sphinx.ext.viewcode',
+    "sphinx_rtd_theme"
 ]
 
 # Add any paths that contain templates here, relative to this directory.
@@ -85,15 +88,18 @@ pygments_style = None
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme_path = [sphinx_readable_theme.get_html_theme_path()]
+# html_theme_path = [sphinx_readable_theme.get_html_theme_path()]
 
-html_theme = 'readable'
+# html_theme = 'readable'
+html_theme = "sphinx_rtd_theme"
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
 #
-# html_theme_options = {}
+html_theme_options = {
+    'sticky_navigation': False,
+    'style_nav_header_background': 'green'}
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
