+++ requirements.txt
@@ -4,7 +4,8 @@ configparser
 coverage
 DendroPy
 DateTime
-opentree
+nexson
+-e git+https://github.com/OpenTreeOfLife/python-opentree@main#egg=opentree
 pandas
 
 requests
