### 2020-10-14 #26 @lwasser
hi there @snacktavish i'm so sorry we have been out of communication here. are you still interested in a review by pyopensci? i've been a bit overwhelemed by current events which has been a hit at work in productivity. We may have an editor for this however if you are still interested in a review.
