### 2020-10-15 #26 @NickleDave
Hi @snacktavish !
I volunteered myself to be editor.

Currently I work in AI / neuro but my background is bio ... and one of the main reasons I made it into science is a really good prof who taught the first non-boring class I took, on bioinformatics where we worked on genome assembly, phylogenetics, and related things.

So I think I get at a high level what `physcraper` does and it definitely looks like a good fit for PyOpenSci.

One thing: since we are still relatively new, we don't have a huge pool of reviewers to draw from right now. I can definitely think of some places to start asking, but do you by any chance have two people you could recommend as reviewers?
