### 2020-10-21 #26 @NickleDave
Hey @snacktavish just want to check back -- please let me know if you need me to start looking for reviewers

> One thing: since we are still relatively new, we don't have a huge pool of reviewers to draw from right now. I can think of some places to start asking, but do you by any chance have two people you could recommend as reviewers?

and I agree with Leah that you have the best GitHub handle ever
