### 2020-10-27 #26 @NickleDave
@snacktavish @lwasser adding initial editor checks.
Very swamped at work right now but I will reach out to reviewers by Thursday at the latest

## Editor checks:

- [x] **License:** The package has an OSI accepted license

---

## Editor comments
* Looks like a perfect fit for pyOpenSci, within scope under categories of data retrieval and munging as maintainer has indicated.
* Package appears well-developed and maintained to me. I see no obvious issues at this time.
* Manuscript on biorxiv
* My impression is review would best address how the package meets its goal of usability by a non-specialist. Input from users and developers that have experience with related tools would be most helpful

---

Reviewers:
Due date:
