### 2020-11-28 #26 @NickleDave
update update: very excited that @bpucker has accepted the invitation to review!!!  🎉🎉🎉
who has a ton of experience developing bioinformatics tools in Python
(see https://github.com/bpucker -- thank you @NatJWalker-Hale for connecting us)

Boas, just a reminder that our reviewer guide is here:
https://www.pyopensci.org/contributing-guide/open-source-software-submissions/reviewer-guide.html
and our review template is here:
https://www.pyopensci.org/contributing-guide/appendices/templates.html#review-template
Thank you!!!
