### 2020-11-29 #26 @bpucker
- [x] As the reviewer I confirm that there are no conflicts of interest for me to review this work.

- [ ] **Function Documentation:** for all user-facing functions
- [ ] **Examples** for all user-facing functions
- [x] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.

Readme requirements

- [x] Badges for continuous integration and test coverage, a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the badge for pyOpenSci peer-review will be provided upon acceptance.)
- [x] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
- [x] Any additional setup required (authentication tokens, etc)
- [ ] Brief demonstration usage
- [x] Direction to more detailed documentation (e.g. your documentation files or website).
- [x] If applicable, how the package compares to other similar packages and/or how it relates to other packages

Package structure should follow general community best-practices. In general please consider:

- [x] The documentation is easy to find and understand

- [x] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

1) Installation instructions are not in the README.md, but there are links to all relevant pages. Documentation and examples are also provided on different pages. I think this is fine though.

2) Some parts (relabeling/rerooting) are labeled as "under construction" thus I have not tested those.

3) Anaconda/virtualenv installation instructions should be added and these requirements should be specified in the README.

4) Support for additional alignment tools e.g. MAFFT could be included to allow user to choose their favorite one.

5) There are some details missing in the function documentation:
self.otu_by_gi
self._to_be_pruned
self.tmpfi

6) The BLAST jobs run for a very long time (as promised in the documentation). I know this is tricky, but some kind of progress bar would be great.

7) Would it be possible to use peptide sequences? These could show better conservation over large phylogenetic distances.

8) Everything looks fine, but I got an issue when running physcraper on a different dataset. Here are the details:

Description:	Ubuntu 18.04.2 LTS
Release:	18.04
Codename:	bionic
Python 3.8.5 (on the system)
Python 3.6.11 (in physcraper_env)

anaconda Command line client (version 1.7.2)

physcraper_run.py -s pg_328 -t tree322 -r -tb -o chenopodia

  File "/home/ubuntu/anaconda3/envs/physcraper_env/bin/physcraper_run.py", line 7, in <module>
    exec(compile(f.read(), __file__, 'exec'))
  File "XXX/physcraper/bin/physcraper_run.py", line 181, in <module>
    dataset = physcraper.opentree_helpers.get_dataset_from_treebase(study_id)
  File "XXX/physcraper/physcraper/opentree_helpers.py", line 390, in get_dataset_from_treebase
    dna = DataSet.get(url=url, schema="nexml")
  File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/site-packages/dendropy/datamodel/datasetmodel.py", line 188, in get
    return cls._get_from(**kwargs)
  File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/site-packages/dendropy/datamodel/basemodel.py", line 160, in _get_from
    return cls.get_from_url(src=src, schema=schema, **kwargs)
  File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/site-packages/dendropy/datamodel/basemodel.py", line 279, in get_from_url
    **kwargs)
  File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/site-packages/dendropy/datamodel/datasetmodel.py", line 129, in _parse_and_create_from_stream
    state_alphabet_factory=charstatemodel.StateAlphabet,
  File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/site-packages/dendropy/dataio/ioservice.py", line 255, in read_dataset
    global_annotations_target=dataset)
  File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/site-packages/dendropy/dataio/nexmlreader.py", line 284, in _read
    subelement_factory=self._subelement_factory)
  File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/site-packages/dendropy/dataio/xmlprocessing.py", line 150, in __init__
    self.parse_file(file_obj)
  File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/site-packages/dendropy/dataio/xmlprocessing.py", line 166, in parse_file
    for event, elem in ElementTree.iterparse(source, events):
  File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/xml/etree/ElementTree.py", line 1221, in iterator
    yield from pullparser.read_events()
  File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/xml/etree/ElementTree.py", line 1296, in read_events
    raise event
  File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/xml/etree/ElementTree.py", line 1268, in feed
    self._parser.feed(data)
xml.etree.ElementTree.ParseError: syntax error: line 1, column 0

These automated tree > These automated trees (https://physcraper.readthedocs.io/en/latest/how_to_start.html)
mapped a unified taxonomy > mapped to a unified taxonomy

There are formatting issue on https://physcraper.readthedocs.io/en/latest/physcraper_run.html involving
1) "physcraper_run.py -tf tests/data/tiny_test_example/test.tre -tfs newick -a tests/data/tiny_test_example/test.fas -as fasta –taxon_info tests/data/tiny_test_example/main.json -o owndata"
2) "Take a look at the tree, the alignemnt and the out_info csv file. It will list all taxa by their unique identifiers."

final output ree > final output tree

cofniguration > configuration

updtaed > updated (https://physcraper.readthedocs.io/en/latest/data_exploration.html)

Genbank/genbank > GenBank (https://physcraper.readthedocs.io/en/latest/apidocs.html)

It appears that there should be a link on the overview figure, but the pages does not exist: https://physcraper.readthedocs.io/en/latest/how_to_start.html#
