### 2020-11-30 #26 @snacktavish
Thank you so much for the thorough review!
We will create issues, and address them, and then link them here where they are all completed, unless there is a different procedure you would like us to follow.
