### 2020-11-30 #26 @NickleDave
yes 🙏 thank you very much @bpucker -- looks quite helpful

@snacktavish creating issues and linking here is perfect, as was done in this previous review https://github.com/pyOpenSci/software-review/issues/20#issuecomment-626467367

I will keep looking for one more reviewer
