### 2020-12-02 McTavishLab/physcraper#143 @snacktavish
reviewer 1 - comment 3 - Anaconda/virtualenv
"Anaconda/virtualenv installation instructions should be added and these requirements should be specified in the README."
