### 2020-12-02 McTavishLab/physcraper#145 @snacktavish
Function documentation
There are some details missing in the function documentation:
self.otu_by_gi
self._to_be_pruned
self.tmpfi
