### 2020-12-02 McTavishLab/physcraper#146 @snacktavish
Progress bar
The BLAST jobs run for a very long time (as promised in the documentation). I know this is tricky, but some kind of progress bar would be great.
