### 2020-12-02 McTavishLab/physcraper#147 @snacktavish
reviewer 1 - comment 8 - Error on treebase tree
```physcraper_run.py -s pg_328 -t tree322 -r -tb -o chenopodia```
