### 2020-12-02 McTavishLab/physcraper#147 @snacktavish
aha! The supertreebase mirror has a file for that study, but the data consists of the single character 'x', so the data set loading crashes. https://raw.githubusercontent.com/TreeBASE/supertreebase/master/data/treebase/S12369.xml

Updated function to look on treebase if supertreebase fails, even if it appears to have a file.
