+++ tests/test_webservices.py
@@ -77,6 +77,14 @@ def test_get_from_treebase():
                             "-no_est",
                             "-o", "tests/data/tmp/pg_55_treebase"])
 
+def test_get_from_treebase_supertreebase_err():
+    subprocess.check_call(["python", "bin/physcraper_run.py", 
+                            "-s" "pg_328",
+                            "-t", "tree322",
+                            "-tb",
+                            "-no_est",
+                            "-o", "tests/data/tmp/pg_328_treebase"])
+
 def test_web_blast():
     scraper.run_blast_wrapper()
     scraper.read_blast_wrapper()
