### 2020-12-02 McTavishLab/physcraper#147 @snacktavish
Fixed and tested in https://github.com/McTavishLab/physcraper/commit/d293dc741318b14118c52f3beca417a3afb985a5
