### 2020-12-22 #26 @NickleDave
Hi all, just checking in here.
I am still looking for a second reviewer, but I am hesitant to email people right before a major holiday.
@snacktavish I plan to start contacting people again the week of Jan. 11.
Please let me know if that's okay.
It looks like you might still be working through issues from the first review, so I'm hoping you are not feeling too impatient with the editor (me) yet.
