### 2021-01-14 1df77c6
keeping docs simple with the alabaster theme
