### 2021-01-21 afc0f48
info on installation checks
