+++ docs/mds/INSTALL.md
@@ -87,3 +87,23 @@ You can check if they are already installed with:
 which muscle
 which raxmlHPC
 ```
+## Checking install success on remote searches
+
+To test a full run with pre-downloaded Blast results, copy the example results using:
+
+    cp -r docs/examples/pg_55_web pg_55_test
+
+and then run:
+
+    physcraper_run.py --study_id pg_55 --tree_id tree5864 --treebase --bootstrap_reps 10 --output pg_55_test
+
+There is more info on all the parameter setting in [Physcraper_run](https://physcraper.readthedocs.io/en/latest/physcraper_run.html), but briefly this gets a tree (tree5864) from study pg_55 on OpenTree, pulls the alignment from tree base, blasts the sequences, and does 10 bootstrap reps on the final phylogeny.
+
+
+This example tests all the components, except the actual remote blast searches (because they can be very sloooow).
+To check if your installation was successful for remote searches, try running an analysis:
+
+    physcraper_run.py --study_id pg_55 --tree_id tree5864 --treebase --bootstrap_reps 10 --output pg_55_new
+
+This run will take a while - once it starts blasting, that means it's working! You can use Ctrl-C to cancel.
+
