### 2021-01-22 f9cd436
removing html theme option from conf.py
