+++ docs/source/conf.py
@@ -91,9 +91,7 @@ html_theme = "alabaster"
 # further.  For a list of options available for each theme, see the
 # documentation.
 #
-html_theme_options = {
-    'sticky_navigation': False,
-    'style_nav_header_background': 'green'}
+html_theme_options = {}
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
