### 2021-01-22 51e4d31
removing html sidebars from conf.py
