### 2021-01-22 899cd4d
commenting out html theme option from conf.py
