+++ docs/source/conf.py
@@ -46,7 +46,7 @@ extensions = [
     'm2r2',
     'sphinx.ext.autodoc',
     'sphinx.ext.intersphinx',
-    'sphinx.ext.viewcode'
+    'sphinx.ext.viewcode',
 ]
 
 # Add any paths that contain templates here, relative to this directory.
