+++ docs/mds/intro-dendropy.md
@@ -1,5 +1,7 @@
 Physcraper relies on:
 
-- [Dendropy](https://dendropy.org/primer/index.html): *Sukumaran, J and MT Holder. 2010. DendroPy: a Python library for phylogenetic computing. Bioinformatics 26: 1569-1571*
-
-- [The Open Tree of Life Project](https://tree.opentreeoflife.org/opentree/argus/opentree12.3@ott93302): *Open Tree of Life, Benjamin Redelings, Luna Luisa Sanchez Reyes, Karen A. Cranston, Jim Allman, Mark T. Holder, & Emily Jane McTavish. (2019). Open Tree of Life Synthetic Tree (Version 12.3). Zenodo. [doi.org/10.5281/zenodo.3937741](https://doi.org/10.5281/zenodo.3937741)*
+- [Dendropy](https://dendropy.org/primer/index.html): *Sukumaran, J and MT Holder. 2010. DendroPy: a Python library for phylogenetic computing. Bioinformatics 26: 1569-1571* [doi.org/10.1093/bioinformatics/btq228]( https://doi.org/10.1093/bioinformatics/btq228)
+<br/><br/>
+- [The Open Tree of Life Project](https://tree.opentreeoflife.org/opentree/argus/opentree12.3@ott93302): *Open Tree of Life, Benjamin Redelings, Luna Luisa Sanchez Reyes, Karen A. Cranston, Jim Allman, Mark T. Holder, & Emily Jane McTavish. (2019). Open Tree of Life Synthetic Tree (Version 12.3). Zenodo* [doi.org/10.5281/zenodo.3937741](https://doi.org/10.5281/zenodo.3937741)
+<br/><br/>
+- [OpenTree](https://github.com/OpenTreeOfLife/python-opentree): *Emily Jane McTavish, Luna Luisa Sanchez Reyes, Mark T. Holder. (2020). OpenTree: A Python package for Accessing and Analyzing data from the Open Tree of Life. BioRxiv 2020.12.14.422759* [doi.org/10.1101/2020.12.14.422759](https://doi.org/10.1101/2020.12.14.422759)
