### 2021-01-25 #26 @NickleDave
Hi @snacktavish and @LunaSare happy to report that @leomrtns has generously accepted our invitation to review

@leomrtns for your convenience here are the links to our reviewer guide and review template:
https://www.pyopensci.org/contributing-guide/open-source-software-submissions/reviewer-guide.html
https://www.pyopensci.org/contributing-guide/appendices/templates.html#review-template
Thank you again

Given everything else that is going on, I have set a due date for this review of five weeks.
I will update my [initial editor checks comment](https://github.com/pyOpenSci/software-review/issues/26#issuecomment-717299442) above

As we discussed I think this should be okay given that the authors are still working through issues from the first review
