### 2021-01-26 McTavishLab/physcraper#152 @LunaSare
reviewer 1 - comment 1 - missing things on readme
Things that should be on readme according to pysci that we are missing:

- Installation instructions
- Brief demonstration usage
- Citation information

Reviewer Comments
"Installation instructions are not in the README.md, but there are links to all relevant pages. Documentation and examples are also provided on different pages. I think this is fine though."

So I think this is ok
