### 2021-01-26 McTavishLab/physcraper#143 @LunaSare
These are already on the installation instructions section at [readthedocs](https://physcraper.readthedocs.io/en/latest/install.html#option-1-install-physcraper-using-conda)
But we can surely make that more clear.

- [ ] add to readme
- [ ] highlight on installation instructions
