### 2021-01-26 McTavishLab/physcraper#154 @LunaSare
reviewer 1 - comment 7 - aminoacid alignments
"Would it be possible to use peptide sequences? These could show better conservation over large phylogenetic distances."
