### 2021-01-26 McTavishLab/physcraper#155 @LunaSare
reviewer 1 - minor comments
- [x] These automated tree > These automated trees (https://physcraper.readthedocs.io/en/latest/how_to_start.html)

- [x] mapped a unified taxonomy > mapped to a unified taxonomy

- [x] There are formatting issue on https://physcraper.readthedocs.io/en/latest/physcraper_run.html involving

"physcraper_run.py -tf tests/data/tiny_test_example/test.tre -tfs newick -a tests/data/tiny_test_example/test.fas -as fasta –taxon_info tests/data/tiny_test_example/main.json -o owndata"
"Take a look at the tree, the alignemnt and the out_info csv file. It will list all taxa by their unique identifiers."

- [x] final output ree > final output tree

- [x] cofniguration > configuration

- [x] updtaed > updated (https://physcraper.readthedocs.io/en/latest/data_exploration.html)

- [x] Genbank/genbank > GenBank (https://physcraper.readthedocs.io/en/latest/apidocs.html)

- [x] It appears that there should be a link on the overview figure, but the pages does not exist: https://physcraper.readthedocs.io/en/latest/how_to_start.html#
