### 2021-01-26 58c441e
highlighting conda virtual env on installation docs
