+++ README.md
@@ -1,12 +1,12 @@
 
 <!-- README.md is generated from README.Rmd; please edit the .Rmd file and then from R do rmarkdown::render("README.Rmd")-->
 
-<img align="left" width="250" src="https://raw.githubusercontent.com/McTavishLab/physcraper/main/docs/physcraper.svg">
+<img align="left" width="250" src="https://raw.githubusercontent.com/McTavishLab/physcraper/cleanup/docs/physcraper-long.png">
 
 # Physcraper
 
 [![Build
-Status](https://travis-ci.org/McTavishLab/physcraper.svg?branch=main)](https://travis-ci.org/McTavishLab/physcraper)[![Documentation](https://readthedocs.org/projects/physcraper/badge/?version=latest&style=flat)](https://physcraper.readthedocs.io/en/latest/)[![codecov](https://codecov.io/gh/McTavishLab/physcraper/branch/main/graph/badge.svg)](https://codecov.io/gh/McTavishLab/physcraper)[![NSF-1759846](https://img.shields.io/badge/NSF-1759846-blue.svg)](https://nsf.gov/awardsearch/showAward?AWD_ID=1759846) 
+Status](https://travis-ci.org/McTavishLab/physcraper.svg?branch=main)](https://travis-ci.org/McTavishLab/physcraper)[![Documentation](https://readthedocs.org/projects/physcraper/badge/?version=latest&style=flat)](https://physcraper.readthedocs.io/en/latest/)[![codecov](https://codecov.io/gh/McTavishLab/physcraper/branch/main/graph/badge.svg)](https://codecov.io/gh/McTavishLab/physcraper)
 
 <p>
 
@@ -26,16 +26,20 @@ Lab](https://mctavishlab.github.io/). Please post a GitHub issue
 [here](https://github.com/McTavishLab/physcraper/issues) or contact
 <ejmctavish@ucmerced.edu> if you need any help or have feedback.
 
-This is the code repository. Please refer to Physcraper’s [documentation
-website](https://physcraper.readthedocs.io/en/latest/) for more details
-on
+This is the code repository. Please refer to `Physcraper`’s
+[documentation website](https://physcraper.readthedocs.io/en/latest/)
+for more details on
 [installation](https://physcraper.readthedocs.io/en/latest/install.html)
-instructions,
+instructions using an Anaconda or a Virtualenv Python virtual
+environment,
 [function](https://physcraper.readthedocs.io/en/latest/apidocs.html)
-usage, running
-[tutorials](https://physcraper.readthedocs.io/en/latest/physcraper_run.html),
-real life
-[examples](https://physcraper.readthedocs.io/en/latest/data_exploration.html),
+usage, tutorials on finding and choosing a [starting
+dataset](https://physcraper.readthedocs.io/en/latest/find_trees.html),
+[running](https://physcraper.readthedocs.io/en/latest/physcraper_run.html)
+a full analysis, and
+[visualizing](https://physcraper.readthedocs.io/en/latest/data_exploration.html)
+results, as well as real life
+[examples](https://physcraper.readthedocs.io/en/latest/examples.html)
 and tools for
 [developers](https://physcraper.readthedocs.io/en/latest/CONTRIBUTING.html).
 
@@ -45,13 +49,21 @@ and tools for
 
 Physcraper relies on:
 
-[Dendropy](https://dendropy.org/primer/index.html) *Sukumaran, J and
-MT Holder. 2010. DendroPy: a Python library for phylogenetic computing.
-Bioinformatics 26: 1569-1571*.
-
[…]
