### 2021-01-26 McTavishLab/physcraper#143 @LunaSare
added to readme 6cd952abc17129ff904e0c9b6786d1a567fe3640 and highlighted on installation instructions at readthedocs 58c441ed1e03181d7b27bec5f88b7b3bc3c5a932
