### 2021-01-26 McTavishLab/physcraper#152 @LunaSare
But maybe we want the citation and the license on the readme, too?
