+++ docs/source/data_exploration.rst
@@ -1,8 +1,8 @@
-Visualizing the Physcraper results
-================
+The Physcraper results
+======================
 
 .. mdinclude:: ../mds/results.md
 
-.. mdinclude:: ../mds/DataExploration.md
-
 .. mdinclude:: ../mds/malvaceae.md
+
+.. mdinclude:: ../mds/DataExploration.md
