### 2021-01-26 efeb56f
updating results documentation section
