+++ docs/mds/results.md
@@ -1,4 +1,4 @@
-## The Physcraper results files
+## Files generated by a Physcraper run
 
 ### - Input files
 
