### 2021-01-27 McTavishLab/physcraper#153 @LunaSare
Flesh out these sections:

- [ ] https://physcraper.readthedocs.io/en/cleanup/find_trees.html#find-the-corresponding-alignment
- [ ] https://physcraper.readthedocs.io/en/cleanup/data_exploration.html#tree-comparison-with-robinson-foulds
- [ ] https://physcraper.readthedocs.io/en/cleanup/data_exploration.html#rerooting-the-trees
