+++ docs/mds/methods_extended.md
@@ -3,7 +3,7 @@
 <div id="the-inputs-a-tree-and-an-alignment" class="section level2">
 <h2>The inputs: a tree and an alignment</h2>
 <p>In order to take advantage of the OpenTree tools, it is recommended that the input tree is either stored in the OpenTree <a href="https://github.com/opentreeoflife/phylesystem">Phylesystem</a>, or submitted via OpenTree’s curator <a href="https://tree.opentreeoflife.org/curator">application</a> <span class="citation">(McTavish <em>et al.</em> <a href="#ref-mctavish2015phylesystem">2015</a>)</span>. Currently, only trees connected to a published study can be stored in the Phylesystem. Users can choose from among the 2, 950 studies in OpenTree’s Phylesystem that have alignments on TreeBASE. If the user is not ready to make the input tree public, tree tip labels can be standardized to the unified OpenTree taxonomy using OpenTree’s bulk Taxonomic Name Resolution Service <a href="https://tree.opentreeoflife.org/curator/tnrs/">TNRS</a> tool. This step is referred to as taxonomic name mapping and Phylesystem stored trees are processed in this way upon submission. Physcraper saves a summary “csv” file with results from the taxon name standardization for advantage of the user, including the mappings to unique identifiers in the OpenTree and NCBI taxonomies. If taxon names can’t be mapped, their taxonomic information is not used in comparison analysis. These taxa will still be used in the sequence search and phylogenetic reconstruction steps. Mapping tip names to OpenTree’s unified taxonomy saves a set of user defined characteristics that are essential for automatizing the phylogeny updating process. The most relevant of these is the standardized taxonomic names and the definition of ingroup and outgroup taxa, allowing to automatically set the root for the updated tree on the final steps of the pipeline.</p>
-<p>The input alignment should be a single locus alignment that was used in part or in whole, to generate the tree. Alignments are often stored in a public repository such as TreeBase <span class="citation">(Piel <em>et al.</em> <a href="#ref-piel2009treebase">2009</a>; Vos <em>et al.</em> <a href="#ref-vos2012nexml">2012</a>)</span>, DRYAD (<a href="http://datadryad.org/">www.datadryad.org</a>), or a data repository associated with the journal where the tree was originally published. If the alignment is stored in TreeBase, Physcraper can download it directly, either from the TreeBASE website (<a href="https://treebase.org/">www.treebase.org</a>) or through the TreeBASE GitHub repository (SuperTreeBASE; <a href="https://github.com/TreeBASE/supertreebase">github.com/TreeBASE/supertreebase</a>). If the alignment is on another repository, or constitutes personal data, a path to a local copy of the alignment has to be provided.</p>
[…]
