### 2021-01-28 d9261ae
cleaning up physcraper run doc
