+++ docs/mds/PhyscraperRun.md
@@ -1,66 +1,82 @@
-## Example commands
 
-The easiest way to run physcraper is using the command line tools. This way, you can directly specify arguments, and a config file will be written down for the sake of reproducibility.
+The easiest way to run Physcraper is using the command line tools. This way, you can directly specify arguments.
+A configuration file will be written down for the sake of reproducibility.
 
-As input, you will minimally need a study and tree ids from a tree uploaded to the Open Tree of Life website. The 'treebase' flag will automatically download an alignment for that tree from treebase.
+## Running Physcraper from the command line
+
+As input, you will minimally need a study and tree ids from a tree uploaded to the Open Tree of Life website. The 'treebase' flag will automatically download an alignment for that tree from TreeBASE.
 
 ```
 physcraper_run_py -s OPENTREE_STUDY_ID -t OPENTREE_TREE_ID --treebase -o OUTPUT_DIRECTORY_NAME/AND/OR/PATH
 ```
 
-e.g.
+e.g.,
 
 
 ```
 physcraper_run.py -s pg_55 -t tree5864 -tb -o pg_55_web
 ```
 
-The output files generated by this run are stored in docs/examples/pg_55_web
+The output files generated by this run are stored in "docs/examples/pg_55_web"
 
-Alternately, you can provide the gene alignment that you want to be updated using the `-a` command:
+Alternatively, you can provide the gene alignment that you want to be updated using the `-a` command:
 
 ```
 physcraper_run_py -s OPENTREE_STUDY_ID -t OPENTREE_TREE_ID -o OUTPUT/DIRECTORY/NAME/AND/OR/PATH -a PATH/TO/GENE/ALIGNMENT/NAME -as ALIGNMENT_SCHEMA
 
 ```
 
-The outputs from this command are avialable
+e.g.,  to update this tree [Crous et al. 2012](https://tree.opentreeoflife.org/curator/study/view/ot_350/?tab=home&tree=Tr53296) using an alignment already downloaded from TreeBASE.
 
-```
-physcraper_run.py -s pg_55 -t tree5864 -tb -o pg55_treebase
-```
-
-
-e.g.  to update this tree [Crous et al. 2012](https://tree.opentreeoflife.org/curator/study/view/ot_350/?tab=home&tree=Tr53296) using an alignment already downloaded from treebase.
 ```
 physcraper_run.py -s ot_350 -t Tr53297 -a docs/examples/inputdata/ot_350Tr53297.aln -as "nexus" -o ot_350
 ```
-If the tree you want to update is not posted to the OpenTree website, you need to match the labels the labels on your tree to taxa using [OpenTree Bulk Taxonomic Name Resolution](https://tree.opentreeoflife.org/curator/tnrs/). Download your matched names, unzip the folder, and pass the .json file as a physcraper argument.
+
+If the tree you want to update is not posted to the OpenTree website, you need to match
+the labels on your tree to taxa using [OpenTree Bulk Taxonomic Name Resolution Service](https://tree.opentreeoflife.org/curator/tnrs/). Download your matched names, unzip the folder, and pass the .json file as `--taxon_info` argument.
 
 ```
 physcraper_run.py -tf TREE_FILE -tfs TREEFILE_SCHEMA -a ALIGNMENT_FILE -as ALIGNMENT_SCHEMA --taxon_info TAXON_INFO_JSONFILE  -o OUTPUT_DIRECTORY_NAME
 ```
 
+e.g.,
 
+```
 physcraper_run.py -tf tests/data/tiny_test_example/test.tre -tfs newick -a tests/data/tiny_test_example/test.fas  -as fasta --taxon_info tests/data/tiny_test_example/main.json -o owndata
 ```
-
-
[…]
