+++ docs/mds/citation.md
@@ -1,3 +1,3 @@
-If you use Physcraper, please cite 
+If you use Physcraper, please cite: 
 
 - [Physcraper](https://www.biorxiv.org/content/10.1101/2020.09.15.299156v1.abstract): *Luna Luisa Sanchez Reyes, Martha Kandziora, Emily Jane McTavish. (2020). Physcraper: A Python package for continual update of evolutionary estimates using the Open Tree of Life. BioRxiv 2020.09.15.299156*. doi: [doi.org/10.1101/2020.09.15.299156](https://doi.org/10.1101/2020.09.15.299156)
