+++ docs/source/index.rst
@@ -11,6 +11,8 @@ Welcome to Physcraper's documentation!
 
 .. mdinclude:: ../mds/intro-dendropy.md
 
+.. mdinclude:: ../mds/citation.md
+
 .. toctree::
    :maxdepth: 3
    :caption: Documentation topics:
