+++ docs/mds/results.md
@@ -15,5 +15,5 @@ The trees are reconstructed using RAxML, with tip labels corresponding to local 
 
 An updated tree in nexson format, containing all types of tip labels is saved in here.
 From this tree, a tree with any kind of label can be produced.
-By default, the updtaed tree with taxon names as tip labels is saved
+By default, the updated tree with taxon names as tip labels is saved
 in the `output_tag` folder as `updated_taxonname.tre`. See section [relabeling the trees](# Relabeling) below for more info on how to generate trees with different types of labels.
