+++ docs/mds/Tutorial.md
@@ -8,7 +8,8 @@ If you have access to a single gene alignment, and a tree, Physcraper automates 
 
 <br/>
 
-![](../img/schematic.svg)
+<img src="https://raw.githubusercontent.com/McTavishLab/physcraper/main/docs/img/schematic.svg" title="schematic" alt="The Phsycraper Framework" width="100%" />
+
 The Physcraper framework consists of 4 general steps. The methodology is extensively described in the *Framework* section.
 
 <br/>
