### 2021-02-04 McTavishLab/physcraper#155 @LunaSare
Done here 06b2fda553a874661e6860f54f9962734ac3042e and here 970860158a9088a76306f56166d2f91b7991bd68
