+++ docs/mds/license.md
@@ -1,0 +1,1 @@
+Physcraper is made available through the [GNU General Public License v3.0](https://github.com/McTavishLab/physcraper/blob/main/LICENSE)
