### 2021-02-04 fc253bf
adding license to credits docs
