+++ README.Rmd
@@ -57,3 +57,7 @@ and tools for [developers](https://physcraper.readthedocs.io/en/latest/CONTRIBUT
 ```{r child="docs/mds/citation.md"}
 
 ```
+
+```{r child="docs/mds/license.md"}
+
+```
