### 2021-02-04 McTavishLab/physcraper#152 @LunaSare
Yes, see https://github.com/McTavishLab/physcraper/issues/157 for citation
License done here
2bb37ea5a8d4fedb78990c3098b22b9bc51bae9b
fc253bf30d7d81b85fdcae936a84b0ff0e691693
d4734e0ea17c7b75f88f42814bbeaeaaf516ea19
9b666faca2d0d9891eaf11b000b2abc5c431b8a8
