### 2021-02-04 8572069
removing duplicate example commands
