+++ docs/mds/PhyscraperRun.md
@@ -2,12 +2,14 @@
 The easiest way to run Physcraper is using the command line tools. This way, you can directly specify arguments.
 A configuration file will be written down for the sake of reproducibility.
 
-## Running Physcraper from the command line
+## Example Physcraper runs from the command line
 
-As input, you will minimally need a study and tree ids from a tree uploaded to the Open Tree of Life website. The 'treebase' flag will automatically download an alignment for that tree from TreeBASE.
+### Starting with only an OpenTree study and tree id
+
+As input, you will minimally need a study and tree ids from a tree uploaded to the Open Tree of Life website. The `--treebase` flag (or `-tb`) will automatically download an alignment for that tree from TreeBASE.
 
 ```
-physcraper_run_py -s OPENTREE_STUDY_ID -t OPENTREE_TREE_ID --treebase -o OUTPUT_DIRECTORY_NAME/AND/OR/PATH
+physcraper_run.py -s OPENTREE_STUDY_ID -t OPENTREE_TREE_ID --treebase -o OUTPUT_DIRECTORY_NAME/AND/OR/PATH
 ```
 
 e.g.,
@@ -19,10 +21,12 @@ physcraper_run.py -s pg_55 -t tree5864 -tb -o pg_55_web
 
 The output files generated by this run are stored in "docs/examples/pg_55_web"
 
-Alternatively, you can provide the gene alignment that you want to be updated using the `-a` command:
+### Starting with an OpenTree study and tree id and an alignment
+
+Alternatively, you can provide the gene alignment that you want to update using the `-a` command:
 
 ```
-physcraper_run_py -s OPENTREE_STUDY_ID -t OPENTREE_TREE_ID -o OUTPUT/DIRECTORY/NAME/AND/OR/PATH -a PATH/TO/GENE/ALIGNMENT/NAME -as ALIGNMENT_SCHEMA
+physcraper_run.py -s OPENTREE_STUDY_ID -t OPENTREE_TREE_ID -o OUTPUT/DIRECTORY/NAME/AND/OR/PATH -a PATH/TO/GENE/ALIGNMENT/NAME -as ALIGNMENT_SCHEMA
 
 ```
 
@@ -32,8 +36,10 @@ e.g.,  to update this tree [Crous et al. 2012](https://tree.opentreeoflife.org/c
 physcraper_run.py -s ot_350 -t Tr53297 -a docs/examples/inputdata/ot_350Tr53297.aln -as "nexus" -o ot_350
 ```
 
+### Starting with your own tree
+
 If the tree you want to update is not posted to the OpenTree website, you need to match
-the labels on your tree to taxa using [OpenTree Bulk Taxonomic Name Resolution Service](https://tree.opentreeoflife.org/curator/tnrs/). Download your matched names, unzip the folder, and pass the .json file as `--taxon_info` argument.
+the labels on your tree to taxa using [OpenTree Bulk Taxonomic Name Resolution Service](https://tree.opentreeoflife.org/curator/tnrs/). Download your matched names, unzip the folder, and pass the "json" file that is output from the OpenTree Bulk TNRS tool as `--taxon_info` argument.
 
 ```
 physcraper_run.py -tf TREE_FILE -tfs TREEFILE_SCHEMA -a ALIGNMENT_FILE -as ALIGNMENT_SCHEMA --taxon_info TAXON_INFO_JSONFILE  -o OUTPUT_DIRECTORY_NAME
@@ -45,8 +51,32 @@ e.g.,
 physcraper_run.py -tf tests/data/tiny_test_example/test.tre -tfs newick -a tests/data/tiny_test_example/test.fas  -as fasta --taxon_info tests/data/tiny_test_example/main.json -o owndata
 ```
 
[…]
