+++ docs/mds/Tutorial.md
@@ -11,7 +11,7 @@ If you have access to a single gene alignment, and a tree, Physcraper automates 
 <img src="https://raw.githubusercontent.com/McTavishLab/physcraper/main/docs/img/schematic.svg" title="schematic" alt="The Phsycraper Framework" width="100%" />
 
 The Physcraper framework consists of 4 general steps. The methodology is extensively described in the
-[Implementation](https://physcraper.readthedocs.io/en/latest/methods_extended.html) section.
+[Implementation](https://physcraper.readthedocs.io/en/latest/methods_extended.html) section of the documentation.
 
 <br/>
 
