+++ docs/mds/methods_extended.md
@@ -23,7 +23,7 @@
 <div id="tree-reconstruction-and-comparison" class="section level2">
 <h2>Tree reconstruction and comparison</h2>
 <p>A Maximum Likelihood (ML) gene tree is reconstructed for each alignment provided, using the software RAxML <span class="citation">(Stamatakis <a href="#ref-stamatakis2014raxml">2014</a>)</span> with default settings, such as a GTRCAT model of molecular evolution and 100 bootstrap replicates with the default algorithm. Currently only the number of bootsrap replicates can be specified by the user. By default, the original tree is used as a starting tree for the ML searches. Alternatively, users can set the original tree as a full topological constraint, or ignore it completely for the searches. Bootstrap results are summarized with the SumTrees module of DendroPy <span class="citation">(current version 4.4.0; Sukumaran &amp; Holder <a href="#ref-sukumaran2010dendropy">2010</a>)</span>.</p>
-<p>Physcraper’s final result is an updated phylogenetic hypothesis for the locus provided in the input alignment. Tips on all trees generated by Physcraper are defined by a taxon “name space”. The taxon metadata captures the NCBI accession information, as well as the taxon identifiers, allowing the user to perform comparisons and conflict analyses. Two ways to compare the updated tree with the original tree are implemented in Physcraper. First, Robinson Foulds weighted and unweighted metrics are estimated using Dendropy functions <span class="citation">(Sukumaran &amp; Holder <a href="#ref-sukumaran2010dendropy">2010</a>)</span>. Second, a conflict analysis is performed. This is a node by node comparison between the the synthetic OpenTree and the original and updated tree individually. This is performed with OpenTree’s conflict Application Programming Interface <span class="citation">(Redelings &amp; Holder <a href="#ref-redelings2017supertree">2017</a>)</span>. For the conflict analysis to be meaningful, the root of the tree needs to be accurately defined. A suggested default rooting based on OpenTree’s taxonomy is implemented for now. This approach uses the taxon labels for all the tips in the updated tree, pulls an inferred subtree from OpenTree’s taxonomy and then applies the same rooting to the inferred updated tree. However, if the updated tree changes expectations from taxonomy, the root may no longer be appropriate. Automatic identification of a phylogenetic tree root is indeed a difficult problem that has not been solved yet. The best way right now is for users to define outgroup directly on the updated tree, so trees are accurately rooted. <!-- *It would be a nice addition to have users give the output of the input tree as an argument at some point, or maybe we could add a super outgroup at random based on the search taxon* --></p>
[…]
