### 2021-02-04 f16c5e8
explain re-rooting
