### 2021-02-04 d662995
explain tree comparison.py
