### 2021-02-05 70024c2
edits to data exploration md
