+++ docs/mds/DataExploration.md
@@ -1,34 +1,52 @@
 ## Analysing the Physcraper results
 
+*Under construction*
+
+The functionalities described in the following sections are still under development.
+The first two functionalities are not accessible through the command line yet.
+This means that you can access them only through Python from now.
+The last functionality can be accessed through the command line, but it has not
+achieved final form yet.
+
+### Relabeling the trees
+
+For downstream analyses and figure making, it can be handy to swap name on tips of
+the updated phylogeny from alternative taxonomies or taxon id numbers.
+
+Do that with`write_labelled` function, e.g.,
+
+    from physcraper import treetaxon
+    pg55 = treetaxon.generate_TreeTax_from_run('docs/examples/pg_55_web')
+    pg55.write_labelled(label='^ot:ottTaxonName', norepeats=False, path='test_podarcis/repeats.tre')
+
 ### Rerooting the trees
 
-*In construction*
-
 A correctly rooted phylogeny is needed to compare relationships between two or more phylogenies.
-Rooting phylogenies can be tricky. Physcraper places a suggested root based on the taxonomic relationships in OpenTree, using this function.
+Rooting phylogenies can be tricky. Physcraper places a suggested root based on the taxonomic relationships in OpenTree, using the `root_tree_from_synth` function.
 
-To root a physcraper tree using either the OpenTree taxonomy, or the OpenTree synth tree, use `root_tree_from_synth`:
-e.g.
+To root a Physcraper tree using either the OpenTree taxonomy, or the OpenTree synthetic tree, first get the tree object:
+
     pg55 = treetaxon.generate_TreeTax_from_run('docs/examples/pg_55_web')
-    # To root based on taxonomy, set base = "ott"
+
+Then, to root based on taxonomy, set `base = "ott"`:
+
     ott_rooted_tree = physcraper.opentree_helpers.root_tree_from_synth(pg55.tre, pg55.otu_dict, base='ott')
     pg55.tre = ott_rooted_tree
     pg55.write_labelled(label="^ot:ottTaxonName", path="ott_root.tre")
 
 
-    # To root based on phylogenetic relationships in the synthetic tree, set base = "synth"
+And, to root based on phylogenetic relationships in the synthetic tree, set `base = "synth"`:
+
     synth_rooted_tree = physcraper.opentree_helpers.root_tree_from_synth(pg55.tre, pg55.otu_dict, base='synth')
     pg55.tre = synth_rooted_tree
     pg55.write_labelled(label="^ot:ottTaxonName", path="synth_root.tre")
 
 
-In this example both trees are the same even though they use the MRCA of differnet pairs of taxa, because those MRCA's map to the same node on the output tree.
+In this example both trees are the same even though they use the MRCA of different pairs of taxa, because those MRCA's map to the same node on the output tree.
 
+However rooting based on OTT can be unreliable, especially if taxonomy is a poor fit to true evolutionary relationships.
+So whenever possible, the root should be specified by the user, for example by choosing from tips in the "otu_dictionary" file, e.g.,
 
-However rooting based on ott can be unreliable, especially if taxonomy is a poor fit to true evolutionary relationships.
-So whenever possible, the root should be specified by the user, for example by choosing from tips in the otu_dictionary file.
-
-e.g 
     pg55 = treetaxon.generate_TreeTax_from_run('docs/examples/pg_55_web')
     outgroup = ['otu376436','otu376444']
     mrca = pg55.tre.mrca(taxon_labels=outgroup)
@@ -36,37 +54,30 @@ e.g
     pg55.write_labelled(label="^ot:ottTaxonName", path="manual_root.tre")
 
 
-### Tree comparison with Robinson-Foulds
[…]
