### 2021-02-05 e255dc4
some more edits to data exploration md
