### 2021-02-06 1b194cc
fleshing out finding alignments section
