### 2021-02-06 b65b388
some edit to homogenize commands
