+++ docs/mds/PhyscraperRun.md
@@ -6,63 +6,62 @@ A configuration file will be written down for the sake of reproducibility.
 
 ### Starting with only an OpenTree study and tree id
 
-As input, you will minimally need a study and tree ids from a tree uploaded to the Open Tree of Life website. The `--treebase` flag (or `-tb`) will automatically download an alignment for that tree from TreeBASE.
+As input, you will minimally need a study and tree ids from a tree uploaded to the OpenTree website (https://tree.opentreeoflife.org/curator). The `--treebase` flag (or `-tb`) will automatically download an alignment for that tree from TreeBASE.
 
-```
-physcraper_run.py -s OPENTREE_STUDY_ID -t OPENTREE_TREE_ID --treebase -o OUTPUT_DIRECTORY_NAME/AND/OR/PATH
-```
+
+    physcraper_run.py [-s OPENTREE_STUDY_ID] [-t OPENTREE_TREE_ID] [-tb] [-o OUTPUT]
+
 
 e.g.,
 
 
-```
-physcraper_run.py -s pg_55 -t tree5864 -tb -o pg_55_web
-```
+    physcraper_run.py -s pg_55 -t tree5864 -tb -o pg_55_web
 
-The output files generated by this run are stored in "docs/examples/pg_55_web"
+
+The output files generated by this example run are stored in "docs/examples/pg_55_web"
 
 ### Starting with an OpenTree study and tree id and an alignment
 
 Alternatively, you can provide the gene alignment that you want to update using the `-a` command:
 
-```
-physcraper_run.py -s OPENTREE_STUDY_ID -t OPENTREE_TREE_ID -o OUTPUT/DIRECTORY/NAME/AND/OR/PATH -a PATH/TO/GENE/ALIGNMENT/NAME -as ALIGNMENT_SCHEMA
 
-```
+    physcraper_run.py [-s OPENTREE_STUDY_ID] [-t OPENTREE_TREE_ID] [-o OUTPUT] [-a ALIGNMENT] [-as ALIGNMENT_SCHEMA]
 
-e.g.,  to update this tree [Crous et al. 2012](https://tree.opentreeoflife.org/curator/study/view/ot_350/?tab=home&tree=Tr53296) using an alignment already downloaded from TreeBASE.
 
-```
-physcraper_run.py -s ot_350 -t Tr53297 -a docs/examples/inputdata/ot_350Tr53297.aln -as "nexus" -o ot_350
-```
+
+For example, to update a tree from [Crous et al. 2012](https://tree.opentreeoflife.org/curator/study/view/ot_350/?tab=home&tree=Tr53296) using an alignment already downloaded from TreeBASE, you can do:
+
+
+    physcraper_run.py -s ot_350 -t Tr53297 -a docs/examples/inputdata/ot_350Tr53297.aln -as "nexus" -o ot_350
+
 
 ### Starting with your own tree
 
 If the tree you want to update is not posted to the OpenTree website, you need to match
-the labels on your tree to taxa using [OpenTree Bulk Taxonomic Name Resolution Service](https://tree.opentreeoflife.org/curator/tnrs/). Download your matched names, unzip the folder, and pass the "json" file that is output from the OpenTree Bulk TNRS tool as `--taxon_info` argument.
+the labels on your tree to taxa using the [OpenTree Bulk Taxonomic Name Resolution Service](https://tree.opentreeoflife.org/curator/tnrs/). Download your matched names, unzip the folder, and pass the "json" file that is output from the OpenTree Bulk TNRS tool as `--taxon_info` argument:
 
-```
[…]
