+++ physcraper/scrape.py
@@ -141,7 +141,7 @@ class PhyscraperScrape():
         self.map_taxa_to_ncbi()
         assert self.mrca_ncbi
         self.write_mrca()
-        self.data.write_labelled(filename="taxonname", label='^ot:ottTaxonName', direc=self.inputsdir)
+#        self.data.write_labelled(filename="taxonname", label='^ot:ottTaxonName', direc=self.inputsdir)
         self.data.write_otus(schema='table', direc=self.inputsdir)
         if not os.path.exists("{}/otu_info_{}.json".format(self.rundir, self.data.tag)):
             self.data.write_otus(schema='json', direc=self.rundir)
