+++ physcraper/scrape.py
@@ -141,7 +141,7 @@ class PhyscraperScrape():
         self.map_taxa_to_ncbi()
         assert self.mrca_ncbi
         self.write_mrca()
-#        self.data.write_labelled(filename="taxonname", label='^ot:ottTaxonName', direc=self.inputsdir)
+        self.data.write_labelled(filename="taxonname", label='^ot:ottTaxonName', direc=self.inputsdir)
         self.data.write_otus(schema='table', direc=self.inputsdir)
         if not os.path.exists("{}/otu_info_{}.json".format(self.rundir, self.data.tag)):
             self.data.write_otus(schema='json', direc=self.rundir)
@@ -347,8 +347,6 @@ class PhyscraperScrape():
                                     gb_acc = get_acc_from_blast(match)
                                     if gb_acc is not None:
                                         full_seq = self.get_full_seq(gb_acc, sseq)
-                                        if full_seq == None or full_seq == 'None':
-                                            full_seq = self.get_full_seq(gi_id, sseq)
                                         query_dict[gb_acc] = {'^ncbi:gi': gi_id,
                                                               'accession': gb_acc,
                                                               'staxids': staxids,
@@ -365,8 +363,6 @@ class PhyscraperScrape():
                                 if gb_acc not in self.ids.acc_ncbi_dict:  # fill up dict with more information.
                                     self.ids.acc_ncbi_dict[gb_acc] = staxids
                                 full_seq = self.get_full_seq(gb_acc, sseq)
-                                if full_seq == None or full_seq == 'None':
-                                    full_seq = self.get_full_seq(gi_id, sseq)
                                 query_dict[gb_acc] = {'^ncbi:gi': gi_id,
                                                       'accession': gb_acc,
                                                       'staxids': staxids,
