### 2021-02-18 5a979fe
pull run out of if
