+++ bin/concatenate_loci.py
@@ -10,6 +10,7 @@ import copy
 #physcraper_run.py -s pg_238 -t tree109 -a physcraper_example/treebase_alns/pg_238tree109_18S.aln -db /branchinecta/shared/local_blast_db -nt 8 -as "nexus" -r -o pg_238_runs/18s
 
 
+#python -i bin/concatenate_loci.py -d ../physcraper_runs/pg_238_runs/
 
 
 
@@ -19,6 +20,8 @@ parser = argparse.ArgumentParser()
 parser.add_argument("-d","--loci_runs_folder", help="folder containing results directories from individual locus runs")
 parser.add_argument("-o","--output", help="folder to write to")
 parser.add_argument("-c","--concatenate_by", help="what aspect of otu info to concatenate by, eitehr OTTid or ...?")
+parser.add_argument("-f","--format", help="output format", choices=["concatenate", "astral"])
+parser.add_argument("-s","--schema", help="alignment file format schema", choices=["fasta", "nexus"], default="fasta")
 parser.add_argument("-m","--include_missing", help="Where unevern numbers of sequences are available, concatenate with gaps", default = False)
 ## Include all seqs or just some???
 
@@ -37,7 +40,7 @@ tax_labels = {}
 
 otu_dict_dict = {}
 
-runs = [f for f in os.listdir(args.loci_runs_folder)]
+runs = [f for f in os.listdir(args.loci_runs_folder) if os.path.isdir("{}/{}".format(args.loci_runs_folder, f))]
 i = 0
 
 all_loci = {}
@@ -46,92 +49,114 @@ for run in runs:
     i += 1
     fp = "{}/{}".format(args.loci_runs_folder, run)
     files = [f for f in os.listdir(fp)]
-    for file in files:
-        if file.startswith('inputs_'):
-            tag = file.split('.')[0].replace('inputs_', '')
-    loc = "Locus{}".format(i)
-    print(loc)
-    locus = aligntreetax.generate_ATT_from_run(fp, start_files='output', tag=tag, run=False)
-    all_loci[loc] = {}
-    all_loci[loc]['data'] = locus
-    all_loci[loc]['path'] = fp
-    for seq in locus.aln:
-        otu_id = seq.label
-        taxon = locus.otu_dict[otu_id]['^ot:ottTaxonName']
-        tax_labels[taxon] = locus.otu_dict[otu_id]['^ot:ottTaxonName']
-        if taxon not in all_taxa:
-            all_taxa[taxon] = {}
-        if loc not in all_taxa[taxon]:
-            all_taxa[taxon][loc] = []
-        all_taxa[taxon][loc].append(otu_id)
+    if files:
+        for file in files:
+            if file.startswith('inputs_'):
+                tag = file.split('.')[0].replace('inputs_', '')
+        loc = "Locus{}".format(i)
+        print(loc)
+        locus = aligntreetax.generate_ATT_from_run(fp, start_files='output', tag=tag, run=False)
+        all_loci[loc] = {}
+        all_loci[loc]['data'] = locus
+        all_loci[loc]['path'] = fp
+        for seq in locus.aln:
+            otu_id = seq.label
+            taxon = locus.otu_dict[otu_id]['^ot:ottTaxonName']
+            tax_labels[taxon] = locus.otu_dict[otu_id]['^ot:ottTaxonName']
+            if taxon not in all_taxa:
+                all_taxa[taxon] = {}
+            if loc not in all_taxa[taxon]:
+                all_taxa[taxon][loc] = []
+            all_taxa[taxon][loc].append(otu_id)
 
 nloci = i
 
-
-#
 #ds.read(path="pythonidae_cytb.fasta", schema="fasta", data_type="dna")
[…]
