### 2021-03-08 251ab3d
fixing blanks and adding a docstring for the module
