+++ bin/find_trees.py
@@ -1,4 +1,28 @@
 #!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+##############################################################################
+##  Physcraper Library.
+##
+##
+##  If you use this work or any portion thereof in published work,
+##  please cite it as:
+##
+##     Luna L. Sanchez Reyes, Martha Kandziora, Emily Jane McTavish. (2020).
+##     Physcraper: A Python package for continual update of evolutionary
+##     estimates using the Open Tree of Life. BioRxiv 2020.09.15.299156.
+##     doi: doi.org/10.1101/2020.09.15.299156
+##
+##############################################################################
+
+"""
+This module handles the command line code to find trees in the Open Tree of Life
+Phylesystem database that contain a biological group of interest and/or that have
+data in TREEbase too.
+"""
+
+
+
 import sys
 import argparse
 from opentree import OT
@@ -6,15 +30,15 @@ from opentree import OT
 
 
 parser = argparse.ArgumentParser()
-parser.add_argument("-t","--taxon_name", help="Name of search taxon")
-parser.add_argument("-ott","--ott_id", help="Name of search taxon")
-parser.add_argument("-tb","--treebase", action = 'store_true', help="Rturn studies with treebase data only")
-parser.add_argument("-o","--output", help="Output file path")
+parser.add_argument("-t", "--taxon_name", help="Name of search taxon")
+parser.add_argument("-ott", "--ott_id", help="Name of search taxon")
+parser.add_argument("-tb", "--treebase", action='store_true', help="Return studies with treebase data only")
+parser.add_argument("-o", "--output", help="Output file path")
 args = parser.parse_args()
 
 
 args = parser.parse_args()
-if len(sys.argv)==1:
+if len(sys.argv) == 1:
     parser.print_help(sys.stderr)
     sys.exit(1)
 
@@ -33,7 +57,7 @@ if args.ott_id:
     ottid = args.ott_id
 
 sys.stdout.write("OTT id {}\n".format(ottid))
-phylesystem_studies_resp = OT.find_trees(ottid, search_property ='ot:ottId')
+phylesystem_studies_resp = OT.find_trees(ottid, search_property='ot:ottId')
 
 
 
@@ -59,7 +83,7 @@ for study in phylesystem_studies_resp.response_dict['matched_studies']:
     data_deposit_url = 'None'
     if data_deposit:
         data_deposit_url = data_deposit[u'@href']
-    studies[study_id]['data_deposit_url'] =  data_deposit_url
+    studies[study_id]['data_deposit_url'] = data_deposit_url
     if 'treebase' in data_deposit_url:
         treebase_studies.add(study_id)
     trees[study_id] = []
@@ -68,10 +92,10 @@ for study in phylesystem_studies_resp.response_dict['matched_studies']:
         trees[study_id].append(treeid)
     cites = []
     studies[study_id]['opentree_url'] = "https://tree.opentreeoflife.org/curator/study/view/{}".format(study_id)
-    studies[study_id]['reference'] = nexson['nexml'].get('^ot:studyPublicationReference', 'no ref') 
+    studies[study_id]['reference'] = nexson['nexml'].get('^ot:studyPublicationReference', 'no ref')
     studies[study_id]['doi'] = nexson['nexml'].get('^ot:studyPublication', 'no study pub')
     if args.treebase and study_id not in treebase_studies:
-            continue
+        continue
     cites_phyl = cites_phyl + "\nStudy {} tree(s) {}\n".format(study_id, ', '.join(trees[study_id]))
[…]
