### 2021-03-09 a5380ea
adding pylinting instructions to contributing documentation
