### 2021-03-09 40a2366
fixing bare except when taxon has no ott id
