### 2021-03-09 9eb0c5e
adding tree auto rooting test
