### 2021-03-10 5d2cf8d
fixing comparisons
