+++ bin/physcraper_run.py
@@ -3,67 +3,77 @@ import argparse
 import os
 import sys
 import physcraper
-from physcraper.opentree_helpers import get_tree_from_study, scraper_from_opentree, get_max_match_aln, generate_ATT_from_phylesystem, bulk_tnrs_load
+from physcraper.opentree_helpers import (get_tree_from_study,
+                                         scraper_from_opentree,
+                                         get_max_match_aln,
+                                         generate_ATT_from_phylesystem,
+                                         bulk_tnrs_load)
 from physcraper.aligntreetax import generate_ATT_from_run, generate_ATT_from_files
 
 parser = argparse.ArgumentParser()
-parser.add_argument("-s","--study_id", help="OpenTree study id")
-parser.add_argument("-t","--tree_id", help="OpenTree tree id")
-parser.add_argument("-a","--alignment", help="path to alignment")
-parser.add_argument("-as","--aln_schema", help="alignment schema (nexus or fasta)")
-parser.add_argument("-o","--output", help="path to output directory")
-parser.add_argument("-c","--configfile", help="path to config file")
-parser.add_argument("-tb","--treebase", action="store_true", help="download alignment from treebase")
-parser.add_argument("-re","--reload_files",  help="reload files and configuration from dir")
+parser.add_argument("-s","--study_id", help="OpenTree study id.")
+parser.add_argument("-t","--tree_id", help="OpenTree tree id.")
+parser.add_argument("-a","--alignment", help="Path to alignment.")
+parser.add_argument("-as","--aln_schema", help="Alignment schema (nexus or fasta).")
+parser.add_argument("-o","--output", help="Path to output directory.")
+parser.add_argument("-c","--configfile", help="Path to config file.")
+parser.add_argument("-tb","--treebase", action="store_true", help="Download alignment from treebase.")
+parser.add_argument("-re","--reload_files",  help="Reload files and configuration from dir.")
 
-parser.add_argument("-tf","--tree_file", help="a path to a tree file")
-parser.add_argument("-tfs","--tree_schema", help="tree file format schema")
-parser.add_argument("-ti","--taxon_info", help="taxon info file from OpenTree TNRS")
+parser.add_argument("-tf","--tree_file", help="A path to a tree file.")
+parser.add_argument("-tfs","--tree_schema", help="Tree file format schema.")
+parser.add_argument("-ti","--taxon_info", help="Taxon info file from OpenTree TNRS.")
 
 
 
-parser.add_argument("-tag","--tag", help="gene name or other specifier")
-parser.add_argument("-st","--search_taxon", help="taxonomic id to constrain blast search. format ott:123 or ncbi:123. Deafult will use ingroup of tree on OpenTree, or MRCA of input tips ")
+parser.add_argument("-tag","--tag", help="Gene name or other name specifier.")
+msg1 = "Taxonomic id to constrain a BLAST search. "
+msg2 = "Format is ott:123 or ncbi:123. Default will use ingroup of tree on OpenTree, or MRCA of input tips."
+parser.add_argument("-st","--search_taxon", help=msg1 + msg2)
+msg3 = "MAXimum "
+msg4 = "MINimum "
+parser.add_argument("-spn","--species_number", help=msg3 + "number of sequences to include per species.")
+parser.add_argument("-bl","--block_list", nargs='+', help="NCBI accession numbers to exclude.")
 
-parser.add_argument("-spn","--species_number", help="max number of seqs to include per species")
-parser.add_argument("-bl","--block_list", nargs='+', help="ncbi accession numbers to exclude")
+parser.add_argument("-tp","--trim_perc", help= msg4 + "percentage of sequences end of alignments.")
 
[…]
