### 2021-03-10 b7b1ace
linting assertions and long functions
