### 2021-03-10 337f5ef
updating modules docstring
