+++ bin/find_trees.py
@@ -16,9 +16,9 @@
 ##############################################################################
 
 """
-This module handles the command line code to find trees in the Open Tree of Life
-Phylesystem database that contain a biological group of interest and/or that have
-data in TREEbase too.
+This module handles the code to genefrate the Physcraper command line code interface
+to find trees in the Open Tree of Life Phylesystem database that contain a biological
+group of interest and/or that have data in TREEbase too.
 """
 
 
