### 2021-03-11 916668d
linting undeclared vars
