### 2021-03-11 7c6aa8d
linting set comprehension and bare except
