### 2021-03-11 6b63da0
linting docstrings
