+++ physcraper/aligntreetax.py
@@ -1,6 +1,8 @@
-"""AlignTreeTax: The core data object for physcraper.
-Holds and links name spaces for a tree, an alignment and
-the taxa and their metadata"""
+"""
+AlignTreeTax: The core data object for Physcraper.
+Holds and links name spaces for a tree, an alignment,
+the taxa and their metadata.
+"""
 
 import sys
 import re
@@ -30,7 +32,8 @@ def generate_ATT_from_files(workdir,
                             otu_json,
                             tree_schema,
                             search_taxon=None):
-    """Build an ATT object without phylesystem, use your own files instead.
+    """
+    Build an ATT object without phylesystem, use your own files instead.
 
     Spaces vs underscores kept being an issue, so all spaces are coerced to underscores when data are read in.
 
@@ -76,7 +79,8 @@ def generate_ATT_from_files(workdir,
                         configfile=configfile, aln_schema=aln_schema, tree_schema=tree_schema)
 
 def generate_ATT_from_run(workdir, start_files='output', tag=None, configfile=None):
-    """Build an ATT object without phylesystem, use your own files instead.
+    """
+    Build an ATT object without phylesystem, use your own files instead.
     :return: object of class ATT
     """
     files = [f for f in os.listdir(workdir)]
@@ -127,12 +131,13 @@ def generate_ATT_from_run(workdir, start_files='output', tag=None, configfile=No
                         tree_schema='newick')
 
 def write_otu_file(treetax, filepath, schema="table"):
-    """Writes out OTU dict as json or table.
-        :param treetax: eitehr a treetaxon object or an alignment tree taxon object
-        :param filename: filename
-        :param schema: either table or json format
-        :return: writes out otu_dict to file
-        """
+    """
+    Writes out OTU dict as json or table.
+    :param treetax: eitehr a treetaxon object or an alignment tree taxon object
+    :param filename: filename
+    :param schema: either table or json format
+    :return: writes out otu_dict to file
+    """
     if schema == "json":
         with open(filepath, "w") as outfile:
             json.dump(treetax.otu_dict, outfile)
@@ -160,12 +165,13 @@ def write_otu_file(treetax, filepath, schema="table"):
 
 
 def write_labelled_tree(treetax, label, filepath, schema="newick", norepeats=True, add_gb_id=False):
-    """output tree and alignment with human readable labels
+    """
+    Output tree and alignment with human readable labels.
     Jumps through a bunch of hoops to make labels unique.
 
     NOT MEMORY EFFICIENT AT ALL
 
-    Has different options available for different desired outputs
+    Has different options available for different desired outputs.
 
     :param label: which information shall be displayed in labelled files: possible options:
                 '^ot:ottTaxonName', '^user:TaxonName', "^ot:originalLabel", "^ot:ottId", "^ncbi:taxon"
@@ -205,12 +211,13 @@ def write_labelled_tree(treetax, label, filepath, schema="newick", norepeats=Tru
 
 
 def write_labelled_aln(aligntreetax, label, filepath, schema="fasta", norepeats=True, add_gb_id=False):
-    """output tree and alignment with human readable labels
+    """
+    Output tree and alignment with human readable labels.
     Jumps through a bunch of hoops to make labels unique.
 
     NOT MEMORY EFFICIENT AT ALL
 
-    Has different options available for different desired outputs
+    Has different options available for different desired outputs.
 
     :param label: which information shall be displayed in labelled files: possible options:
                 '^ot:ottTaxonName', '^user:TaxonName', "^ot:originalLabel", "^ot:ottId", "^ncbi:taxon"
@@ -246,80 +253,72 @@ def write_labelled_aln(aligntreetax, label, filepath, schema="fasta", norepeats=
 
[…]
