### 2021-03-11 eb0862c
linting set comprehension
