+++ physcraper/configobj.py
@@ -1,4 +1,6 @@
-""" Physcraper run Configuration object generator"""
+"""
+Physcraper run Configuration object generator
+"""
 import sys
 import os
 import datetime
@@ -17,7 +19,7 @@ PHYSCRAPER_DIR = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
 
 
 def is_number(inputstr):
-    """test if string can be coerced to float"""
+    """Test if string can be coerced to float"""
     try:
         float(inputstr)
         return True
@@ -244,7 +246,8 @@ max_length = {maxlen}
         if not os.path.exists("{}/nt.23.nhr".format(self.blastdb)):
             sys.stderr.write("Errors with local Blast DB at {}, \
                 may be incomplete. please use a remote search, \
-                or update as described in 'https://physcraper.readthedocs.io/en/latest/install.html#local-databases'\n".format(self.blastdb))
+                or update as described in \
+                'https://physcraper.readthedocs.io/en/latest/install.html#local-databases'\n".format(self.blastdb))
             sys.exit()
         else:
             download_date = os.path.getmtime("{}/nt.23.nhr".format(self.blastdb))
@@ -255,10 +258,12 @@ max_length = {maxlen}
                 sys.stderr.write("Your databases might not be up to date anymore. \
                     You downloaded them {} days ago. \
                     Continuing, but perhaps use a remote search, \
-                    or update as decribed in 'https://physcraper.readthedocs.io/en/latest/install.html#local-databases'\n".format(time_passed))
+                    or update as decribed in \
+                    'https://physcraper.readthedocs.io/en/latest/install.html#local-databases'\n".format(time_passed))
         if not os.path.exists(self.ncbi_nodes):
             sys.stderr.write("NCBI taxonomy not found at {} - please update nodes and names.dmp, \
-                as described in 'https://physcraper.readthedocs.io/en/latest/install.html#local-databases'\n".format(self.ncbi_nodes))
+                as described in \
+                'https://physcraper.readthedocs.io/en/latest/install.html#local-databases'\n".format(self.ncbi_nodes))
             sys.exit()
         else:
             download_date = os.path.getmtime(self.ncbi_nodes)
@@ -268,6 +273,7 @@ max_length = {maxlen}
             if time_passed >= 90:
                 sys.stderr.write("Your taxonomy databases from NCBI were dowloaded {} days ago. \
                     Please update nodes and names.dmp, \
-                    as described in 'https://physcraper.readthedocs.io/en/latest/install.html#local-databases'\n".format(time_passed))
+                    as described in \
+                    'https://physcraper.readthedocs.io/en/latest/install.html#local-databases'\n".format(time_passed))
         assert(shutil.which("blastn")), "blastn  not found in path"
         self.url_base = None
