### 2021-03-11 f3ec86b
linting unidiomatic type check
