+++ physcraper/ncbi_data_parser.py
@@ -231,7 +231,7 @@ class Parser:
         #        debug("get downtorank")
         if nodes is None:
             self.initialize()
-        if type(tax_id) != int:
+        if not isinstance(tax_id, int):
            # sys.stdout.write(
            #     "WARNING: tax_id {} is no integer. Will convert value to int\n".format(
            #         tax_id
@@ -242,11 +242,9 @@ class Parser:
         # following statement is to get id of taxa if taxa is higher ranked than specified
         if nodes[nodes["tax_id"] == tax_id]["rank"].values[0] != "species":
             if downtorank == "species":
-                if (
-                    nodes[nodes["tax_id"] == tax_id]["rank"].values[0] != "varietas"
-                    and nodes[nodes["tax_id"] == tax_id]["rank"].values[0]
-                    != "subspecies"
-                ):
+                if (nodes[nodes["tax_id"] == tax_id]["rank"].values[0] != "varietas"
+                        and nodes[nodes["tax_id"] == tax_id]["rank"].values[0]
+                        != "subspecies"):
                     return tax_id
         if nodes[nodes["tax_id"] == tax_id]["rank"].values[0] == downtorank:
             # debug("found right rank")
