### 2021-03-11 d6c805b
linting set comprehension, singleton comparison and unidiomatic typecheck
