### 2021-03-11 49bcfad
linting docstrings, imports and lines
