### 2021-03-11 7ded9b1
linting set comprehension, singleton expr and format strings
