### 2021-03-11 12594c2
linting set comprehension, singleton expr and format strings
