### 2021-03-11 140f4d2
doc on how to count functions, classes and methods
