+++ bin/concatenate_loci.py
@@ -27,31 +27,31 @@ from physcraper import aligntreetax
 parser = argparse.ArgumentParser()
 parser.add_argument("-d","--locus_runs_folder", help="folder containing results directories from individual locus runs")
 parser.add_argument("-o","--output", help="folder to write to")
-parser.add_argument("-c","--concatenate_by", help="what aspect of otu info to concatenate by, eitehr OTTid or ...?")
+#parser.add_argument("-c","--concatenate_by", help="what aspect of otu info to concatenate by, eitehr OTTid or ...?")
 parser.add_argument("-f","--format", help="output format", choices=["concatenate", "astral"])
 parser.add_argument("-s","--schema", help="alignment file format schema", choices=["fasta", "nexus"], default="fasta")
-parser.add_argument("-m","--include_missing", help="Where unevern numbers of sequences are available, concatenate with gaps", default = False)
+parser.add_argument("-m","--include_missing", help="Where uneven numbers of sequences are available, concatenate with gaps", default = False)
 ## Include all seqs or just some???
-
 
 args = parser.parse_args()
 
-
-
-all_taxa = {}
-
-#args.output = "test_concat_res"
 if not os.path.exists(args.output):
     os.mkdir(args.output)
 
+
+# Track what taxa are present across all loci
+all_taxa = {}
+
+# find otu_ids by taxon
 tax_labels = {}
 
-otu_dict_dict = {}
-
+# Runs is the set of completed physcraper runs. 
+# They should all be in the folder given by -d aka locus_runs_folder
 runs = [f for f in os.listdir(args.locus_runs_folder) if os.path.isdir("{}/{}".format(args.locus_runs_folder, f))]
-i = 0
 
-all_loci = {}
+
+i = 0 #iterator to count runs (loci)
+all_loci = {} # dictionary of ATT objects
 for run in runs:
     print(run)
     i += 1
@@ -61,16 +61,15 @@ for run in runs:
         for file in files:
             if file.startswith('inputs_'):
                 tag = file.split('.')[0].replace('inputs_', '')
-        loc = "Locus{}".format(i)
+        loc = "Locus{}_{}".format(i, tag)
         print(loc)
-        locus = aligntreetax.generate_ATT_from_run(fp, start_files='output', tag=tag, run=False)
+        locus = aligntreetax.generate_ATT_from_run(fp, start_files='output', tag=tag, run=False) #build ATT opject for each run
         all_loci[loc] = {}
-        all_loci[loc]['data'] = locus
+        all_loci[loc]['data'] = locus #all_loci[loc]['data'] has an alignemnt, and otuu_dictionary, and a tree
         all_loci[loc]['path'] = fp
         for seq in locus.aln:
             otu_id = seq.label
             taxon = locus.otu_dict[otu_id]['^ot:ottTaxonName']
-            tax_labels[taxon] = locus.otu_dict[otu_id]['^ot:ottTaxonName']
             if taxon not in all_taxa:
                 all_taxa[taxon] = {}
             if loc not in all_taxa[taxon]:
@@ -79,46 +78,37 @@ for run in runs:
 
 nloci = i
 
-#ds.read(path="pythonidae_cytb.fasta", schema="fasta", data_type="dna")
-#ds.read(schema="pythonidae.mle.tre", "nexus", taxon_namespace=ds.taxon_namespaces[0])
-#ds.write_to_path("pythonidae_combined.nex", "nexus")
-
-
 
 if args.format == "concatenate":
     concat_dict = {}
-    re_label_dict = {}
+    re_label_dict = {} # What are we calling each concatneated taxon, and what OTUs are part of it.
     for loc in all_loci:
[…]
