### 2021-03-18 493e426
pylint defining 'synonyms' global var at the module level
