+++ docs/mds/MultiLoci.md
@@ -1,1 +1,13 @@
-Single locus analyses only porvide a narrow view of teh evolutionary history of a group.
+Single locus analyses only provide a narrow view of the evolutionary history of a group.
+
+After assembling individual gene data sets and phylogenies using Physcraper,
+it is straigtforward to combine the data from those analyses to generate species tree estimates.
+
+
+# Astral
+
+
+# Concatenation
+
+
+# SVD quartets
