### 2021-03-22 41295b8
fixing pylint potentially undefined loop variable 'node'
