+++ physcraper/scrape.py
@@ -272,7 +272,8 @@ class PhyscraperScrape():
             else:
                 time_passed = abs((datetime.datetime.strptime(today, "%Y/%m/%d") -
                                    datetime.datetime.strptime(last_blast, "%Y/%m/%d")).days)
-            if not self.data.otu_dict[otu_id].get("^physcraper:ingroup"):
+            if self.data.otu_dict[otu_id].get("^physcraper:ingroup") == False: # pylint: disable
+            #this should stay == False, because sometimes ingroup status is just unknown 
                 sys.stdout.write("tip {} not in ingroup. Will not blast, \n".format(otu_id))
                 continue
             if time_passed > delay:
