+++ physcraper/scrape.py
@@ -418,7 +418,7 @@ class PhyscraperScrape():
             except subprocess.CalledProcessError as grepexc:
                 sys.stderr.write("error code {}, {},\n".format(grepexc.returncode, grepexc.output))
                 sys.stderr.write("downloading from NCBI\n")
-                read_handle = self.ids.entrez_efetch(gb_id)
+                read_handle = Entrez.efetch(db="nucleotide", id=gb_id, retmode="xml")
                 seq = read_handle[0][u'GBSeq_sequence']
                 with open(seq_path, 'w') as fi:
                     fi.write("> {} \n".format(gb_id))
