### 2021-03-23 e9f3f6f
pylint bare except IndexError
