### 2021-03-23 ae24f24
fixing some more zombie code part 2
