### 2021-03-23 888f489
fixing indentation and undefined variable from zombie code
