+++ bin/multi_loci.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+#!/usr/bin/env python
 import dendropy
 import argparse
 import os
@@ -28,12 +29,16 @@ parser = argparse.ArgumentParser()
 parser.add_argument("-d","--locus_runs_folder", help="folder containing results directories from individual locus runs")
 parser.add_argument("-o","--output", help="folder to write to")
 #parser.add_argument("-c","--concatenate_by", help="what aspect of otu info to concatenate by, eitehr OTTid or ...?")
-parser.add_argument("-f","--format", help="output format", choices=["concatenate", "astral"])
-parser.add_argument("-s","--schema", help="alignment file format schema", choices=["fasta", "nexus"], default="fasta")
-parser.add_argument("-m","--include_missing", help="Where uneven numbers of sequences are available, concatenate with gaps", default = False)
+parser.add_argument("-f","--format", help="output format", choices=["concatenate", "astral", "svdq"])
+parser.add_argument("-s","--schema", help="output alignment file format schema", choices=["fasta", "nexus"], default="fasta")
+parser.add_argument("-m","--include_missing", action="store_true", help="For concatenation, where uneven numbers of sequences are available, concatenate with gaps", default = False)
 ## Include all seqs or just some???
 
 args = parser.parse_args()
+
+
+assert args.output, "-o output folder is required"
+assert args.locus_runs_folder, "-d input folder with completed runs is required"
 
 if not os.path.exists(args.output):
     os.mkdir(args.output)
@@ -49,44 +54,40 @@ tax_labels = {}
 # They should all be in the folder given by -d aka locus_runs_folder
 runs = [f for f in os.listdir(args.locus_runs_folder) if os.path.isdir("{}/{}".format(args.locus_runs_folder, f))]
 
-
-i = 0 #iterator to count runs (loci)
-all_loci = {} # dictionary of ATT objects
-for run in runs:
-    print(run)
-    i += 1
-    fp = "{}/{}".format(args.locus_runs_folder, run)
-    files = [f for f in os.listdir(fp)]
-    if files:
-        for file in files:
-            if file.startswith('inputs_'):
-                tag = file.split('.')[0].replace('inputs_', '')
-        loc = "Locus{}_{}".format(i, tag)
-        print(loc)
-        locus = aligntreetax.generate_ATT_from_run(fp, start_files='output', tag=tag, run=False) #build ATT opject for each run
-        all_loci[loc] = {}
-        all_loci[loc]['data'] = locus #all_loci[loc]['data'] has an alignemnt, and otuu_dictionary, and a tree
-        all_loci[loc]['path'] = fp
-        for seq in locus.aln:
-            otu_id = seq.label
-            taxon = locus.otu_dict[otu_id]['^ot:ottTaxonName']
-            if taxon not in all_taxa:
-                all_taxa[taxon] = {}
-            if loc not in all_taxa[taxon]:
-                all_taxa[taxon][loc] = []
-            all_taxa[taxon][loc].append(otu_id)
-
-nloci = i
+def setup_dicts(runs):
+    i = 0 #iterator to count runs (loci)
+    all_loci = {} # dictionary of ATT objects
+    for run in runs:
+        print(run)
+        i += 1
+        fp = "{}/{}".format(args.locus_runs_folder, run)
+        files = [f for f in os.listdir(fp)]
+        if files:
+            for file in files:
+                if file.startswith('inputs_'):
+                    tag = file.split('.')[0].replace('inputs_', '')
+            loc = "Locus{}_{}".format(i, tag)
+            locus = aligntreetax.generate_ATT_from_run(fp, start_files='output', tag=tag, run=False) #build ATT opject for each run
+            all_loci[loc] = {}
[…]
