+++ bin/multi_loci.py
@@ -1,37 +1,33 @@
 #!/usr/bin/env python
-#!/usr/bin/env python
-import dendropy
+""" 
+Example using the data from https://github.com/McTavishLab/physcraper_example
+physcraper_run.py -s pg_238 -t tree109 -a physcraper_example/treebase_alns/pg_238tree109_18S.aln -db /branchinecta/shared/local_blast_db -nt 8 -as "nexus" -r -o pg_238_runs/18s
+physcraper_run.py -s pg_238 -t tree109 -a physcraper_example/treebase_alns/pg_238tree109_18S.aln -db /branchinecta/shared/local_blast_db -nt 8 -as "nexus" -r -o pg_238_runs/18s
+
+Examples:
+python bin/multi_loci.py -d ../physcraper_runs/pg_238_runs/ -f concatenate -o test_concat_res
+ra
+
+python bin/multi_loci.py -d ../physcraper_runs/pg_238_runs/ -f astral -o test_concat_res
+java -jar astral.5.7.5.jar -i ../physcraper/test_concat_res/genetrees.new -a ../physcraper/test_concat_res/mappings.txt
+
+"""
 import argparse
 import os
 import copy
+import dendropy
 
 from physcraper import aligntreetax
-
-# Example using the data from https://github.com/McTavishLab/physcraper_example
-#physcraper_run.py -s pg_238 -t tree109 -a physcraper_example/treebase_alns/pg_238tree109_18S.aln -db /branchinecta/shared/local_blast_db -nt 8 -as "nexus" -r -o pg_238_runs/18s
-#physcraper_run.py -s pg_238 -t tree109 -a physcraper_example/treebase_alns/pg_238tree109_18S.aln -db /branchinecta/shared/local_blast_db -nt 8 -as "nexus" -r -o pg_238_runs/18s
-
-#Examples
-#python bin/concatenate_loci.py -d ../physcraper_runs/pg_238_runs/ -f concatenate -o test_concat_res
-
-
-
-#python bin/concatenate_loci.py -d ../physcraper_runs/pg_238_runs/ -f astral -o test_concat_res
-#java -jar astral.5.7.5.jar -i ../physcraper/test_concat_res/genetrees.new -a ../physcraper/test_concat_res/mappings.txt 
-
-
-#ToDo - clean up dendropy part, write Charsets
-# Generate taxon sets for SVD quratets inputs
 
 
 
 parser = argparse.ArgumentParser()
-parser.add_argument("-d","--locus_runs_folder", help="folder containing results directories from individual locus runs")
-parser.add_argument("-o","--output", help="folder to write to")
+parser.add_argument("-d", "--locus_runs_folder", help="folder containing results directories from individual locus runs")
+parser.add_argument("-o", "--output", help="folder to write to")
 #parser.add_argument("-c","--concatenate_by", help="what aspect of otu info to concatenate by, eitehr OTTid or ...?")
-parser.add_argument("-f","--format", help="output format", choices=["concatenate", "astral", "svdq"])
-parser.add_argument("-s","--schema", help="output alignment file format schema", choices=["fasta", "nexus"], default="fasta")
-parser.add_argument("-m","--include_missing", action="store_true", help="For concatenation, where uneven numbers of sequences are available, concatenate with gaps", default = False)
+parser.add_argument("-f", "--format", help="output format", choices=["concatenate", "astral", "svdq"])
+parser.add_argument("-s", "--schema", help="output alignment file format schema", choices=["fasta", "nexus"], default="fasta")
+parser.add_argument("-m", "--include_missing", action="store_true", help="For concatenation, where uneven numbers of sequences are available, concatenate with gaps", default=False)
[…]
