+++ tests/test_script_run.py
@@ -51,7 +51,7 @@ def test_comparison():
                             "-o", "tests/tmp/pg_55_comparison"])
 
 def test_concat():
-    subprocess.check_call(["python", "bbin/multi_loci.py",
+    subprocess.check_call(["python", "bin/multi_loci.py",
                             "-d", "tests/data/precooked/multi_loc/",
                             "-f", "astral",
                             "-o", "tests/tmp/mini_astral"
