### 2021-03-25 ae1c4e7
write json otu_info to inputs
