+++ physcraper/aligntreetax.py
@@ -99,7 +99,7 @@ def generate_ATT_from_run(workdir, start_files='output', tag=None, configfile=No
     try:
         alnfi = "{}/physcraper_{}.fas".format(sdir, tag)
         treefile = "{}/physcraper_{}.tre".format(sdir, tag)
-        otu_json = "{}/otu_info_{}.json".format(rundir, tag)
+        otu_json = "{}/otu_info_{}.json".format(sdir, tag)
         assert os.path.exists(alnfi)
         assert os.path.exists(treefile)
         assert os.path.exists(otu_json)
@@ -111,7 +111,7 @@ def generate_ATT_from_run(workdir, start_files='output', tag=None, configfile=No
             sys.stdout.write("No output files found in {}, loading files from {}\n".format(outputsdir, inputsdir))
             alnfi = "{}/physcraper_{}.fas".format(inputsdir, tag)
             treefile = "{}/physcraper_{}.tre".format(inputsdir, tag)
-            otu_json = "{}/otu_info_{}.json".format(rundir, tag)
+            otu_json = "{}/otu_info_{}.json".format(inputsdir, tag)
             assert os.path.exists(alnfi)
             assert os.path.exists(treefile)
             assert os.path.exists(otu_json)
