### 2021-03-25 10bfced
trying some excepts on cd function + more docs on it
