### 2021-03-26 5a61b19
multiple excepts did not work, disabling bare except for pylint
