### 2021-03-26 067e5ea
fixing pylint bare excepts
