### 2021-03-26 a20b3b1
fixing pylint bare excepts
