### 2021-03-26 405c6d6
pylinting white space, importing Entrez and disabling singleton comparison
