### 2021-03-26 3457760
pylint removing unused variables
