### 2021-03-26 c00a8ef
pylint-disabling on the whole module
