### 2021-03-26 e757670
pylint removing trailing stuff
