+++ bin/multi_loci.py
@@ -16,6 +16,10 @@ python bin/multi_loci.py -d ../physcraper_runs/pg_238_runs/ -f astral -o test_co
 java -jar astral.5.7.5.jar -i ../physcraper/test_concat_res/genetrees.new -a ../physcraper/test_concat_res/mappings.txt
 
 """
+
+# Disabling the following globally for this module
+# pylint: disable=redefined-outer-name, protected-access, too-many-branches, too-many-statements
+
 import argparse
 import os
 import copy
