### 2021-03-26 42c729b
returning back necessary unused variables and pylint disabling them
