+++ tests/test_fromfile.py
@@ -15,8 +15,8 @@ def test_generate_ATT_from_files():
     configfi = "tests/data/test.config"
 
     sys.stdout.write("\nTesting 'generate_ATT_from_files (fromfile.py)'\n")
-    data_obj = generate_ATT_from_files(alnfile=seqaln, 
-                                     aln_schema=mattype, 
+    data_obj = generate_ATT_from_files(alnfile=seqaln,
+                                     aln_schema=mattype,
                                      workdir=workdir,
                                      configfile=configfi,
                                      treefile=treefile,
@@ -32,90 +32,89 @@ def test_generate_ATT_from_run():
 
     sys.stdout.write("\nTesting 'generate_ATT_from_run '\n")
     data_obj = generate_ATT_from_run(workdir=workdir)
-  
+
 
 def test_example():
-    indir="docs/examples/pg_55_web"
-    workdir = "tests/tmp/example_test"
-    data_obj = generate_ATT_from_run(workdir=indir, start_files = "input")
-    data_obj.workdir = workdir
-    ids = IdDicts(data_obj.config)
-    scraper = PhyscraperScrape(data_obj, ids)
-    scraper.read_blast_wrapper("docs/examples/pg_55_web/blast_run_pg_55tree5864/")
-    assert(len(scraper.new_seqs_otu_id) == 3)
-
-
-"""
-def test_webbased_blast():
-    indir="docs/examples/pg_55_web"
-    workdir = "tests/tmp/example_test"
-    data_obj = generate_ATT_from_run(workdir=indir, start_files = "input")
-    data_obj.workdir = workdir
-    ids = IdDicts(data_obj.config)
-    scraper = PhyscraperScrape(data_obj, ids)
-    scraper.read_webbased_blast_query('docs/examples/pg_55_web/blast_run_pg_55tree5864/otu376425.xml')
-    len(scraper.new_seqs)
-    scraper.read_blast_wrapper('docs/examples/pg_55_web/blast_run_pg_55tree5864')
-
-def test_local_blast():
-    indir="docs/examples/pg_55_web"
-    workdir = "tests/tmp/example_loc_test"
-    data_objloc = generate_ATT_from_run(workdir=indir, start_files = "input")
-    data_objloc.workdir = workdir
-    data_objloc.config.blastdb = "/home/ejmctavish/ncbi/localblastdb"
-    data_objloc.config.set_local()
-    idsloc = IdDicts(data_obj.config)
-    idsloc.ncbi_parser = ncbi_data_parser.Parser(names_file=data_objloc.config.ncbi_names,
-                            nodes_file=data_objloc.config.ncbi_nodes)
-    scraperloc = PhyscraperScrape(data_objloc, idsloc)
-    scraperloc.read_local_blast_query('docs/examples/pg_55_local/blast_run_pg_55tree5864_ndhf/otu376425.txt')
-    scraperloc.read_blast_wrapper('docs/examples/pg_55_local/blast_run_pg_55tree5864_ndhf')
-
-len(scraperloc.new_seqs)
-
-"""
-
-
-
-"""
-def test_check_complement():
 indir="docs/examples/pg_55_web"
 workdir = "tests/tmp/example_test"
 data_obj = generate_ATT_from_run(workdir=indir, start_files = "input")
 data_obj.workdir = workdir
 ids = IdDicts(data_obj.config)
 scraper = PhyscraperScrape(data_obj, ids)
-scraper.data.gb_dict = {}
-result_handle = open('docs/examples/pg_55_web/blast_run_pg_55tree5864/otu376425.xml')
-blast_records = NCBIXML.parse(result_handle)
-i = 0
-ok = 0
-for blast_record in blast_records:
-    i += 1
-    print(i)
[…]
