### 2021-03-26 98723d7
pylint disabling no self use
