### 2021-03-26 202b7f7
iterating the new_seqs_otu_id dictionary directly
