### 2021-03-28 #26 @leomrtns
- [X] **Function Documentation:** for all user-facing functions
- [ ] **Examples** for all user-facing functions
- [X] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.

Readme requirements

- [X] Badges for continuous integration and test coverage, a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the badge for pyOpenSci peer-review will be provided upon acceptance.)
- [X] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
- [X] Any additional setup required (authentication tokens, etc)
- [ ] Brief demonstration usage
- [X] Direction to more detailed documentation (e.g. your documentation files or website).
- [ ] If applicable, how the package compares to other similar packages and/or how it relates to other packages

Package structure should follow general community best-practices. In general please consider:

- [X] The documentation is easy to find and understand

- [X] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

8

---

I will focus my comments from the point of view of a nonspecialist end user. I had no problem installing the software, however I could not complete the analysis since running from the OToL took too long, and my trees did not seem appropriate (several potentially from same species, without easy taxonomic mapping). Please keep in mind that most (all) my comments can be addressed by changing the documentation, since they may reflect a misunderstanding on my part.

* My overall impression is that the requested familiarity with open ToL, treeBASE, and TNRS might hinder a broader  usability of the tool. If the user is unfamiliar with these, they will really consider the learning curve in order to  decide if using this tool is worth; if they are familiar, they are probably doing it without `physcraper` already.  My suggestion is that you must show how the tool can be useful even if one is does not want to familiarise with the above mentioned tools.

* In this sense I see several alternative tools, that although doing phylogenetics from scratch also returns an 'updated' phylogeny (in less than 12 hours perhaps?):

  - https://github.com/AstrobioMike/GToTree
  - https://github.com/ropensci/phylotaR
  - https://github.com/dongzhang0725/PhyloSuite
  - https://lmap-s.sourceforge.io/
  - https://github.com/dportik/SuperCRUNCH
  - https://github.com/rdcrawford/cognac
  - https://github.com/yangfangs/physpetools

* Alternatively to the discussion above, it is fair and absolutely fine to describe `physcraper` as a tool within the open ToL ecosystem instead of a tool for the nonspecialist. From my understanding physcraper is great at automating OToL tasks   (as in continuous updating), but a disclaimer or an observation at the begining (of the documentation and README) would be necessary.   The objective here is to manage user expectation, and I will describe more below about hidden assumptions and knowledge.

* When using the term "homologous", maybe the authors mean "orthologous"? I understand this is a single-alignment tool but since you rely on taxonomic classifications I think that the tool is not ready for paralogous sequences.   In  other words, if you allow several tips to map to the same taxon (with a further indication of gene/locus name perhaps)  then you indeed accept homologous sequences (orthologs **and** paralogs). The documentation mentions "single locus alignment", and thus you should prefer the term ortholog/orthologous whenever possible (or mention that this   distinction is deferred to the study depositing the data, maybe? I am unaware of whole gene family data sets on OToL).   If `physcraper` does indeed handle all homologous sequences (as suggested in https://physcraper.readthedocs.io/en/latest/how_to_start.html)   then you need to explain how it reconciles the gene (locus) tree with the species tree, or how it decides between paralog alternatives.   Still I think it's safer to replace "homologous" by "orthologous" therefore implying that the responsability for  disambiguation is on the original publications, and not `physcraper`.

* The need for mapping the tree tip names to a broader taxonomy may be a serious impediment to the nonspecialist,  since in many analyses the taxonomical classification is a conclusion, and not an assumption of the study.  In which case a tool capable of updating my tree and alignment with data from known studies would be most welcome! Furthermore filling the mapping table by hand is impractical except for toy examples (with less than a few dozen taxa).  (https://physcraper.readthedocs.io/en/latest/how_to_start.html#updating-your-own-tree-and-alignment). I've seen in another piece of documentation that this might not be necessary, I can't find it now... So could this be made clearer?

* In the [Results section of the documentation](https://physcraper.readthedocs.io/en/latest/data_exploration.html) it seems I need to know python (and use `physcraper`'s module) to be able to relabel the tree. Which is not exactly user-friendly or useful to a general, nonspecialist audience. I see no reason not to save several files, one of them with easily interpretable names. The same goes for the [scary mention of _rewriting_ of input files](https://physcraper.readthedocs.io/en/stable/data_exploration.html#an-overview-of-the-physcraper-results-files) &mdash;I am sure you mean that you copy the input files with modification somewhere else?

**minor comments:**

* How does it handle polytomies, are all trees in OToL fully resolved? Furthermore, is it possible to know when there are conflicting trees across studies? This is an OToL question actually, but still it would be good to be adressed in the documentation here.

* Some `readthedocs` internal links are not working (for instance the text mentions https://physcraper.readthedocs.io/en/latest/data_exploration.html#Relabeling but the correct link is  https://physcraper.readthedocs.io/en/latest/data_exploration.html#id1 ). Furthermore in  https://physcraper.readthedocs.io/en/latest/methods_extended.html, `Fig. @ref(fig:framework)` in the documentation is missing
