### 2021-03-29 43d1908
commenting out bare except
