### 2021-03-29 efad472
initializing attributes
