+++ physcraper/aligntreetax.py
@@ -423,7 +423,7 @@ class AlignTreeTax():
         for leaf in self.tre.leaf_nodes():
             treed_tax.add(leaf.taxon)
         aln_tax = set()
-        for tax, seq in self.aln.items():
+        for tax, seq in self.aln.items(): #pylint: disable=unused-variable
             aln_tax.add(tax)
         if not aln_tax.intersection(treed_tax):
             self.write_files()
