### 2021-03-30 bd9ee57
pylint fixing singleton comparison
