### 2021-03-30 a2c0957
pylint global disabling too many locals and branches
