### 2021-03-30 896aa59
pylint glocal disabling attributes defined outside init
