### 2021-03-30 c77600f
pylint glocal disabling attributes defined outside init
