### 2021-03-30 191cad2
pylint removing unused argument 'all_taxa'
