+++ docs/source/multi_locus.rst
@@ -1,4 +1,4 @@
 How to combine analyses across multiple loci
 ============================================
 
-.. mdinclude:: ../mds/MultLoci.md
+.. mdinclude:: ../mds/MultiLoci.md
