+++ docs/mds/MultiLoci.md
@@ -7,7 +7,7 @@ it is straigtforward to combine the data from those analyses to generate species
 The multi_loci.py script can be used to combine results from single locus runs.
 
 
-usage: 
+usage:
     multi_loci.py [-h] [-d LOCUS_RUNS_FOLDER] [-o OUTPUT] [-f {concatenate,astral}]
                      [-s {fasta,nexus}] [-m INCLUDE_MISSING]
 
