### 2021-05-06 fc31e1d
polishing the rooting and comparison docs
