### 2021-05-07 034259d
specifics about otu info file
