+++ docs/mds/PhyscraperRun.md
@@ -111,7 +111,7 @@ output directory as "run.config", e.g.,
 ### Input Data
 
 
-Tree information (required)
+Tree information (required):
 
   -s STUDY_ID, --study_id STUDY_ID
                         OpenTree study id
@@ -120,16 +120,22 @@ Tree information (required)
 
 OR
 
-  -tf TREE_FILE, --tree_file TREE_FILE
-                        A path to a tree file
-  -tfs TREE_SCHEMA, --tree_schema TREE_SCHEMA
-                        Tree file format schema
-  -ti TAXON_INFO, --taxon_info TAXON_INFO
-                        Taxon info file from OpenTree TNRS
+<blockquote>
+<div><dl class="option-list">
+<dt><kbd><span class="option">-tf <var>TREE_FILE</var></span>, <span class="option">--tree_file <var>TREE_FILE</var></span></kbd></dt>
+<dd><p>A name (and path) to a tree file.</p>
+</dd>
+<dt><kbd><span class="option">-tfs <var>{newick,nexus}</var></span>, <span class="option">--tree_schema <var>{newick,nexus}</var></span></kbd></dt>
+<dd><p>Tree file format schema.</p>
+</dd>
+<dt><kbd><span class="option">-ti <var>FILE_NAME</var></span>, <span class="option">--taxon_info <var>FILE_NAME</var></span></kbd></dt>
+<dd><p>Name (and path) of a taxon info file from an OpenTree TNRS run.</p>
+</dd>
+</dl></div>
+</blockquote>
 
 
-
-Alignment information (required)
+Alignment information (required):
 
   -a ALIGNMENT, --alignment ALIGNMENT
                         Gives the path to alignment file
@@ -138,13 +144,26 @@ Alignment information (required)
 
 OR
 
-  -tb, --treebase       Downloads alignment from TreeBASE
+<blockquote>
+<div><dl class="option-list">
+<dt><kbd><span class="option">-tb </span>, <span class="option">--treebase </span></kbd></dt>
+<dd><p>Downloads alignment from TreeBASE.</p>
+</dd>
+</dl>
+</div></blockquote>
+
+<br/>
 
 Tree and alignment information are required.
 After an analysis has been run, they can be reloaded from a directory from a previous run.
 
-  -re RELOAD_FILES, --reload_files RELOAD_FILES
-                        Reloads files and configuration from the output directory specified in -o
+<blockquote>
+<div><dl class="option-list">
+<dt><kbd><span class="option">-re <var>RELOAD_FILES</var></span>, <span class="option">--reload_files <var>RELOAD_FILES</var></span></kbd></dt>
+<dd><p>Reloads files and configuration from the output directory specified in <code class="docutils literal notranslate"><span class="pre">-o, --output</span></code>.
+</p></dd>
+</dl></div>
+</blockquote>
 
 
 REQUIRED:
@@ -154,72 +173,106 @@ REQUIRED:
 
 Optional:
 
-  -st SEARCH_TAXON, --search_taxon SEARCH_TAXON
-                        Specifies the taxonomic id to constrain the BLAST search. Format ott:123
-                        or ncbi:123. By default it will use the ingroup of the tree from
-                        OpenTree, or the MRCA of all tips if the former is not specified.
-
-
+<blockquote>
+<div><dl class="option-list">
+<dt><kbd><span class="option">-st <var>SEARCH_TAXON</var></span>, <span class="option">--search_taxon <var>SEARCH_TAXON</var></span></kbd></dt>
+<dd><p>Specifies the taxonomic id to constrain the BLAST search. Format  <code class="docutils literal notranslate"><span class="pre">ott:123</span></code> or
[…]
