### 2021-06-17 McTavishLab/physcraper#165 @LunaSare
reviewer 2 - comment 1 - clarify usability 
"My overall impression is that the requested familiarity with open ToL, treeBASE, and TNRS might hinder a broader usability of the tool. If the user is unfamiliar with these, they will really consider the learning curve in order to decide if using this tool is worth; if they are familiar, they are probably doing it without physcraper already. My suggestion is that you must show how the tool can be useful even if one is does not want to familiarise with the above mentioned tools."
