### 2021-06-17 McTavishLab/physcraper#171 @LunaSare
reviewer 2 - comment 7 - explain polytomy handling in docs
"How does it handle polytomies, are all trees in OToL fully resolved? Furthermore, is it possible to know when there are conflicting trees across studies? This is an OToL question actually, but still it would be good to be adressed in the documentation here."
