+++ docs/mds/changelog.md
@@ -1,0 +1,16 @@
+All notable changes to the Physcraper project are documented here.
+
+## `0.6.0`
+
+- add citation to README
+- add links to installation instructions and demonstration usage (documentation and examples) to README
+- Pylinting code
+- Multilocus script
+- Fix supertreebase error
+
+## `0.5.0`
+
+- New contributor [Luna L. Sanchez Reyes](https://github.com/LunaSare)
+- Fixing lack of ingroup
+- Adding documentation
+- Rerooting functions
