+++ docs/source/index.rst
@@ -26,8 +26,9 @@ Welcome to Physcraper's documentation!
    Start <find_trees>
    Run <physcraper_run>
    Results <data_exploration>
-   Combining multiple loci <multi_locus>   
+   Combining multiple loci <multi_locus>
    Examples <examples>
    Implementation <methods_extended>
    CONTRIBUTING
    CONTRIBUTORS
+   changelog
