### 2021-07-20 d370b9d
cleaning function documentation
