+++ physcraper/opentree_helpers.py
@@ -50,7 +50,7 @@ synthref = "Redelings BD, Holder MT. A supertree pipeline for summarizing phylog
 def root_tree_from_synth(tree, otu_dict, base='ott'):
     # Disable all the too many locals and branches for this function
     # pylint: disable=too-many-locals, too-many-branches
-    """Uses information from OpenTree of Life to suggest root.
+    """Uses information from OpenTree of Life to suggest a root.
     :param tree: dendropy :class:`Tree
     :param otu_dict: a dictionary of tip label metadata, inculding an '^ot:ottId'attribute
     'param base: either `synth` or `ott.` If `synth` will use OpenTree synthetic tree relationships to root input tree,
@@ -117,8 +117,8 @@ def root_tree_from_synth(tree, otu_dict, base='ott'):
     return tree
 
 def ottids_in_synth(synthfile=None):
-    """Checks if ottids are present in current synth tree,
-    using a file listing all current otts in synth (v12.3)
+    """Checks if OTT ids are present in current synthetic tree,
+    using a file listing all current OTT ids in synth (v12.3)
     :param synthfile: defaults to taxonomy/ottids_in_synth.txt
     """
     if synthfile is None:
@@ -132,7 +132,7 @@ def ottids_in_synth(synthfile=None):
 
 
 def check_if_ottid_in_synth(ottid):
-    """Web call to check if ott id in synth tree. NOT USED.
+    """Web call to check if OTT id in synthetic tree. NOT USED.
     """
     url = 'https://api.opentreeoflife.org/v3/tree_of_life/node_info'
     payload = json.dumps({"ott_id":int(ottid)})
@@ -152,8 +152,8 @@ def check_if_ottid_in_synth(ottid):
         sys.stderr.write("Connection Error - coud not get taxon information from OpenTree\n")
 
 def get_ottid_from_gbifid(gbif_id):
-    """Returns a dictionary mapping gbif_ids to ott_ids.
-    ott_id is set to 'None' if the gbif id is not found in the Open Tree Txanomy
+    """Returns a dictionary mapping GBIF ids to OTT ids.
+    ott_id is set to 'None' if the GBIF id is not found in the Open Tree Taxanomy
     """
     url = 'https://api.opentreeoflife.org/v3/taxonomy/taxon_info'
     headers = {'content-type':'application/json'}
@@ -163,14 +163,14 @@ def get_ottid_from_gbifid(gbif_id):
     if res.status_code == 200:
         ott_id = int(res.json()['ott_id'])
         return ott_id
-    sys.stderr.write("error getting ott_id for gbif id {}, {}, {}".format(tax, res.status_code, res.reason))
+    sys.stderr.write("error getting OTT id for GBIF id {}, {}, {}".format(tax, res.status_code, res.reason))
     return None
 
 
 
 def bulk_tnrs_load(filename):
     """Read in outputs from OpenTree Bulk TNRS,
-    translates to a Physcraper otu_dictionary.
+    translates to a Physcraper OTU dictionary.
     :param filename: input json file
     """
     otu_dict = {}
@@ -216,9 +216,9 @@ def bulk_tnrs_load(filename):
 # res = requests.post(url, data=payload, headers=headers)
 
 def get_citations_from_json(synth_response, citations_file):
-    """Get ciattions for studies in an induced synth tree repsonse.
-    :param synth_response: Webservice call record
-    :param citations_file: output file
+    """Get ciattions for studies in an induced synthetic tree repsonse.
+    :param synth_response: Web service call record
+    :param citations_file: Output file
     """
     assert isinstance(citations_file, str)
     f = open(citations_file, "w+")
[…]
