### 2021-07-21 894cb59
trapping dendropy error when dataset contains ambiguous characters defined with the symbol {
