+++ docs/mds/Tutorial.md
@@ -8,9 +8,9 @@ If you have access to a single gene alignment, and a tree, Physcraper automates 
 
 <br/>
 
-<img src="https://raw.githubusercontent.com/McTavishLab/physcraper/main/docs/img/schematic.svg" title="schematic" alt="The Phsycraper Framework" width="100%" />
+<img src="https://raw.githubusercontent.com/McTavishLab/physcraper/main/docs/img/schematic-final.svg" title="schematic" alt="The Phsycraper Framework" width="100%" />
 
-The Physcraper framework consists of 4 general steps. The methodology is extensively described in the
+Figure 1 from the [Physcraper publication](https://doi.org/10.1186/s12859-021-04274-6). The Physcraper framework consists of 4 general steps. The methodology is extensively described in the
 [Implementation](https://physcraper.readthedocs.io/en/latest/methods_extended.html) section of the documentation.
 
 <br/>
