+++ docs/source/faq.rst
@@ -1,0 +1,5 @@
+FAQs
+====
+
+
+.. mdinclude:: ../mds/faq.md
