+++ docs/mds/faq.md
@@ -1,0 +1,6 @@
+Frequently asked questions
+
+
+# Polytomies
+
+How does Physcraper handle polytomies?
