### 2021-07-22 dd017c9
handling polytomies
