+++ docs/mds/citation.md
@@ -1,3 +1,5 @@
-If you use Physcraper, please cite: 
+## Citation
 
-- [Physcraper](https://www.biorxiv.org/content/10.1101/2020.09.15.299156v1.abstract): *Luna Luisa Sanchez Reyes, Martha Kandziora, Emily Jane McTavish. (2020). Physcraper: A Python package for continual update of evolutionary estimates using the Open Tree of Life. BioRxiv 2020.09.15.299156*. doi: [doi.org/10.1101/2020.09.15.299156](https://doi.org/10.1101/2020.09.15.299156)
+If you use Physcraper, please cite:
+
+- [Physcraper](https://doi.org/10.1186/s12859-021-04274-6): *Luna Luisa Sanchez Reyes, Martha Kandziora, Emily Jane McTavish. (2020). Physcraper: a Python package for continually updated phylogenetic trees using the Open Tree of Life. BMC Bioinformatics 22, 355 (2021)* doi: [doi.org/10.1186/s12859-021-04274-6](https://doi.org/10.1186/s12859-021-04274-6)
