+++ docs/mds/citation.md
@@ -2,4 +2,4 @@
 
 If you use Physcraper, please cite:
 
-- [Physcraper](https://doi.org/10.1186/s12859-021-04274-6): *Luna Luisa Sanchez Reyes, Martha Kandziora, Emily Jane McTavish. (2020). Physcraper: a Python package for continually updated phylogenetic trees using the Open Tree of Life. BMC Bioinformatics 22, 355 (2021)* doi: [doi.org/10.1186/s12859-021-04274-6](https://doi.org/10.1186/s12859-021-04274-6)
+- *Sánchez-Reyes, L.L., Kandziora, M. & McTavish, E.J. Physcraper: a Python package for continually updated phylogenetic trees using the Open Tree of Life. BMC Bioinformatics 22, 355 (2021)* doi: [doi.org/10.1186/s12859-021-04274-6](https://doi.org/10.1186/s12859-021-04274-6)
