+++ docs/mds/intro.md
@@ -1,4 +1,1 @@
-Use a tree and a single locus alignment to find and add homologous sequences to improve and advance phylogenetic inference in a group.
-
-The tool is under active development in the [McTavish Lab](https://mctavishlab.github.io/).
-Please post a GitHub issue [here](https://github.com/McTavishLab/physcraper/issues) or contact ejmctavish@ucmerced.edu if you need any help or have feedback.
+Use a phylogenetic tree, Open Tree of Life tools, and a single locus alignment to automatically find and add homologous sequences to improve and advance phylogenetic inference in a group.
