### 2021-07-27 0ecf891
updating anaconda install
