### 2021-07-27 947c620
updating requirements on intro
