+++ docs/mds/intro-dendropy.md
@@ -1,7 +1,17 @@
-Physcraper relies on:
+## Requirements
 
-- [Dendropy](https://dendropy.org/primer/index.html): *Sukumaran, J and MT Holder. 2010. DendroPy: a Python library for phylogenetic computing. Bioinformatics 26: 1569-1571* [doi.org/10.1093/bioinformatics/btq228]( https://doi.org/10.1093/bioinformatics/btq228)
+Physcraper requires the user to install:
+
+- [Anaconda](https://docs.anaconda.com/anaconda/install/): *Anaconda Software Distribution. Computer software. Vers. 4.8.0. Anaconda, July. 2021. Web. <https://anaconda.com>*
+- OR [Virtualenv](https://pypi.org/project/virtualenv/),
+- [MUSCLE](https://www.ebi.ac.uk/Tools/msa/muscle/): *Edgar RC. MUSCLE: multiple sequence alignment with high accuracy and high throughput. Nucleic Acids Res. 2004 Mar 19;32(5):1792-7. doi: [10.1093/nar/gkh340](https://doi.org/10.1093/nar/gkh340)*
+- AND [RAxML](https://cme.h-its.org/exelixis/web/software/raxml/): *Stamatakis, Alexandros. "RAxML version 8: a tool for phylogenetic analysis and post-analysis of large phylogenies." Bioinformatics 30.9 (2014): 1312-1313.* doi: [10.1093/bioinformatics/btu033](https://doi.org/10.1093/bioinformatics/btu033)
+-
+
+Physcraper relies on the following packages that are automatically installed:
+
+- [DendroPy](https://dendropy.org/primer/index.html): *Sukumaran, J and MT Holder. 2010. DendroPy: a Python library for phylogenetic computing. Bioinformatics 26: 1569-1571* doi:  [10.1093/bioinformatics/btq228](https://doi.org/10.1093/bioinformatics/btq228)
 <br/><br/>
-- [The Open Tree of Life Project](https://tree.opentreeoflife.org/opentree/argus/opentree12.3@ott93302): *Open Tree of Life, Benjamin Redelings, Luna Luisa Sanchez Reyes, Karen A. Cranston, Jim Allman, Mark T. Holder, & Emily Jane McTavish. (2019). Open Tree of Life Synthetic Tree (Version 12.3). Zenodo* [doi.org/10.5281/zenodo.3937741](https://doi.org/10.5281/zenodo.3937741)
+- [The Open Tree of Life Project](https://tree.opentreeoflife.org/opentree/argus/opentree12.3@ott93302): *Open Tree of Life, Benjamin Redelings, Luna Luisa Sanchez Reyes, Karen A. Cranston, Jim Allman, Mark T. Holder, & Emily Jane McTavish. (2019). Open Tree of Life Synthetic Tree (Version 12.3). Zenodo* doi: [10.5281/zenodo.3937741](https://doi.org/10.5281/zenodo.3937741)
 <br/><br/>
-- [OpenTree](https://github.com/OpenTreeOfLife/python-opentree): *Emily Jane McTavish, Luna Luisa Sanchez Reyes, Mark T. Holder. (2020). OpenTree: A Python package for Accessing and Analyzing data from the Open Tree of Life. BioRxiv 2020.12.14.422759* [doi.org/10.1101/2020.12.14.422759](https://doi.org/10.1101/2020.12.14.422759)
+- [OpenTree](https://github.com/OpenTreeOfLife/python-opentree): *Emily Jane McTavish, Luna Luisa Sanchez-Reyes, Mark T. Holder. (2020). OpenTree: A Python package for Accessing and Analyzing data from the Open Tree of Life. BioRxiv 2020.12.14.422759* doi: [10.1101/2020.12.14.422759](https://doi.org/10.1101/2020.12.14.422759)
