### 2021-07-27 b960b8f
moving OpenTree citation to the beginning
