+++ docs/mds/intro.md
@@ -1,1 +1,3 @@
 Use a phylogenetic tree, Open Tree of Life tools, and a single locus alignment to automatically find and add homologous sequences to improve and advance phylogenetic inference in a group.
+
+Physcraper relies on [The Open Tree of Life Project](https://tree.opentreeoflife.org/opentree/argus/opentree12.3@ott93302): *Open Tree of Life, Benjamin Redelings, Luna Luisa Sanchez Reyes, Karen A. Cranston, Jim Allman, Mark T. Holder, & Emily Jane McTavish. (2019). Open Tree of Life Synthetic Tree (Version 12.3). Zenodo* doi: [10.5281/zenodo.3937741](https://doi.org/10.5281/zenodo.3937741)
