### 2021-07-27 f6379d0
updating contact, license and requirements
