+++ docs/mds/license.md
@@ -1,1 +1,3 @@
+## License
+
 Physcraper is made available through the [GNU General Public License v3.0](https://github.com/McTavishLab/physcraper/blob/main/LICENSE)
