+++ README.md
@@ -1,7 +1,7 @@
 
 <!-- README.md is generated from README.Rmd; please edit the .Rmd file and then from R do rmarkdown::render("README.Rmd")-->
 
-<img align="left" width="250" src="https://raw.githubusercontent.com/McTavishLab/physcraper/master/docs/physcraper-long.png">
+<img align="left" width="250" src="https://raw.githubusercontent.com/McTavishLab/physcraper/cleanup/docs/physcraper-long.png">
 
 # Physcraper
 
@@ -18,64 +18,121 @@ Status](https://travis-ci.org/McTavishLab/physcraper.svg?branch=main)](https://t
 
 ## Automated gene tree updating\!
 
-Use a tree and a single locus alignment to find and add homologous
-sequences to improve and advance phylogenetic inference in a group.
+Use a phylogenetic tree, Open Tree of Life tools, and a single locus
+alignment to automatically find and add homologous sequences to improve
+and advance phylogenetic inference in a group.
+
+Physcraper relies on [The Open Tree of Life
+Project](https://tree.opentreeoflife.org/opentree/argus/opentree12.3@ott93302):
+*Open Tree of Life, Benjamin Redelings, Luna Luisa Sanchez Reyes, Karen
+A. Cranston, Jim Allman, Mark T. Holder, & Emily Jane McTavish. (2019).
+Open Tree of Life Synthetic Tree (Version 12.3). Zenodo* doi:
+[10.5281/zenodo.3937741](https://doi.org/10.5281/zenodo.3937741)
+
+This is the code repository. Please refer to Physcraper’s [documentation
+website](https://physcraper.readthedocs.io/en/latest/) for more details
+on:
+
+  - [installation](https://physcraper.readthedocs.io/en/latest/install.html)
+    instructions using an Anaconda or a Virtualenv Python virtual
+    environment,
+  - [function](https://physcraper.readthedocs.io/en/latest/apidocs.html)
+    usage,
+  - tutorials on finding and choosing a [starting
+    dataset](https://physcraper.readthedocs.io/en/latest/find_trees.html),
+    [running](https://physcraper.readthedocs.io/en/latest/physcraper_run.html)
+    a full analysis,
+  - [visualizing and
+    analysing](https://physcraper.readthedocs.io/en/latest/data_exploration.html)
+    results,
+  - real life
+    [examples](https://physcraper.readthedocs.io/en/latest/examples.html),
+    and
+  - tools for
+    [developers](https://physcraper.readthedocs.io/en/latest/CONTRIBUTING.html).
+
+:hamster: :palm\_tree: :frog: :ear\_of\_rice: :panda\_face: :tulip:
+:octopus: :blossom: :whale: :mushroom: :ant: :cactus: :fish:
+:maple\_leaf: :water\_buffalo: 🦠 :shell: :bug: :octocat:
+
+## Citation
+
+If you use Physcraper, please cite:
+
+Sánchez-Reyes, L.L., Kandziora, M. & McTavish, E.J. *Physcraper: a
+Python package for continually updated phylogenetic trees using the Open
+Tree of Life.* BMC Bioinformatics 22, 355 (2021) doi:
+[doi.org/10.1186/s12859-021-04274-6](https://doi.org/10.1186/s12859-021-04274-6)
+
+## License
+
+Physcraper is made available through the [GNU General Public License
+v3.0](https://github.com/McTavishLab/physcraper/blob/main/LICENSE)
+
+## Contact
 
 The tool is under active development in the [McTavish
[…]
