+++ README.Rmd
@@ -33,7 +33,7 @@ IS_README <- TRUE
 
 ```
 
-This is the code repository. Please refer to Physcraper’s
+You are now on the code repository. Please refer to Physcraper’s
 [documentation website](https://physcraper.readthedocs.io/en/latest/)
 for more details on:
 
@@ -42,9 +42,8 @@ instructions using an Anaconda or a Virtualenv Python virtual
 environment,
 - [function](https://physcraper.readthedocs.io/en/latest/apidocs.html)
 usage,
-- tutorials on finding and choosing a [starting dataset](https://physcraper.readthedocs.io/en/latest/find_trees.html),
-[running](https://physcraper.readthedocs.io/en/latest/physcraper_run.html)
-a full analysis,
+- finding and choosing a [starting dataset](https://physcraper.readthedocs.io/en/latest/find_trees.html),
+- [running](https://physcraper.readthedocs.io/en/latest/physcraper_run.html) a full analysis,
 - [visualizing and analysing](https://physcraper.readthedocs.io/en/latest/data_exploration.html) results,
 - real life [examples](https://physcraper.readthedocs.io/en/latest/examples.html), and
 - tools for [developers](https://physcraper.readthedocs.io/en/latest/CONTRIBUTING.html).
@@ -52,6 +51,7 @@ a full analysis,
 
 :hamster: :palm_tree: :frog: :ear_of_rice: :panda_face: :tulip: :octopus: :blossom: :whale: :mushroom: :ant: :cactus: :fish: :maple_leaf: :water_buffalo: 🦠 :shell: :bug: :octocat:
 
+<br>
 
 ```{r child="docs/mds/citation.md"}
 
