### 2021-07-28 McTavishLab/physcraper#146 @LunaSare
This is a great suggestion. We will look at implementing progress bars in a future release.
