### 2021-07-28 McTavishLab/physcraper#154 @LunaSare
The current version of the software does not accept aminoacid alignments. But it is doable to add this functionality to a future version.
