### 2021-07-28 bdbfa28
updating 'methods_extended' file name to 'implementation'
