+++ docs/mds/implementation.md
@@ -9,7 +9,7 @@
 <div id="dna-sequence-search-and-filtering" class="section level2">
 <h2>DNA sequence search and filtering</h2>
 <p>Physcraper uses the GenBank DNA database as source to search for new sequences. The DNA sequence search can be performed on the GenBank remote database or in a GenBank local database set up by the user, which can speed up the search process. Detailed instructions to setup a local database are provided on Physcraper’s software documentation.</p>
-<p>The next step is to identify a “search taxon” to constrain the sequence search on the GenBank database within that taxonomic group. The search taxon can be chosen by the user from the NCBI taxonomy. If none is provided, then the search taxon is automatically set using the taxa in the input tree labeled as the “ingroup” (see Physcraper's framework <a href="https://physcraper.readthedocs.io/en/latest/how_to_start.html#the-physcraper-framework">figure</a>). The search taxon is The Most Recent Common Ancestor (MRCA) of the ingroup taxa in the OpenTree synthetic tree, that is also a named clade in the NCBI taxonomy. <!-- FIGURE RECOMMENDED: Figure \@ref(fig:search)  --> This is known in the OpenTree as the Most Recent Common Ancestral Taxon (MRCAT; also referred as the Least Inclusive Common Ancestral taxon - LICA). The MRCAT can be different from the phylogenetic MRCA when the latter is an unnamed clade in the synthetic tree. To identify the MRCAT of a group of taxon names, we use the OpenTree <a href="https://github.com/OpenTreeOfLife/germinator/wiki/Taxonomy-API-v3#mrca">API</a> <span class="citation">(Rees &amp; Cranston <a href="#ref-rees2017automated">2017</a>)</span>.</p>
+<p>The next step is to identify a “search taxon” to constrain the sequence search on the GenBank database within that taxonomic group. The search taxon can be chosen by the user from the NCBI taxonomy. If none is provided, then the search taxon is automatically set using the taxa in the input tree labeled as the “ingroup” (see Physcraper's framework <a href="https://physcraper.readthedocs.io/en/latest/description.html#the-physcraper-framework">figure</a>). The search taxon is The Most Recent Common Ancestor (MRCA) of the ingroup taxa in the OpenTree synthetic tree, that is also a named clade in the NCBI taxonomy. <!-- FIGURE RECOMMENDED: Figure \@ref(fig:search)  --> This is known in the OpenTree as the Most Recent Common Ancestral Taxon (MRCAT; also referred as the Least Inclusive Common Ancestral taxon - LICA). The MRCAT can be different from the phylogenetic MRCA when the latter is an unnamed clade in the synthetic tree. To identify the MRCAT of a group of taxon names, we use the OpenTree <a href="https://github.com/OpenTreeOfLife/germinator/wiki/Taxonomy-API-v3#mrca">API</a> <span class="citation">(Rees &amp; Cranston <a href="#ref-rees2017automated">2017</a>)</span>.</p>
 <p>Users can provide a search taxon that is either a more or a less inclusive clade relative to the ingroup of the original phylogeny. If the search taxon is more inclusive, the sequence search will be performed outside the MRCAT of the matched taxa, e.g., including all taxa within the family or the order that the ingroup belongs to. If the search taxon is a less inclusive clade, the users can focus on enriching a particular clade/region within the ingroup of the phylogeny.</p>
[…]
