+++ docs/source/description.rst
@@ -1,4 +1,4 @@
 Introduction to Physcraper
 ==========================
 
-.. mdinclude:: ../mds/Tutorial.md
+.. mdinclude:: ../mds/description.md

[…]
