### 2021-07-28 McTavishLab/physcraper#165 @LunaSare
So, I think that the decision to invest the time to learn how to use Physcraper depends on the goals of the user.

To our knowledge, existing tools that automatize dataset construction for phylogenetics focus on mining and filtering homolog sequences.
The underlying goal of Physcraper is interoperability between biological databases. To achieve that, the OpenTree Taxonomy and associated tools are key.

In this sense Physcraper offers the unique advantage to automatically connect a phylogenetic tree to other databases and services, including the conflicting service provided by OpenTree, which permits to rapidly identify regions that have been enriched in a tree with phylogenetic information, regions that are coherent with other phylogenetic estimates, as well as regions that conflict with other phylogenetic estimates.
Users could update phylogenetic trees using any other existing tool, but if they wanted to connect them to other databases, particularly in a reproducible way, the OpenTree Taxonomy would allow them to do this programmatically, instead of by hand.

For this goal, having a minimum familiarity with the OpenTree tools is needed.

We realize that this might initially reduce the pool of users, but we believe that the benefits that are added by connecting taxonomic data with the OpenTree services will encourage users to familiarize with the OpenTree services, potentially increasing the adoption of Physcraper.
