+++ docs/mds/faq.md
@@ -1,5 +1,24 @@
 Frequently asked questions
 
+## I have to learn to use OpenTree to use Physcraper, is the learning curve worth it?
+
+We think that this decision depends on the goals of the user.
+
+To our knowledge, existing tools that automatize dataset construction for phylogenetics focus on assembling an alignment _de novo_ and/or mining and filtering homolog sequences.
+The main goal of Physcraper is to construct upon the knowledge contained in previous expertly-curated and peer-reviwed homology hypothesis (alignments), as well as establish interoperability between biological databases and phylogenetic knowledge. To achieve that, understanding the OpenTree Taxonomy and associated tools is key.
+
+In this sense Physcraper offers the unique advantage to automatically connect a phylogenetic tree to other databases and services, including alignment databases such as TreeBASE, and the conflicting service provided by OpenTree, which permits to rapidly identify regions in the tree that:
+- have been enriched in a tree with phylogenetic information,
+- are coherent with other phylogenetic estimates, as well as
+- conflict with other phylogenetic estimates.
+-
+Users can update phylogenetic trees using any other existing tool, and if they want to connect them to other biological databases, particularly in a reproducible way, matching their taxon names to the OpenTree Taxonomy allows them to do this programmatically, instead of by hand.
+
+For this goal, having a minimum familiarity with the OpenTree tools is needed.
+
+We realize that this might initially discourage some users, but we believe that the benefits brought by connecting taxonomic data with the OpenTree services will encourage users to familiarize with the OpenTree services, and to adopt the use of Physcraper.
+
+
 ## How does Physcraper handle polytomies of starting trees?
 
 The Physcraper starting tree is a phylogeny whose tip labels must have been standardized to the OpenTree Taxonomy (as described in the Introduction section:
