### 2021-07-30 #26 @LunaSare
Here it goes for the first reviewer!

Thanks for the thorough review @bpucker, it was most helpful!
@NickleDave, @snacktavish and I address all points below:

> ## Package Review
> * [x]  As the reviewer I confirm that there are no conflicts of interest for me to review this work.
>
> #### Documentation
> The package includes all the following forms of documentation:
>
> * [x]  **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
> * [ ]  **Installation instructions:** for the development version of package and any non-standard dependencies in README

Installation instructions can be found in Physcraper's documentation at readthedocs here: https://physcraper.readthedocs.io/en/latest/install.html,
and a link to it has been added to the README.

> * [x]  **Vignette(s)** demonstrating major functionality that runs successfully locally
> * [ ]  **Function Documentation:** for all user-facing functions

Documentation for all user-facing functions can be found in Physcraper's documentation at readthedocs here: https://physcraper.readthedocs.io/en/latest/apidocs.html

> * [ ]  **Examples** for all user-facing functions

Examples for all command line user-facing functions can be found in the [Start section](https://physcraper.readthedocs.io/en/latest/find_trees.html) and the [Run section](https://physcraper.readthedocs.io/en/latest/physcraper_run.html) of Physcraper's documentation at readthedocs.

> * [x]  **Community guidelines** including contribution guidelines in the README or CONTRIBUTING.
> * [x]  **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.
>
> Readme requirements
> The package meets the readme requirements below:
>
> * [x]  Package has a README.md file in the root directory.
>
> The README should include, from top to bottom:
>
> * [x]  The package name
> * [x]  Badges for continuous integration and test coverage, a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the badge for pyOpenSci peer-review will be provided upon acceptance.)
> * [x]  Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
> * [ ]  Installation instructions

A link to the [Installation section](https://physcraper.readthedocs.io/en/latest/install.html) of Physcraper's documentation at readthedocs is provided in the [README](https://github.com/McTavishLab/physcraper/tree/main#readme).

> * [x]  Any additional setup required (authentication tokens, etc)
> * [ ]  Brief demonstration usage

A link to [a set of demonstration usage examples](https://physcraper.readthedocs.io/en/latest/examples.html) in Physcraper's documentation at readthedocs is provided in the [README](https://github.com/McTavishLab/physcraper/tree/main#readme), as "real life examples".

> * [x]  Direction to more detailed documentation (e.g. your documentation files or website).
> * [x]  If applicable, how the package compares to other similar packages and/or how it relates to other packages
> * [ ]  Citation information

Citation information is now included in the [README](https://github.com/McTavishLab/physcraper#readme).

>
> #### Usability
> Reviewers are encouraged to submit suggestions (or pull requests) that will improve the usability of the package as a whole.
> Package structure should follow general community best-practices. In general please consider:
>
> * [x]  The documentation is easy to find and understand
> * [x]  The need for the package is clear
> * [x]  All functions have documentation and associated examples for use
>
> #### Functionality
> * [x]  **Installation:** Installation succeeds as documented.
> * [x]  **Functionality:** Any functional claims of the software been confirmed.
> * [x]  **Performance:** Any performance claims of the software been confirmed.
> * [x]  **Automated tests:** Tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
> * [x]  **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.
> * [x]  **Packaging guidelines**: The package conforms to the pyOpenSci [packaging guidelines](https://www.pyopensci.org/dev_guide/authoring/overview.html).
>
> #### Review Comments
> 1. Installation instructions are not in the README.md, but there are links to all relevant pages. Documentation and examples are also provided on different pages. I think this is fine though.

Awesome!
We made sure all relevenat section were linked in the [README](https://github.com/McTavishLab/physcraper#readme) and addressed it here
https://github.com/McTavishLab/physcraper/issues/152

> 2. Some parts (relabeling/rerooting) are labeled as "under construction" thus I have not tested those.

Thanks for noting that! We addressed it here https://github.com/McTavishLab/physcraper/issues/153

Briefly, we have updated those in here: https://physcraper.readthedocs.io/en/latest/data_exploration.html?highlight=under%20construction#analysing-the-physcraper-results. They are still labeled as "under construction", but they are runnable and testable.

> 3. Anaconda/virtualenv installation instructions should be added and these requirements should be specified in the README.

We adressed those here https://github.com/McTavishLab/physcraper/issues/143

Briefly, we added links to installation instructions for Anaconda (https://physcraper.readthedocs.io/en/latest/install.html#anaconda-virtual-environment) and virtualenv (https://physcraper.readthedocs.io/en/latest/install.html#virtualenv-virtual-environment) on the installation section of Physcraper's documentation at readthedocs.

Also, a requirements section was added to the [README](https://github.com/McTavishLab/physcraper/tree/main#readme), at the end.

> 4. Support for additional alignment tools e.g. MAFFT could be included to allow user to choose their favorite one.

There are many alignment tools available. MUSCLE is one of the most widely used ones and we integrate it to the Physcraper workflow mainly as a proof-of-concept.

In the [Implementation section](https://physcraper.readthedocs.io/en/latest/methods_extended.html): __New DNA sequence alignment__, at Physcraper's readthedocs we make it more clear that the main goal of Physcraper is to write unaligned sequences which then can be aligned with any software preferred by the user.

> 5. There are some details missing in the function documentation:
>    self.otu_by_gi
>    self._to_be_pruned
>    self.tmpfi

Thanks! We were actually not using those, so we deleted them. We addressed that here https://github.com/McTavishLab/physcraper/issues/145

> 6. The BLAST jobs run for a very long time (as promised in the documentation). I know this is tricky, but some kind of progress bar would be great.

This is a great suggestion. Although we are not able to implement it right now, we added it to our enhancement list https://github.com/McTavishLab/physcraper/issues/146

> 7. Would it be possible to use peptide sequences? These could show better conservation over large phylogenetic distances.

This is also a great suggestion. As of now, Physcraper only implements DNA BLAST, so updating an aminoacid alignment is not possible as of now. However, programming-wise, it is very possible to implement a workflow that accepts aminoacid alignments. We have added this to our future enhancement list https://github.com/McTavishLab/physcraper/issues/154

> 8. Everything looks fine, but I got an issue when running physcraper on a different dataset. Here are the details:
>
> Description: Ubuntu 18.04.2 LTS
> Release: 18.04
> Codename: bionic
> Python 3.8.5 (on the system)
> Python 3.6.11 (in physcraper_env)
>
> anaconda Command line client (version 1.7.2)
>
> physcraper_run.py -s pg_328 -t tree322 -r -tb -o chenopodia
>
> Traceback (most recent call last):
> File "/home/ubuntu/anaconda3/envs/physcraper_env/bin/physcraper_run.py", line 7, in
> exec(compile(f.read(), **file**, 'exec'))
> File "XXX/physcraper/bin/physcraper_run.py", line 181, in
> dataset = physcraper.opentree_helpers.get_dataset_from_treebase(study_id)
> File "XXX/physcraper/physcraper/opentree_helpers.py", line 390, in get_dataset_from_treebase
> dna = DataSet.get(url=url, schema="nexml")
> File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/site-packages/dendropy/datamodel/datasetmodel.py", line 188, in get
> return cls._get_from(**kwargs)
> File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/site-packages/dendropy/datamodel/basemodel.py", line 160, in _get_from
> return cls.get_from_url(src=src, schema=schema, **kwargs)
> File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/site-packages/dendropy/datamodel/basemodel.py", line 279, in get_from_url
> **kwargs)
> File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/site-packages/dendropy/datamodel/datasetmodel.py", line 129, in _parse_and_create_from_stream
> state_alphabet_factory=charstatemodel.StateAlphabet,
> File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/site-packages/dendropy/dataio/ioservice.py", line 255, in read_dataset
> global_annotations_target=dataset)
> File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/site-packages/dendropy/dataio/nexmlreader.py", line 284, in _read
> subelement_factory=self._subelement_factory)
> File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/site-packages/dendropy/dataio/xmlprocessing.py", line 150, in **init**
> self.parse_file(file_obj)
> File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/site-packages/dendropy/dataio/xmlprocessing.py", line 166, in parse_file
> for event, elem in ElementTree.iterparse(source, events):
> File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/xml/etree/ElementTree.py", line 1221, in iterator
> yield from pullparser.read_events()
> File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/xml/etree/ElementTree.py", line 1296, in read_events
> raise event
> File "/home/ubuntu/anaconda3/envs/physcraper_env/lib/python3.6/xml/etree/ElementTree.py", line 1268, in feed
> self._parser.feed(data)
> xml.etree.ElementTree.ParseError: syntax error: line 1, column 0

We really appreciate the feedback! We addressed and solved that issue here https://github.com/McTavishLab/physcraper/issues/147

> Minor comments:
> These automated tree > These automated trees (https://physcraper.readthedocs.io/en/latest/how_to_start.html)
> mapped a unified taxonomy > mapped to a unified taxonomy
>
> There are formatting issue on https://physcraper.readthedocs.io/en/latest/physcraper_run.html involving
>
> 1. "physcraper_run.py -tf tests/data/tiny_test_example/test.tre -tfs newick -a tests/data/tiny_test_example/test.fas -as fasta –taxon_info tests/data/tiny_test_example/main.json -o owndata"
> 2. "Take a look at the tree, the alignemnt and the out_info csv file. It will list all taxa by their unique identifiers."
>
> final output ree > final output tree
>
> cofniguration > configuration
>
> updtaed > updated (https://physcraper.readthedocs.io/en/latest/data_exploration.html)
>
> Genbank/genbank > GenBank (https://physcraper.readthedocs.io/en/latest/apidocs.html)
>
> It appears that there should be a link on the overview figure, but the pages does not exist: https://physcraper.readthedocs.io/en/latest/how_to_start.html#

We addressed and solved all minor comments here https://github.com/McTavishLab/physcraper/issues/155
