### 2021-08-06 #26 @NickleDave
Hi @bpucker thank you again for your review -- when you get a chance can you let us know whether the changes made address your comments?

> Here it goes for the first reviewer!
>
> Thanks for the thorough review @bpucker, it was most helpful!
> @NickleDave, @snacktavish and I address all points below:
