### 2021-08-06 #26 @NickleDave
Just following up on this.
@leomrtns do you by any chance have code to reproduce this issue?

> We have one question for @leomrtns. You mentioned you were unable to complete an analysis because your trees did not seem appropriate:
>
> > ... I had no problem installing the software, however I could not complete the analysis since running from the OToL took too long, and my trees did not seem appropriate (several potentially from same species, without easy taxonomic mapping).
>
> Would you be able to share the Physcraper code or workflow you were using? and what was the issue exactly? (just too long of a runtime, maybe, or an actual error? We would like to be able to reproduce the issue.
> This would allows us to potentially fix it – if it is a bug, or address it appropriately otherwise. Thanks!
