### 2021-08-06 #26 @leomrtns
@LunaSare @NickleDave For the OToL run, I picked up a random study using `find_trees.py`, probably using _salmonella_. Let's say the command was `physcraper_run.py -s ot_274  -t Tr47737 -tb` . Definitely not a bug, just slower than I expected (specially if you don'y know what's going on: "is it downloading? It doesn't seem to be using a lot of memory! is this one BLAST or an all-against-all?" etc.)

As for my trees, one set of trees is from OMA (for instance, https://omabrowser.org/oma/hog/HOG:A0785954/LUCA/), and another is from https://www.biorxiv.org/content/10.1101/2021.02.26.433005v1 . Both sets have trees with over 100 species, and  contain paralogs. I'll paste below one example tree from the biorxiv ms, the species names are in supplementary table 1 (the whole set is on dryad but it's quite big to download). If you say "homologous" I hear "orthologs **and** paralogs".

My observation was not an issue report, BTW, but a comment in line with "managing expectations" in the documentation (limitation to orthologs, expectations regarding amount of data downloaded, local computational cost). OTOH it would be nice, although not necessary obviously, to be a bit more verbose about the run progress...

```
(Sr_di_Shan_XX_Contig_20493_Len1191_Cov12_OG5_126843_Trans1:0.93507385393802833207,((((((Am_my_Asub_XX_Contig_4999_length_1431_OG5_126843:0.17925096907545831626,Am_my_Ppal_XX_Contig_5405_Len1464_Cov25_OG5_126843_Trans1:0.21569370339896629107):0.19580340488178876734,(Am_my_ddis_pdi2_OG5_126843:0.17182410886971846709,Am_my_Dpur_330843883:0.12133198464510544512):0.22380255139140953213):0.35613587961841497442,((Am_is_Sram_XX_Contig_5426_Len1509_Cov37_OG5_126843_Trans1:1.04664610892177445400,Sr_rh_Cten_XX_Contig_155_Len1272_OG5_126843:0.56287279219726327195):0.14646817109171722526,(((((Op_fu_aory_AO090001000733_OG5_126843:0.28061071963647415961,(Op_fu_calb_calbsc5314orf19.5130_OG5_126843:0.00000100000050002909,Op_fu_calb_CAWT05658_OG5_126843:0.00000100000050002909):0.71088977757999072260):0.10953901854897206414,(((Op_fu_Pgra_331221285:0.31257165799644082149,Op_fu_Mmel_328856226:0.21714029986314753451):0.17044131379768695522,((Op_fu_cneg_CNAG06240T0_OG5_126843:0.00490335073675200999,Op_fu_cneo_CNBG4941T0_OG5_126843:0.03101474889812528765):0.30752402927941652200,(Op_fu_Dspa_402226425:0.24747620839039843887,(Op_fu_pchr_egww2.22.48.1_OG5_126843:0.23030555575465275209,Op_fu_lbic_eu2.Lbscf0004g07820_OG5_126843:0.18610329221431809454):0.14364851235972994625):0.09212084084122429406):0.04223821693980913533):0.10610145858545251552,Op_fu_Mglo_164661137:0.44190202986900423143):0.10941629592353219191):0.05324886709773616472,Op_fu_Pspa_XX_GT911314_Contig_1098_length_784_Op_fu_Pspa_OG5_126843:0.53299748739514674067):0.12695648887507282598,Am_is_Fnol_XX_Contig_1711_Len1437_Cov243_OG5_126843_Trans1:0.37866676707546248482):0.06899241669144999534,EE_ap_Rram_XX_Contig_3883_Len1449_Cov190_OG5_126843:0.49130421279016206482):0.03226552797335935363):0.04885050315031935764):0.05451735536390352188,((((((Am_ar_ehis_EHI110350_OG5_126843:0.00896701697087026739,Am_ar_Enut_672808606:0.00000100000050002909):1.00048428259544408547,(Op_me_Hrob_555703477:0.34749231851745499000,(Op_me_Hvul_221104341:0.42308763764593515511,Op_ic_Cowc_EFW46739.1:0.49239266211896459691):0.06604986008873386938):0.74205749399907960928):0.11837733876277976852,EE_ce_Acsp_XX_Contig_2601_Len1599_Cov17_OG5_126843_Trans1:0.81214758458767388749):0.11785084209421833878,((Am_my_Ppal_XX_Contig_2537_Len1509_Cov28_OG5_126843_Trans1:0.52749934495786621191,Am_my_Asub_XX_Contig_3775_length_1701_OG5_126843:0.30101032680370692907):0.75686174789342697267,(Am_hi_Gfon_XX_Contig_5450_length_1423_Am_hi_Gfon_OG5_126843:1.00326624363791272465,EE_ce_Acsp_XX_Contig_4369_Len1248_Cov15_OG5_126843_Trans1:0.87553559380058842976):0.07633639786075907541):0.09394973854661196888):0.03435752962535621596,(((((Ex_fo_glab_GL50581862_OG5_126843:0.03813908247187795808,Ex_fo_glam_GL508039413_OG5_126843:0.01143799685546321304):0.00393051277463650917,Ex_fo_glae_GLP151786_OG5_126843:0.01320511605728901840):0.87585043983993016159,Ex_fo_Ssal_XX_Contig_3505_length_909_OG5_126843:0.58524976738558243117):0.08550626727401361427,(Ex_fo_Ssal_XX_Contig_2929_length_1047_OG5_126843:0.92479901549774101710,Ex_fo_Ssal_XX_Contig_2676_length_1119_OG5_126843:0.82114259684530865879):0.10806374774690984120):0.38855888289137296976,((EE_ap_Ttra_XX_Contig_5807_length_1302_OG5_126843:0.73126093108478107396,Sr_rh_Lamo_XX_Contig_6027_Len1272_Cov31_OG5_126843_Trans1:0.91786508669572386410):0.19893412045747563277,EE_ce_Acsp_XX_Contig_4072_Len1410_Cov20_OG5_126843_Trans1:0.90779381806820491629):0.03977079443285653354):0.22537278272106056609):0.08980613712940611892,(((Pl_rh_Emad_XX_Contig_1346_Len1461_Cov548_OG5_126843_Trans1:0.81307075844187215452,(Pl_rh_Rmar_XX_Contig_5789_Len1296_Cov247_OG5_126843_Trans1:0.23880664541671864365,Pl_rh_Rhod_XX_Contig_4236_Len1470_Cov444_OG5_126843_Trans1:0.28187722546828763992):0.22496190694446868696):0.15422057899154020810,Pl_rh_cmer_CMC180C_OG5_126843:0.60770196554481703544):0.20194312853579279854,((Pl_gl_Gnos_XX_EC123363_Contig_2154_length_566_Pl_gl_Gnos_sequence_OG5_126843:0.48229984916893209324,EE_ap_Ttra_XX_Contig_4500_length_1596_OG5_126843:1.23114496778288629031):0.20823349162466239792,(((((((Sr_di_Amas_XX_Contig_8721_Len1503_Cov13_OG5_126843_Trans1:0.22000298007940599643,((Sr_di_Gspi_XX_Contig_1546_Len1539_Cov10_OG5_126843_Trans1:0.20658714149380372205,Sr_di_Pbei_XX_Contig_8880_Len1395_Cov9_OG5_126843_Trans1:0.33594827798944587594):0.04719377508013879174,Sr_di_Bnut_XX_Contig_4886_Len1566_Cov16_OG5_126843_Trans1:0.21042200520234821681):0.04675598122599654988):0.79683890447052296846,(Sr_di_Pbei_XX_Contig_13578_Len1392_Cov10_OG5_126843_Trans1:0.08669763001230035293,(Sr_di_Bnut_XX_Contig_12773_Len1368_Cov8_OG5_126843_Trans1:0.15418121533974440962,(Sr_di_Amas_XX_Contig_16023_Len1446_Cov13_OG5_126843_Trans1:0.33468365133300032044,Sr_di_Gspi_XX_Contig_3497_Len1401_Cov5_OG5_126843_Trans1:0.09092593411446361906):0.03870768158425601235):0.03117587948753398550):0.40070809137965962554):0.15997625899171194486,((Sr_ci_Otri_XX_EJY74011_Contig_10626_Len1677_OG5_126843:1.26908954778811966513,(Sr_ci_tthe_71.m00217_OG5_126843:0.71725726870035044769,(EE_cr_Rlen_XX_Contig_3533_Len1569_Cov7_OG5_126843_Trans1:0.88876841582310861511,(EE_cr_Ccur_XX_Contig_11269_Len810_Cov14_OG5_126843_Trans1:0.39473295025413851489,EE_cr_Htep_XX_Contig_669_Len924_Cov210_OG5_126843_Trans1:0.38107785873453520198):0.08275021765313551880):0.29953133554171801345):0.04243628256110524438):0.08024801064308244303,((Sr_ci_Ehar_XX_Contig_3478_Len1443_Cov82_OG5_126843_Trans1:0.13550220855753106197,Sr_ci_Ehar_XX_Contig_3890_Len1395_Cov9_OG5_126843_Trans1:0.27131912899004723139):0.22056771399770325193,(((Sr_ci_Otri_XX_EJY66972_Contig_12477_Len1464_OG5_126843:0.00000100000050002909,Sr_ci_Otri_XX_EJY76614_Contig_12468_Len1464_OG5_126843:0.00000100000050002909):0.06458872007287089212,Sr_ci_Slem_XX_CDW72484_Contig_9453_Len1461_OG5_126843:0.10195458868921110362):0.21879120320419923051,((Sr_ci_Otri_XX_EJY71853_Contig_12531_Len1458_OG5_126843:0.70782552320815872449,Sr_ci_Slem_XX_CDW74310_Contig_8147_Len1623_OG5_126843:0.54229814632263495522):0.42087477380658905313,Sr_ci_Otri_XX_EJY75278_Contig_10997_Len1629_OG5_126843:0.71856335868730158989):0.12119370392944027459):0.13148612335414314178):0.13829528395855178458):0.07345094621854977857):0.07093460018689877467,(((((Sr_st_Ospa_XX_Contig_5597_Len1053_Cov88_OG5_126843_Trans1:0.77965668850437352955,(Pl_gr_Csti_XX_Contig_2386_Len1389_Cov18_OG5_126843_Trans1:0.53773285015531924191,(Pl_gr_Bpra_424513427:0.75425914540268312880,((Pl_gr_atha_NP568926_OG5_126843:0.00197306264377207280,Pl_gr_atha_NP851234_OG5_126843:0.00000100000050002909):0.16851454196935142238,(Pl_gr_atha_NP191056_OG5_126843:0.00000100000050002909,Pl_gr_atha_NP001118842_OG5_126843:0.00000100000050002909):0.28849845202939300437):0.35824074789470650160):0.02167666002880524329):0.13192342992586075789):0.03453032156420222237,(EE_cr_Rlen_XX_Contig_5980_Len1557_Cov37_OG5_126843_Trans1:0.83030089436051290885,(EE_cr_Gspb_XX_Contig_3843_Len1509_Cov70_OG5_126843_Trans1:0.42478298276001907885,(EE_cr_Rlen_XX_Contig_5607_Len1449_Cov94_OG5_126843_Trans1:0.26658639140406120704,(EE_cr_Hphi_XX_Contig_6511_Len1425_Cov100_OG5_126843_Trans1:0.21950953829815897378,EE_cr_Ccur_XX_Contig_3766_Len1437_Cov143_OG5_126843_Trans1:0.29697100928922776086):0.06044691965664333172):0.13710185620650303107):0.07216010764730113047):0.08974498323266388855):0.02861214174343811381,(EE_ap_Ftro_XX_Contig_939_Len1908_Cov9_OG5_126843:0.52223563950191342897,(Op_ch_mbre_gw1.5.103.1_OG5_126843:0.48452829265612001874,Sr_rh_Lamo_XX_Contig_5909_Len1386_Cov218_OG5_126843_Trans1:0.45283810950641756321):0.07124592224608709834):0.07164353522498337346):0.05851922290863217913,((Op_ic_Cowc_EFW42887.1:0.48992055916531895843,((((Sr_rh_Phya_XX_Contig_673_Len465_OG5_126843:0.23480158978152981808,Sr_rh_Phya_XX_Contig_383_Len393_OG5_126843:0.19754554785564973352):0.13532867479998195170,(((Op_me_amel_XP623831_OG5_126843:0.29766855608305387015,(Op_me_agam_AGAP012407-PAm_OG5_126843:0.08766623818104515853,(Op_me_cpip_CPIJ005219_OG5_126843:0.00000100000050002909,Op_me_cpip_CPIJ013559_OG5_126843:0.00405938558615993698):0.08962091197457160152):0.20709418094888973338):0.09696904937520058054,((Op_me_Hrob_555690080:0.73988726287380035540,Op_me_Hrob_555708103:0.22108310709907205704):0.10267224664379459576,(Op_me_Bflo_260821328:0.00133802684242644975,Op_me_Bflo_260801796:0.00303712892655254828):0.20241733698977010247):0.04230858317754827153):0.04384455466827966857,(Op_me_ggal_ENSGALP00000011689_OG5_126843:0.01565408367958995642,(Op_me_oana_ENSOANP00000009768_OG5_126843:0.05910773061984994636,((Op_me_ptro_ENSPTRP00000016636_OG5_126843:0.00456545081313346549,Op_me_mmul_XP002800695_OG5_126843:0.10111563731540890165):0.00930629534656852477,(Op_me_ecab_XP001489541_OG5_126843:0.03993790595810710331,Op_me_rnor_ENSRNOP00000051841_OG5_126843:0.01914954237110791727):0.00499374672537351329):0.02843600151571777299):0.05106494251989562166):0.15967075426250332271):0.03720200964808671046):0.05436947599901767458,(Op_me_Lbai_XX_GH556967_Contig_582_length_643_Op_me_Lbai_OG5_126843:0.24592844467252827911,Op_me_Cfol_XX_GO085666_Contig_2186_length_744_Op_me_Cfol_OG5_126843:0.31140821813735097834):0.02707315614969139916):0.03902821443173278698,(Op_me_Odio_XX_CBY14738_Contig_4533_length_1422_OG5_126843:0.25591371058123546511,Op_me_Odio_XX_CBY09721_Contig_9401_length_819_OG5_126843:0.26084212548758972661):0.20499236077392740674):0.04752672026381277864):0.05431623376143203397,(Op_ch_mbre_estExtGenewise1.C70127_OG5_126843:0.43400328738286625407,Pl_gr_Cvar_552844396:0.26346053243553146928):0.10870108709372983125):0.09005029903216778264):0.06106366562359076544,(((EE_ce_Acsp_XX_Contig_3585_Len1503_Cov10_OG5_126843_Trans1:0.84303770936043165296,(EE_ce_Rhet_XX_Contig_7471_Len1464_Cov918_OG5_126843_Trans1:0.22874127366990892885,(EE_ce_Acsp_XX_Contig_4875_Len1047_Cov35_OG5_126843_Trans1:0.75902932740368922460,EE_ce_Acsp_XX_Contig_1610_Len1455_Cov352_OG5_126843_Trans1:0.10177285771858810004):0.10053708673582030519):0.12441907821065165130):0.07863766934111071638,((Sr_st_Chys_XX_Contig_10983_Len828_Cov32_OG5_126843:0.54736866793597227510,(Sr_st_Npun_XX_Contig_11354_Len768_Cov26_OG5_126843:0.43093955173102582723,Sr_st_Apal_XX_Contig_15331_Len747_Cov19_OG5_126843_Trans1:0.53037098445015207204):0.28131117190707599640):1.05660621135191723496,((Sr_st_Esil_XX_CBJ29913_Contig_5980_Len1485_OG5_126843:0.41746138850145508847,(Sr_st_Rmar_XX_Contig_2258_Len1263_Cov192_OG5_126843_Trans1:0.19627849498873059519,Sr_st_Fspa_XX_Contig_3573_Len1497_Cov20_OG5_126843:0.27394948057347789927):0.16032520205781358724):0.18390708111169259964,(Sr_st_Ospa_XX_Contig_2149_Len1449_Cov884_OG5_126843_Trans1:0.69563873643702289762,Sr_st_Arad_XX_Contig_5707_Len1482_Cov25_OG5_126843_Trans1:0.31714462097998125101):0.11383483959994317170):0.02203620620348853767):0.06610977634145646931):0.06824461789640910714,EE_ha_Unid_XX_Contig_833_Len1449_Cov18_OG5_126843_Trans1:1.15039603633753939604):0.00000100000050002909):0.04341986964178908004):0.04093501382906210295,(Pl_gr_Bpra_412989178:0.39331438456639783441,Pl_gr_otau_0500010276_OG5_126843:0.42841141750791283949):0.67550284426564166651):0.04004655643103614132,((((Am_di_Vrob_XX_Contig_3939_Len1422_Cov10_OG5_126843_Trans1:0.65821190964230802400,(Am_va_Usch_XX_Contig_643_Len1380_Cov59_OG5_126843_Trans1:0.18562667732721888081,Am_di_Vrob_XX_Contig_4503_Len1410_Cov661_OG5_126843_Trans1:0.23718084912832596012):0.11844981902166248910):0.31041127044730754125,(Am_ar_Mbal_XX_EC712446_Contig_261_length_785_Am_ar_Mbal_EST_OG5_126843:0.51481323702778325480,Am_di_Mspa_XX_Contig_5619_Len1431_Cov30_OG5_126843_Trans1:0.35932330450176391823):0.20930197228582372859):0.11953278196923261545,(Am_di_Acas_XX_Contig_3095_length_1824_OG5_126843:0.65605465601823653721,(Am_hi_Gfon_XX_Contig_5050_length_1425_Am_hi_Gfon_OG5_126843:0.42889695917026010585,(Am_di_Acas_XX_Contig_4726_length_1449_OG5_126843:0.28409789995179640654,Am_di_Acas_XX_Contig_4825_length_1431_OG5_126843:0.69833040123726086712):0.20144048977236889142):0.04801934849544194084):0.13740847489037019691):0.06431517007014632414,(((((EE_ha_Ehux_551538645:0.02028104710377271955,(EE_ha_Ehux_551608513:0.00000100000050002909,EE_ha_Ehux_551583649:0.00478355174776149035):0.00000100000050002909):0.67742157199520236510,EE_ha_Ehux_551538581:0.41048988535461544691):0.26580662763279433447,Pl_gr_Pspa_XX_Contig_129_Len1791_Cov34_OG5_126843_Trans1:0.88007758804240809880):0.29460158956795218232,((Pl_gl_Cglo_XX_Contig_1478_Len1476_Cov10_OG5_126843_Trans1:0.91128460060060034742,(Am_is_Vant_XX_Contig_4656_Len1707_Cov53_OG5_126843_Trans1:0.88842305898425499855,Am_th_Tqua_XX_Contig_3385_Len1593_Cov16_OG5_126843_Trans1:0.88935907143897363625):0.11764294394467690597):0.06703635864196730298,Am_th_Tqua_XX_Contig_4489_Len1428_Cov57_OG5_126843_Trans1:0.84725472921131628201):0.04403375975537040382):0.06148353816571344721,Am_is_Vant_XX_Contig_5937_Len1572_Cov1606_OG5_126843_Trans1:0.52817640231220275293):0.03093906543706720705):0.13181592355717078324):0.05197687989529831826,(((EE_cr_Gspb_XX_Contig_3715_Len1737_Cov190_OG5_126843_Trans1:0.79236342668034676073,((EE_cr_Hphi_XX_Contig_1366_Len1671_Cov67_OG5_126843_Trans1:0.24399920193600757590,EE_cr_Rlen_XX_Contig_2843_Len1704_Cov32_OG5_126843_Trans1:0.29047547403618551032):0.08694529461861133512,(EE_cr_Ccur_XX_Contig_2522_Len1746_Cov35_OG5_126843_Trans1:0.22014807202947750397,EE_cr_Ccur_XX_Contig_2575_Len1374_Cov84_OG5_126843_Trans1:0.44473865275187579593):0.14285698852500994249):0.40647004797333813331):0.07822412222133164061,((EE_cr_Htep_XX_Contig_2441_Len1632_Cov71_OG5_126843_Trans1:0.40142512562762483475,EE_cr_Rlen_XX_Contig_4927_Len1392_Cov8_OG5_126843_Trans1:0.41228341308729021808):0.72964821289537507010,((Pl_gr_Mspa_XX_Contig_2390_Len1470_Cov20_OG5_126843_Trans1:0.24778612067252805962,Pl_gr_Mpus_303278490:0.31098414900819887352):0.38051836417941514723,(((Pl_gr_atha_NP173594_OG5_126843:0.00244698369359632644,Pl_gr_atha_NP849696_OG5_126843:0.01770474004138434393):0.05350889309395497062,Pl_gr_atha_NP177875_OG5_126843:0.09523667682532073464):0.36998464359043559124,Pl_gr_Psal_XX_Contig_5095_Len1047_Cov573_OG5_126843_Trans1:0.52966539359033315559):0.06794143231966622010):0.11913268752801614569):0.01280030390874148036):0.06550980773473404073,((((Op_me_Odio_XX_CBY06760_Contig_3320_length_1677_OG5_126843:0.65123068294253660948,((Op_me_Bflo_260793862:0.26805759663732181552,((Op_me_oana_ENSOANP00000000478_OG5_126843:0.07990894731346698254,(Op_me_ecab_XP001915778_OG5_126843:0.01789430280051933567,(Op_me_rnor_ENSRNOP00000008728_OG5_126843:0.04849074542465598575,(Op_me_mmul_XP002803559_OG5_126843:0.00562502902342595403,Op_me_ptro_ENSPTRP00000033973_OG5_126843:0.00921244567582421917):0.03200084395261786757):0.01784281433435137842):0.02244365462615059703):0.01390616428805972228,Op_me_ggal_ENSGALP00000020268_OG5_126843:0.07960662261023357389):0.22527522671622962624):0.04967074251110372568,(Op_me_Tspi_339258360:0.48753184140200839236,Op_me_Hrob_555699477:0.46815760695283559212):0.07247286394929790265):0.11155284879161198042):0.13801557549425250149,((((Op_me_Hech_XX_DT621374_Contig_758_length_619_Op_me_Hech_OG5_126843:0.11194209579809089039,Op_me_Hvul_449671849:0.19965686596679838027):0.20515640875818694222,((Op_me_Bflo_260814670:0.23188958144046448173,(Op_me_ggal_ENSGALP00000013574_OG5_126843:0.06081437137346058786,(Op_me_oana_ENSOANP00000015379_OG5_126843:0.01696560378055240320,(((Op_me_ptro_ENSPTRP00000051552_OG5_126843:0.00229570116428205436,Op_me_mmul_XP001109061_OG5_126843:0.00369653866817176453):0.00727963065800657505,Op_me_ecab_XP001503038_OG5_126843:0.00964679277349525174):0.01099008030989092818,Op_me_rnor_ENSRNOP00000020478_OG5_126843:0.04051476017167427585):0.03669131407779051279):0.03314320492700274789):0.25200152395189306942):0.05633823654788074065,((Op_me_amel_XP623282_OG5_126843:0.22279943914672889482,(Op_me_agam_AGAP007393-PAm_OG5_126843:0.03953117336756295508,Op_me_cpip_CPIJ015893_OG5_126843:0.08148135546614584945):0.19829213853218682906):0.20593176224569187394,(Op_me_Hrob_555708167:0.92608407005873583095,Op_me_Hrob_555688557:0.22449683805738951392):0.09763452970500137440):0.08696974108730569364):0.05748054808436078161):0.08997772846284714243,(Op_me_Odio_XX_CBY21606_Contig_4421_length_1440_OG5_126843:0.44293642068316618765,Op_me_Odio_XX_CBY13112_Contig_4059_length_1506_OG5_126843:0.91117131005146023703):0.08707051486757666492):0.06677918546908605257,Op_ch_mbre_fgenesh2pg.scaffold18000062_OG5_126843:0.60132462401989128509):0.04271156331858583233):0.05079460229938542043,((EE_ap_Ftro_XX_Contig_1543_Len1467_Cov6_OG5_126843:0.42828475064567800290,Am_di_Pess_XX_Contig_3609_Len1449_Cov289_OG5_126843_Trans1:0.52475467427807065413):0.09514410348237813120,(Sr_st_Rmar_XX_Contig_6769_Len1326_Cov42_OG5_126843_Trans1:0.38710598600915463452,(Sr_st_Ospa_XX_Contig_4505_Len1530_Cov39_OG5_126843_Trans1:0.50411657137676746387,EE_ha_Clep_XX_Contig_3601_Len1188_Cov47_OG5_126843_Trans1:0.43590324953893855309):0.10335223960042735913):0.38863201043383288624):0.03485025275434509850):0.08914434800645586188,(Pl_gr_Pspf_XX_Contig_2539_Len1554_Cov759_OG5_126843_Trans1:1.08711021252421913275,(EE_cr_Ccur_XX_Contig_3940_Len1500_Cov88_OG5_126843_Trans1:0.72733679933057238109,(Am_ar_Mbal_XX_EC711598_Contig_799_length_742_Am_ar_Mbal_EST_OG5_126843:0.68550626907951661693,(Ex_eu_Ndes_XX_Contig_4516_Len1656_Cov8_OG5_126843_Trans1:0.40338781981896976170,Ex_eu_Ndes_XX_Contig_6569_Len1161_Cov11_OG5_126843_Trans1:0.17564405743526448345):0.78208784752821702835):0.15921761883045101871):0.05757220932537771702):0.09634309401868793032):0.10851115127740661170):0.08267518146536147194):0.02958856243487643575):0.04097173113074938089):0.02388577088262061610):0.03268842518058130697):0.07090484927858912989,(((((Ex_eu_lbra_LbrM.35.7320_OG5_126843:0.07916481344087761274,(Ex_eu_lmex_LmxM.36.6940_OG5_126843:0.05258269067965758159,Ex_eu_lmaj_LmjF.36.6940_OG5_126843:0.04318347867183600780):0.07360662524853982258):0.32183829803996200081,((Ex_eu_Scul_XX_Contig_7043_length_843_OG5_126843:0.00000100000050002909,(Ex_eu_Scul_XX_Contig_3127_length_1413_OG5_126843:0.00000100000050002909,Ex_eu_Scul_XX_Contig_3124_length_1413_OG5_126843:0.00000100000050002909):0.00000100000050002909):0.00000100000050002909,Ex_eu_Scul_XX_Contig_9633_length_540_OG5_126843:0.00000100000050002909):0.51864095479334104688):0.11213049862268091583,((Ex_eu_tcru_Tc00.1047053507611.370_OG5_126843:0.01028043173119128247,Ex_eu_tcru_Tc00.1047053506247.10_OG5_126843:0.00000100000050002909):0.19424826511544932006,(Ex_eu_tviv_TvY4861008110_OG5_126843:0.39832192414217637566,(Ex_eu_tbru_Tb927.10.8230_OG5_126843:0.00202714645578040871,(Ex_eu_tbrg_Tbg972.10.10050_OG5_126843:0.00000100000050002909,Ex_eu_tbru_Tb427.10.8230_OG5_126843:0.00000100000050002909):0.00000100000050002909):0.38964294536454541218):0.105289053511258
[…]
