### 2021-08-09 b879e6c
run physcraper
