+++ docs/mds/results.md
@@ -1,29 +1,0 @@
-## Files generated by a Physcraper run
-
-Within the output directory defined by the user, a Physcraper run generates three subdirectories which are labeled with a "tag", corresponding to the file name of the input alignment.
-
-The subdirectories consist of:
-
-
-### - Input files
-
-Within the `inputs_TAG` directory, Physcraper writes tree and alignment files used in a Physcraper run for the sake of reproducibility, taxon name matching, and taxon reconciliation.
-It also writes the ".config" file down if none was provided, as well as the results of the mapping of the tree taxon labels, saved as "otu_info.csv"
-
-### - Run files
-
-Within the `run_TAG` directory, all run files are also automatically written down: Blast runs, alignments, RAxML trees, bootstrap results, etc.
-
-Intermediate processing files, and the json formatted otu information are also stored here. Many fo these files are re-used in the event that the analysis crashes and is restarted. Make sure you use a new output directory or otherwise empty this folder if you want to modify the initial run parameters.
-
-The trees are reconstructed using RAxML, with tip labels corresponding to local ids (e.g., otu42009, otuPS1) and not taxon names (e.g., *Ceiba*), nor taxonomic ids (e.g., ott or ncbi). Branch lengths are proportional to relative substitution rates.
-
-### - Output files
-
-In the `output_TAG` folder, an updated tree in nexson format, containing all types of tip labels is saved in here.
-From this tree, a tree with any kind of label can be produced.
-By default, the updated tree with taxon names as tip labels is saved as "updated_taxonname.tre". See section [relabeling the trees](# Relabeling) below for more info on how to generate trees with different types of labels.
-
-The "seqlen_mismatch.txt" file contains the acession numbers, taxa, and sequence lengths of BLAST matches that didn't meet the sequence length cutoffs.
-
-Finally, a summary CSV file with information about original and newly added sequences is generated.
