+++ docs/mds/localDB.md
@@ -4,8 +4,8 @@ The BLAST tool can be run using local databases, which can be downloaded and upd
 
 ### Installing BLAST command line tools
 
-To blast locally you will need to install blast command line tools first.  
-If you perfomed the dependency installation using conda, the command line tools will already be installed.
+To BLAST locally you will need to install BLAST command line tools first.
+If you perfomed the Physcraper installation using conda, the BLAST command line tools will already be installed.
 
 Find general instructions at
 https://www.ncbi.nlm.nih.gov/books/NBK279671/
@@ -18,7 +18,7 @@ e.g. installing BLAST command line tools on **linux**:
     wget https://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/LATEST/ncbi-blast-2.11.0+-x64-linux.tar.gz
     tar -xzvf ncbi-blast-2.11.0+-x64-linux.tar.gz
  ```
-This link may be broken by NCBI blast executables updates - if so check https://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/LATEST/ for the newest version.
+This link may be broken by NCBI BLAST executables updates - if so check https://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/LATEST/ for the newest version.
 
 The binaries/scripts/executables will be installed in the `/bin` folder.
 
@@ -29,7 +29,7 @@ If your terminal uses zshell instead of bash, make sure you're running the .bash
 
 ### Downloading the NCBI database
 
-If you want to download the NCBI blast database and taxonomy for faster local searches
+If you want to download the NCBI BLAST database and taxonomy for faster local searches
 note that the download can take several hours, depending on your internet connection.
 
 This is what you should do:
@@ -51,10 +51,10 @@ This is what you should do:
     gunzip -f -cd taxdump.tar.gz | (tar xvf - names.dmp nodes.dmp)
 ```
 
-#### Updating an existing blast db
+#### Updating an existing BLAST database
 
-``` 
-    cd local_blast_db  # move to the nblast db folder
+```
+    cd local_blast_db  # move to the nucleotide database folder
     update_blastdb nt  # download the NCBI nucleotide databases
     # update_blastdb.pl nt  # on Mac OS
     cat *.tar.gz | tar -xvzf - --ignore-zeros  # unzip the nucleotide databases
@@ -63,25 +63,25 @@ This is what you should do:
     gunzip -cd taxdb.tar.gz | (tar xvf - )  # unzip the taxonomy database
 ```
 
-### Checking install success of local blast DB
+### Checking install success of local BLAST database
 
     physcraper_run.py --study_id pg_55 --tree_id tree5864 --treebase --bootstrap_reps 10 -db local_blast_db --output pg_55_local
 
-This should start running a query using your local blast DB.
+This should start running a query using your local BLAST database.
 
 
-### Setting up an AWS blast db
+### Setting up an AWS BLAST database
 
-To run blast searches without NCBI's required time delays, you can set up your own server on AWS (for $).
-See instructions at (AWS marketplace NCBI blast)[https://aws.amazon.com/marketplace/pp/NCBI-NCBI-BLAST/B00N44P7L6]
+To run BLAST searches without NCBI's required time delays, you can set up your own server on AWS (for $).
+See instructions at [AWS marketplace NCBI BLAST](https://aws.amazon.com/marketplace/pp/NCBI-NCBI-BLAST/B00N44P7L6)
 
 ### Create an NCBI API key
 
-Generating an NCBI API key will speed up downloading full sequences following blast searches.
[…]
