### 2021-08-10 015c852
adding blast + citation to requirements
