+++ docs/mds/description.md
@@ -7,33 +7,34 @@ These data are often appropriate for looking at phylogenetic relationships, and
 have the advantage of being homologous to genetic sequences used to construct existing
 trees.
 
-If you have access to a single gene DNA alignment and a tree, Physcraper automates
-adding new lineage samples into your tree by using [Open Tree of Life](#the-open-tree-of-life) tools coupled to the [Blast](https://blast.ncbi.nlm.nih.gov/Blast.cgi) algorithm to search for loci in [GenBank](https://www.ncbi.nlm.nih.gov/genbank/statistics/) that are likely to be homologous to sequences in the initial DNA alignment.
+If you have access to a single gene or multilocus DNA alignment, and a phylogenetic tree, Physcraper automates
+adding DNA sequences of new lineage samples into your tree by using [Open Tree of Life](#the-open-tree-of-life) tools to reconcile taxonomies, and the [BLAST](https://blast.ncbi.nlm.nih.gov/Blast.cgi) algorithm to search for loci in the [GenBank](https://www.ncbi.nlm.nih.gov/genbank/statistics/) genetic database that are likely to be
+homologous (both orthologs and paralogs, see FAQs) to sequences in the initial DNA alignment.
+
+By using a starting alignment and tree, Physcraper takes advantage of DNA loci and homology hypotheses (ideally orthology) that
+previous researchers have assessed, curated, and deemed appropriate for the phylogenetic scope.
+The sequences added during the BLAST search are limited either to a user specified taxon or
+monophyletic group, or within the taxonomic scope of the ingroup of the starting tree.
+
+These automated trees can provide a quick inference of potential phylogenetic relationships,
+of problems in the taxonomic assignments of sequences, paralogy and orthology, as well as flag areas of potential systematic interest.
 
 <br/>
 
 ![](../img/schematic-final.svg)
 
 Figure 1 from [Sanchez-Reyes et al. 2021](https://doi.org/10.1186/s12859-021-04274-6):
-The Physcraper framework consists of 4 general steps. The methodology is extensively described in the [Implementation](https://physcraper.readthedocs.io/en/latest/implementation.html) section of this documentation.
+The Physcraper framework consists of 4 general steps. The methodology is further described in the [Implementation](https://physcraper.readthedocs.io/en/latest/implementation.html) section of this documentation.
 
 <br/>
-
-By using a starting alignment and tree, Physcraper takes advantage of loci that
-previous researchers have assessed and deemed appropriate for the phylogenetic scope.
-The sequences added in the search are limited either to a user specified taxon or
-monophyletic group, or within the taxonomic scope of the ingroup of the starting tree.
-
-These automated trees can provide a quick inference or potential relationships,
-of problems in the taxonomic assignments of sequences, and flag areas of potential systematic interest.
-
 <br/>
 
 
 ## The Open Tree of Life
 
-The Open Tree of Life ([OpenTree](https://tree.opentreeoflife.org/opentree/argus/opentree13.4@ott93302)) is a project that unites phylogenetic inferences and taxonomy
-to provide a synthetic estimate of species relationships across the entire tree of life.
+The Open Tree of Life (OpenTree) is a project that unites expert, peer-reviwed [phylogenetic inferences](https://github.com/OpenTreeOfLife/phylesystem-1) and
+[taxonomy](https://tree.opentreeoflife.org/about/taxonomy-version/ott3.3)
+to generate a [synthetic tree](https://tree.opentreeoflife.org/opentree/argus/opentree13.4@ott93302) estimate of species relationships across all life.
 
 <br/>
 
@@ -50,7 +51,7 @@ Currently the tree comprises 2.3 million tips.
[…]
