+++ docs/mds/faq.md
@@ -1,5 +1,14 @@
 Frequently asked questions
 
+## How does Physcraper handle paralogs?
+
+Physcraper allows multiple tips to be mapped to the same taxon, with an added unique identifier that allows linking the sequence back to the original one on GenBank.
+This means that newly added homologous DNA sequences can include both orthologs
+and paralogs and that a phylogenetic analysis can be performed.
+While we expect that most curated alignments are composed of ortholog sequences, the algorithm used to find new sequences is unable to distinguish paralogs, so it is likely that -- if existing, they will be automatically added to the dataset.
+Users should check the output phylogenetic trees to detect paralogs and filter them
+appropriately if needed.
+
 ## I have to learn to use OpenTree to use Physcraper, is the learning curve worth it?
 
 We think that this decision depends on the goals of the user.
@@ -19,7 +28,7 @@ For this goal, having a minimum familiarity with the OpenTree tools is needed.
 We realize that this might initially discourage some users, but we believe that the benefits brought by connecting taxonomic data with the OpenTree services will encourage users to familiarize with the OpenTree services, and to adopt the use of Physcraper.
 
 
-## How does Physcraper handle polytomies of starting trees?
+## How does Physcraper handle polytomies in a starting trees?
 
 The Physcraper starting tree is a phylogeny whose tip labels must have been standardized to the OpenTree Taxonomy (as described in the Introduction section:
 [Mapping names to taxa](https://physcraper.readthedocs.io/en/latest/quick-start.html#updating-your-own-tree-and-alignment)).
@@ -45,8 +54,7 @@ around the early diverging branches,
 the automatic rooting is problematic and can have multiple solutions.
 
 
-
 ## How does Physcraper use the starting alignment?
 
-Physcraper uses the input DNA alignment (single or multiple marker) to mine the GenBank database with the goal of increasing the
+Physcraper uses all unique DNA sequences in the input alignment to mine a genetic database using the BLAST algorithm, with the goal of increasing the
 lineage sampling of the alignment within a given biological group.
