+++ docs/mds/intro.md
@@ -1,3 +1,5 @@
+## Automated gene tree updating <br> with the Open Tree of Life\!
+
 Use a phylogenetic tree and a DNA alignment to automatically find and add nucleotide sequences from a genetic database, to reproducibly improve and advance phylogenetic knowledge in a group.
 
 Physcraper relies on [taxonomic](https://tree.opentreeoflife.org/about/taxonomy-version/ott3.3) and [phylogenetic](https://github.com/OpenTreeOfLife/phylesystem-1) resources and [programmatic tools](https://github.com/OpenTreeOfLife/germinator/wiki/Open-Tree-of-Life-Web-APIs) from the [Open Tree of Life](https://tree.opentreeoflife.org/opentree/argus/opentree12.3@ott93302) project:
