+++ README.Rmd
@@ -16,7 +16,7 @@ IS_README <- TRUE
 
 <img align="left" width="250" src="https://raw.githubusercontent.com/McTavishLab/physcraper/main/docs/physcraper-long.png">
 
-# Physcraper
+# Welcome to Physcraper’s repository!
 
 [![Build Status](https://travis-ci.org/McTavishLab/physcraper.svg?branch=main)](https://travis-ci.org/McTavishLab/physcraper)[![Documentation](https://readthedocs.org/projects/physcraper/badge/?version=latest&style=flat)](https://physcraper.readthedocs.io/en/latest/)[![codecov](https://codecov.io/gh/McTavishLab/physcraper/branch/main/graph/badge.svg)](https://codecov.io/gh/McTavishLab/physcraper)
 
@@ -24,9 +24,6 @@ IS_README <- TRUE
 <p></p>
 
 <p></p>
-
-
-## Automated gene tree updating with the Open Tree of Life!
 
 
 ```{r child="docs/mds/intro.md"}
