+++ docs/mds/description.md
@@ -4,19 +4,19 @@ While genome scale data is increasing rapidly, there are still large quantities
 of gene-sequence data being uploaded to the US National Center on Biotechnology
 Information (NCBI) database [GenBank](https://www.ncbi.nlm.nih.gov/genbank/statistics/).
 These data are often appropriate for looking at phylogenetic relationships, and
-have the advantage of being homologous to genetic sequences used to construct existing
+have the advantage of being orthologous to genetic sequences used to construct existing
 trees.
 
 If you have access to a single gene or multilocus DNA alignment, and a phylogenetic tree, Physcraper automates
 adding DNA sequences of new lineage samples into your tree by using [Open Tree of Life](#the-open-tree-of-life) tools to reconcile taxonomies, and the [BLAST](https://blast.ncbi.nlm.nih.gov/Blast.cgi) algorithm to search for loci in the [GenBank](https://www.ncbi.nlm.nih.gov/genbank/statistics/) genetic database that are likely to be
-homologous (both orthologs and paralogs, see FAQs) to sequences in the initial DNA alignment.
+homologous (both orthologs and paralogs, see [FAQs](https://physcraper.readthedocs.io/en/stable/faq.html)) to sequences in the initial DNA alignment.
 
 By using a starting alignment and tree, Physcraper takes advantage of DNA loci and homology hypotheses (ideally orthology) that
 previous researchers have assessed, curated, and deemed appropriate for the phylogenetic scope.
 The sequences added during the BLAST search are limited either to a user specified taxon or
 monophyletic group, or within the taxonomic scope of the ingroup of the starting tree.
 
-These automated trees can provide a quick inference of potential phylogenetic relationships,
+These automated, reproducible trees can provide a quick inference of potential phylogenetic relationships,
 of problems in the taxonomic assignments of sequences, paralogy and orthology, as well as flag areas of potential systematic interest.
 
 <br/>
@@ -32,7 +32,7 @@ The Physcraper framework consists of 4 general steps. The methodology is further
 
 ## The Open Tree of Life
 
-The Open Tree of Life (OpenTree) is a project that unites expert, peer-reviwed [phylogenetic inferences](https://github.com/OpenTreeOfLife/phylesystem-1) and
+The Open Tree of Life (OpenTree) is a project that unites expert, peer-reviewed [phylogenetic inferences](https://github.com/OpenTreeOfLife/phylesystem-1) and
 [taxonomy](https://tree.opentreeoflife.org/about/taxonomy-version/ott3.3)
 to generate a [synthetic tree](https://tree.opentreeoflife.org/opentree/argus/opentree13.4@ott93302) estimate of species relationships across all life.
 
