+++ docs/mds/citation.md
@@ -2,4 +2,6 @@
 
 If you use Physcraper, please cite:
 
-Sánchez-Reyes, L.L., Kandziora, M. & McTavish, E.J. *Physcraper: a Python package for continually updated phylogenetic trees using the Open Tree of Life.* BMC Bioinformatics 22, 355 (2021) doi: [doi.org/10.1186/s12859-021-04274-6](https://doi.org/10.1186/s12859-021-04274-6)
+- Sánchez-Reyes, L.L., M. Kandziora, & E.J McTavish. (2021). *Physcraper: a Python package for continually updated phylogenetic trees using the Open Tree of Life*. BMC Bioinformatics 22, 355. doi: [doi.org/10.1186/s12859-021-04274-6](https://doi.org/10.1186/s12859-021-04274-6).
+<br><br>
+- Open Tree of Life, B. Redelings, L.L. Sanchez Reyes, K.A. Cranston, J. Allman, M.T. Holder, & E.J. McTavish. (2019). *Open Tree of Life Synthetic Tree (Version 12.3)*. Zenodo. doi: [10.5281/zenodo.3937741](https://doi.org/10.5281/zenodo.3937741)
