+++ docs/mds/intro-header.md
@@ -1,4 +1,4 @@
 
 [![Build Status](https://travis-ci.org/McTavishLab/physcraper.svg?branch=main)](https://travis-ci.org/McTavishLab/physcraper)[![Documentation](https://readthedocs.org/projects/physcraper/badge/?version=latest&style=flat)](https://physcraper.readthedocs.io/en/latest/)[![codecov](https://codecov.io/gh/McTavishLab/physcraper/branch/main/graph/badge.svg)](https://codecov.io/gh/McTavishLab/physcraper)
 
-<img align="right" width="180" src="https://raw.githubusercontent.com/McTavishLab/physcraper/main/docs/physcraper-long.png">
+<img align="right" width="200" src="https://raw.githubusercontent.com/McTavishLab/physcraper/main/docs/physcraper-long.png">
