+++ docs/mds/intro.md
@@ -2,6 +2,4 @@
 
 Use a phylogenetic tree and a DNA alignment to automatically find and add nucleotide sequences from a genetic database, to reproducibly improve and advance phylogenetic knowledge in a group.
 
-Physcraper relies on [taxonomic](https://tree.opentreeoflife.org/about/taxonomy-version/ott3.3) and [phylogenetic](https://github.com/OpenTreeOfLife/phylesystem-1) resources and [programmatic tools](https://github.com/OpenTreeOfLife/germinator/wiki/Open-Tree-of-Life-Web-APIs) from the [Open Tree of Life](https://tree.opentreeoflife.org/opentree/argus/opentree12.3@ott93302) project:
-<br/>
-*Open Tree of Life, Benjamin Redelings, Luna Luisa Sanchez Reyes, Karen A. Cranston, Jim Allman, Mark T. Holder, & Emily Jane McTavish. (2019). Open Tree of Life Synthetic Tree (Version 12.3). Zenodo* doi: [10.5281/zenodo.3937741](https://doi.org/10.5281/zenodo.3937741)
+Physcraper relies on [taxonomic](https://tree.opentreeoflife.org/about/taxonomy-version/ott3.3) and [phylogenetic](https://github.com/OpenTreeOfLife/phylesystem-1) resources and [programmatic tools](https://github.com/OpenTreeOfLife/germinator/wiki/Open-Tree-of-Life-Web-APIs) from the [Open Tree of Life](https://tree.opentreeoflife.org/opentree/argus/opentree12.3@ott93302) project.
