+++ docs/mds/description.md
@@ -1,23 +1,22 @@
 ## The Physcraper framework
 
 While genome scale data is increasing rapidly, there are still large quantities
-of gene-sequence data being uploaded to the US National Center on Biotechnology
+of single locus nucleotide sequence data being uploaded to the US National Center on Biotechnology
 Information (NCBI) database [GenBank](https://www.ncbi.nlm.nih.gov/genbank/statistics/).
 These data are often appropriate for looking at phylogenetic relationships, and
-have the advantage of being orthologous to genetic sequences used to construct existing
-trees.
+have the advantage of being orthologous to genetic sequences that have been used to construct existing phylogenetic trees.
 
-If you have access to a single gene or multilocus DNA alignment, and a phylogenetic tree, Physcraper automates
-adding DNA sequences of new lineage samples into your tree by using [Open Tree of Life](#the-open-tree-of-life) tools to reconcile taxonomies, and the [BLAST](https://blast.ncbi.nlm.nih.gov/Blast.cgi) algorithm to search for loci in the [GenBank](https://www.ncbi.nlm.nih.gov/genbank/statistics/) genetic database that are likely to be
-homologous (both orthologs and paralogs, see [FAQs](https://physcraper.readthedocs.io/en/stable/faq.html)) to sequences in the initial DNA alignment.
+If you have access to a single gene or multilocus nucleotide alignment, and a phylogenetic tree, Physcraper automates
+adding nucleotide sequences of new lineage samples into your tree by using [Open Tree of Life](#the-open-tree-of-life) tools to reconcile Taxonomy, and the [BLAST](https://blast.ncbi.nlm.nih.gov/Blast.cgi) algorithm to search for loci in the [GenBank](https://www.ncbi.nlm.nih.gov/genbank/statistics/) genetic database that are likely to be
+locally similar to sequences in the initial DNA alignment.
 
-By using a starting alignment and tree, Physcraper takes advantage of DNA loci and homology hypotheses (ideally orthology) that
+By using a starting alignment and tree, Physcraper takes advantage of DNA loci alignments as  homology hypotheses (ideally orthology, see [FAQs](https://physcraper.readthedocs.io/en/stable/faq.html)) that
 previous researchers have assessed, curated, and deemed appropriate for the phylogenetic scope.
-The sequences added during the BLAST search are limited either to a user specified taxon or
+The sequences added during a BLAST search are limited either to a user specified taxon or
 monophyletic group, or within the taxonomic scope of the ingroup of the starting tree.
 
 These automated, reproducible trees can provide a quick inference of potential phylogenetic relationships,
-of problems in the taxonomic assignments of sequences, paralogy and orthology, as well as flag areas of potential systematic interest.
+as well as flag problems in the taxonomic assignments of sequences, paralogy and orthology, and areas of potential systematic interest.
 
 <br/>
 
