### 2021-08-11 McTavishLab/physcraper#165 @LunaSare
Added this Q to FAQ section here https://physcraper.readthedocs.io/en/stable/faq.html
