+++ docs/mds/outputs.md
@@ -1,6 +1,6 @@
 ## Files generated by a Physcraper run
 
-Within the output directory defined by the user, a Physcraper run generates three subdirectories which are labeled with a "tag", corresponding to the file name of the input alignment.
+Within the output directory defined by the user, a Physcraper run generates various subdirectories which are labeled with a TAG name that corresponds to the file name of the input alignment.
 
 The subdirectories consist of:
 
@@ -18,12 +18,17 @@ Intermediate processing files, and the json formatted otu information are also s
 
 The trees are reconstructed using RAxML, with tip labels corresponding to local ids (e.g., otu42009, otuPS1) and not taxon names (e.g., *Ceiba*), nor taxonomic ids (e.g., ott or ncbi). Branch lengths are proportional to relative substitution rates.
 
+### - BLAST files
+
+The `blast_TAG` folder contains XML files with BLAST results for each sequence in the input alignment.
+
 ### - Output files
 
-In the `output_TAG` folder, an updated tree in nexson format, containing all types of tip labels is saved in here.
-From this tree, a tree with any kind of label can be produced.
-By default, the updated tree with taxon names as tip labels is saved as "updated_taxonname.tre". See section [relabeling the trees](# Relabeling) below for more info on how to generate trees with different types of labels.
+In the `output_TAG` folder, the updated tree with taxon names as tip labels is saved as "updated_taxonname.tre".
+A version of the updated tree in nexson format containing several types of tip labels is also saved here. From the nexson tree, a tree with any
+label can be produced.
+See section [relabeling the trees](#id1) below for more info on how to generate trees with different types of labels.
 
 The "seqlen_mismatch.txt" file contains the acession numbers, taxa, and sequence lengths of BLAST matches that didn't meet the sequence length cutoffs.
 
-Finally, a summary CSV file with information about original and newly added sequences is generated.
+Finally, the CSV file "otu_info_TAG.csv" contains a summary of information about original and newly added sequences.
