+++ docs/mds/faq.md
@@ -1,6 +1,6 @@
-Frequently asked questions
+## Frequently asked questions
 
-## How does Physcraper handle paralogs?
+### How does Physcraper handle paralogs?
 
 Physcraper allows multiple tips to be mapped to the same taxon, with an added unique identifier that allows linking the sequence back to the original one on GenBank.
 This means that newly added homologous DNA sequences can include both orthologs
@@ -9,26 +9,26 @@ While we expect that most curated alignments are composed of ortholog sequences,
 Users should check the output phylogenetic trees to detect paralogs and filter them
 appropriately if needed.
 
-## I have to learn to use OpenTree to use Physcraper, is the learning curve worth it?
+### I have to learn to use OpenTree to use Physcraper, *is the learning curve worth it*?
 
 We think that this decision depends on the goals of the user.
 
 To our knowledge, existing tools that automatize dataset construction for phylogenetics focus on assembling an alignment _de novo_ and/or mining and filtering homolog sequences.
-The main goal of Physcraper is to construct upon the knowledge contained in previous expertly-curated and peer-reviwed homology hypothesis (alignments), as well as establish interoperability between biological databases and phylogenetic knowledge. To achieve that, understanding the OpenTree Taxonomy and associated tools is key.
+The main goal of Physcraper is to construct upon the knowledge contained in existing expertly-curated and peer-reviewed homology hypothesis (alignments), as well as establish a framework for interoperability between biological databases and phylogenetic knowledge. To achieve that, understanding the OpenTree Taxonomy and associated tools is key.
 
 In this sense Physcraper offers the unique advantage to automatically connect a phylogenetic tree to other databases and services, including alignment databases such as TreeBASE, and the conflicting service provided by OpenTree, which permits to rapidly identify regions in the tree that:
 - have been enriched in a tree with phylogenetic information,
 - are coherent with other phylogenetic estimates, as well as
 - conflict with other phylogenetic estimates.
--
-Users can update phylogenetic trees using any other existing tool, and if they want to connect them to other biological databases, particularly in a reproducible way, matching their taxon names to the OpenTree Taxonomy allows them to do this programmatically, instead of by hand.
+
+Users can update phylogenetic trees using any other existing tool. If they want to connect phylogenetic data to other biological databases, particularly in a reproducible way, matching their taxon names to the OpenTree Taxonomy allows them to do this programmatically and automatically, instead of by hand.
 
 For this goal, having a minimum familiarity with the OpenTree tools is needed.
 
 We realize that this might initially discourage some users, but we believe that the benefits brought by connecting taxonomic data with the OpenTree services will encourage users to familiarize with the OpenTree services, and to adopt the use of Physcraper.
 
 
-## How does Physcraper handle polytomies in a starting trees?
+### How does Physcraper handle polytomies in a starting tree?
 
 The Physcraper starting tree is a phylogeny whose tip labels must have been standardized to the OpenTree Taxonomy (as described in the Introduction section:
 [Mapping names to taxa](https://physcraper.readthedocs.io/en/latest/quick-start.html#updating-your-own-tree-and-alignment)).
@@ -54,7 +54,7 @@ around the early diverging branches,
 the automatic rooting is problematic and can have multiple solutions.
 
 
-## How does Physcraper use the starting alignment?
+### How does Physcraper use the starting alignment?
 
 Physcraper uses all unique DNA sequences in the input alignment to mine a genetic database using the BLAST algorithm, with the goal of increasing the
 lineage sampling of the alignment within a given biological group.
