+++ docs/mds/changelog.md
@@ -1,12 +1,19 @@
 All notable changes to the Physcraper project are documented here.
 
+## `next`
+
+- progress bars
+- aminoacid alignments
+- cleaner screen print
+
 ## `0.6.0`
 
 - add citation to README
 - add links to installation instructions and demonstration usage (documentation and examples) to README
 - Pylinting code
 - Multilocus script
-- Fix supertreebase error
+- Trapping supertreebase error
+- Trapping DendroPy error when reading in alignments with alternative states
 
 ## `0.5.0`
 
