+++ docs/mds/intro-requirements.md
@@ -14,7 +14,7 @@ This software is only needed if using a local genetic database.
 Note that BLAST + is automatically installed when [installing Physcraper using Anaconda](https://physcraper.readthedocs.io/en/stable/install.html#anaconda-virtual-environment). <br>
 *Camacho, C., Coulouris, G., Avagyan, V. et al. BLAST+: architecture and applications. BMC Bioinformatics 10, 421 (2009).* doi: [10.1186/1471-2105-10-421](https://doi.org/10.1186/1471-2105-10-421)
 
-
+<br>
 Physcraper relies on the following Python packages that are <b>automatically</b>
  installed:
 
