### 2021-08-11 #26 @LunaSare
No worries @leomrtns. Thanks for the comment!

Oh, gotcha! So, there's probably several things going on there.

First, the OpenTree synthetic tree for prokaryotes in general, such as _Salmonella_, is not well resolved, so Physcraper probably struggles more there.
The example alignment you share has more than 500 sequences, so it will probably be a long run.
I started running  your example (started 17:03 Tuesday August 10th), I will update once it is done!

In the [implementation](https://physcraper.readthedocs.io/en/stable/implementation.html) section of the documentation we mention that BLAST is implemented as an all-against-all BLAST.
We certainly agree that a progress bar is always a nice piece of information for users, and it was also suggested by @bpucker, so we will be adding it in a following release of Physcraper. We will also be more verbose on what software is being run at any time, as well as having a somewhat cleaner output to screen, so it is easier for users to know what is going on behind the scenes.

Regarding homology, Physcraper allows multiple tips to be mapped to the same taxon, with an added unique identifier that allows linking the sequence back to the original one on GenBank. This means that after the BLAST run, Physcraper can actually recover both paralogs and orthologs. We prompt the users to check and curate output phylogenetic trees, to filter potential paralogs.
If the original alignment and tree contain both orthologs and paralogs (such as the example tree you shared), then Physcraper will recover both types of homologs for sure.

We clarify more throughout the text in the [Readme](https://github.com/McTavishLab/physcraper/tree/main#readme), [intro page of the documentation](https://physcraper.readthedocs.io/en/stable/index.html), and [description](https://physcraper.readthedocs.io/en/stable/description.html) and we elaborated on that subject in the [faq section](https://physcraper.readthedocs.io/en/stable/faq.html)

We really appreciate the thorough review @leomrtns, and @NickleDave follow up! it was most helpful!
We address all points form the review below:

> ## Package Review
> * [x]  As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).
>
> #### Documentation
> The package includes all the following forms of documentation:
>
> * [x]  **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
> * [x]  **Installation instructions:** for the development version of package and any non-standard dependencies in README
> * [x]  **Vignette(s)** demonstrating major functionality that runs successfully locally
> * [x]  **Function Documentation:** for all user-facing functions
> * [ ]  **Examples** for all user-facing functions

Examples for all command line user-facing functions can be found in the [Start section](https://physcraper.readthedocs.io/en/latest/find_trees.html) and the [Run section](https://physcraper.readthedocs.io/en/latest/physcraper_run.html) of Physcraper's documentation at readthedocs.

> * [x]  **Community guidelines** including contribution guidelines in the README or CONTRIBUTING.
> * [x]  **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.
>
> Readme requirements
> The package meets the readme requirements below:
>
> * [x]  Package has a README.md file in the root directory.
>
> The README should include, from top to bottom:
>
> * [x]  The package name
> * [x]  Badges for continuous integration and test coverage, a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the badge for pyOpenSci peer-review will be provided upon acceptance.)
> * [x]  Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
> * [x]  Installation instructions
> * [x]  Any additional setup required (authentication tokens, etc)
> * [ ]  Brief demonstration usage

A link to [a set of demonstration usage examples](https://physcraper.readthedocs.io/en/latest/examples.html) in Physcraper's documentation at readthedocs is provided in the [README](https://github.com/McTavishLab/physcraper/tree/main#readme), as "real life examples".

> * [x]  Direction to more detailed documentation (e.g. your documentation files or website).
> * [ ]  If applicable, how the package compares to other similar packages and/or how it relates to other packages

We address this below.

> * [x]  Citation information
>
> #### Usability
> Reviewers are encouraged to submit suggestions (or pull requests) that will improve the usability of the package as a whole.
> Package structure should follow general community best-practices. In general please consider:
>
> * [x]  The documentation is easy to find and understand
> * [ ]  The need for the package is clear

We address this in the [Readme](https://github.com/McTavishLab/physcraper/tree/main#readme), [intro page of the documentation](https://physcraper.readthedocs.io/en/stable/index.html), and [description](https://physcraper.readthedocs.io/en/stable/description.html).

> * [x]  All functions have documentation and associated examples for use
>
> #### Functionality
> * [x]  **Installation:** Installation succeeds as documented.
> * [ ]  **Functionality:** Any functional claims of the software been confirmed.
> * [ ]  **Performance:** Any performance claims of the software been confirmed.
> * [ ]  **Automated tests:** Tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.

We provide examples and test runs [here](https://physcraper.readthedocs.io/en/clean-docs/install.html#checking-installation-success-on-remote-searches).

> * [x]  **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.
> * [x]  **Packaging guidelines**: The package conforms to the pyOpenSci [packaging guidelines](https://www.pyopensci.org/dev_guide/authoring/overview.html).
>
> Estimated hours spent reviewing:
> 8
>
> #### Review Comments
> I will focus my comments from the point of view of a nonspecialist end user. I had no problem installing the software, however I could not complete the analysis since running from the OToL took too long, and my trees did not seem appropriate (several potentially from same species, without easy taxonomic mapping). Please keep in mind that most (all) my comments can be addressed by changing the documentation, since they may reflect a misunderstanding on my part.
>

Thank you! We address this in a follow up comment.

> * My overall impression is that the requested familiarity with open ToL, treeBASE, and TNRS might hinder a broader  usability of the tool. If the user is unfamiliar with these, they will really consider the learning curve in order to  decide if using this tool is worth; if they are familiar, they are probably doing it without `physcraper` already.  My suggestion is that you must show how the tool can be useful even if one is does not want to familiarise with the above mentioned tools.

Thank you for your comment! Physcraper's main goal differs from most existing tools to update phylogenetic inferences. We more on the [FAQs section](https://physcraper.readthedocs.io/en/clean-docs/faq.html) of Physcraper's documentation at readthedocs. Solved here https://github.com/McTavishLab/physcraper/issues/165

> * In this sense I see several alternative tools, that although doing phylogenetics from scratch also returns an 'updated' phylogeny (in less than 12 hours perhaps?):
>
>   * https://github.com/AstrobioMike/GToTree
>   * https://github.com/ropensci/phylotaR
>   * https://github.com/dongzhang0725/PhyloSuite
>   * https://lmap-s.sourceforge.io/
>   * https://github.com/dportik/SuperCRUNCH
>   * https://github.com/rdcrawford/cognac
>   * https://github.com/yangfangs/physpetools

There are indeed many tools available that are designed to automate different elements of phylogenetic analyses.
Some are designed to automate the process of gathering a DNA dataset, like phylotaR or superCRUNCH, but none use the information contained in existing homology hypotheses.
Users should be able to choose Physcraper or any other tool accordingly to their goals. We elaborate more on this in the [FAQ section](https://physcraper.readthedocs.io/en/clean-docs/faq.html). I am also working on a post that compares available tools, but that is work in progress https://github.com/McTavishLab/physcraper_ms/blob/master/99-tools-review.md

> * Alternatively to the discussion above, it is fair and absolutely fine to describe `physcraper` as a tool within the open ToL ecosystem instead of a tool for the nonspecialist. From my understanding physcraper is great at automating OToL tasks   (as in continuous updating), but a disclaimer or an observation at the begining (of the documentation and README) would be necessary.   The objective here is to manage user expectation, and I will describe more below about hidden assumptions and knowledge.

Yes! We are more explicit now on the connection of Physcraper to OpenTree in the [Readme](https://github.com/McTavishLab/physcraper/tree/main#readme), [intro page of the documentation](https://physcraper.readthedocs.io/en/stable/index.html), and [description](https://physcraper.readthedocs.io/en/stable/description.html).

> * When using the term "homologous", maybe the authors mean "orthologous"? I understand this is a single-alignment tool but since you rely on taxonomic classifications I think that the tool is not ready for paralogous sequences.   In  other words, if you allow several tips to map to the same taxon (with a further indication of gene/locus name perhaps)  then you indeed accept homologous sequences (orthologs **and** paralogs). The documentation mentions "single locus alignment", and thus you should prefer the term ortholog/orthologous whenever possible (or mention that this   distinction is deferred to the study depositing the data, maybe? I am unaware of whole gene family data sets on OToL).   If `physcraper` does indeed handle all homologous sequences (as suggested in https://physcraper.readthedocs.io/en/latest/how_to_start.html)   then you need to explain how it reconciles the gene (locus) tree with the species tree, or how it decides between paralog alternatives.   Still I think it's safer to replace "homologous" by "orthologous" therefore implying that the responsability for  disambiguation is on the original publications, and not `physcraper`.

Thanks for pointing this out! We addressed this above and in the [FAQ section](https://physcraper.readthedocs.io/en/clean-docs/faq.html)

> * The need for mapping the tree tip names to a broader taxonomy may be a serious impediment to the nonspecialist,  since in many analyses the taxonomical classification is a conclusion, and not an assumption of the study.  In which case a tool capable of updating my tree and alignment with data from known studies would be most welcome! Furthermore filling the mapping table by hand is impractical except for toy examples (with less than a few dozen taxa).  (https://physcraper.readthedocs.io/en/latest/how_to_start.html#updating-your-own-tree-and-alignment). I've seen in another piece of documentation that this might not be necessary, I can't find it now... So could this be made clearer?

We certainly agree! Physcraper updates the tree and alignment with data that has been deposited in GenBank from known studies. The users and non-specialists do not have to deal with taxonomy at all. The taxonomic mapping is done automatically using OpenTree tools [here](https://github.com/OpenTreeOfLife/germinator/wiki/Open-Tree-of-Life-Web-APIs#taxonomy-methods) that have been integrated to Phsycraper's workflow. The main goal of this is to document all steps of the analysis, so it can be reproduced and checked by anyone.
In the link you [shared](https://physcraper.readthedocs.io/en/latest/how_to_start.html#updating-your-own-tree-and-alignment) we described the option to match taxonomies using the graphical interface of OpenTree. We do not suggest doing it by hand, that would indeed be very time consuming and non-reproducible, although most researchers still do it like that! We added some more words to hopefully make it clearer.
[This](https://physcraper.readthedocs.io/en/clean-docs/quick-start.html#updating-a-tree-from-open-tree-of-life) is the section where we explain how to use trees that have already been matched to the taxonomy, so no matching is required by the user.

> * In the [Results section of the documentation](https://physcraper.readthedocs.io/en/latest/data_exploration.html) it seems I need to know python (and use `physcraper`'s module) to be able to relabel the tree. Which is not exactly user-friendly or useful to a general, nonspecialist audience. I see no reason not to save several files, one of them with easily interpretable names.

Well, there are many alternative tip labels that can be used to write out the updated tree (including those from other taxonomies, such as NCBI or GBIF). For now, we decided to just print the tree with taxon labels standardized to the OpenTree reference taxonomy.
We think it is a good idea to write down more output trees, however, it can be potentially more confusing for users. We will also work on the command line code to facilitate changing tip labels in the updated trees for users that are not familiar with Python. We created an issue for enhancement https://github.com/McTavishLab/physcraper/issues/124

> The same goes for the [scary mention of _rewriting_ of input files](https://physcraper.readthedocs.io/en/stable/data_exploration.html#an-overview-of-the-physcraper-results-files) —I am sure you mean that you copy the input files with modification somewhere else?

Yes! We do mean that-- we changed the wording in [that section](https://physcraper.readthedocs.io/en/clean-docs/results.html#files-generated-by-a-physcraper-run) from "rewriting" to simply "writing"

>
> **minor comments:**
>
> * How does it handle polytomies, are all trees in OToL fully resolved? Furthermore, is it possible to know when there are conflicting trees across studies? This is an OToL question actually, but still it would be good to be adressed in the documentation here.

Thanks for the question!
We elaborated on the subject of polytomies on the [FAQ section](https://physcraper.readthedocs.io/en/clean-docs/faq.html) of Physcraper's documentation at readthedocs. More info here https://github.com/McTavishLab/physcraper/issues/171

> * Some `readthedocs` internal links are not working (for instance the text mentions https://physcraper.readthedocs.io/en/latest/data_exploration.html#Relabeling but the correct link is  https://physcraper.readthedocs.io/en/latest/data_exploration.html#id1 ). Furthermore in  https://physcraper.readthedocs.io/en/latest/methods_extended.html, `Fig. @ref(fig:framework)` in the documentation is missing

We checked all links in Physcraper's Documentation, including the ones pointed out by the reviewer, to make sure none are broken.
We also fixed the link to the figure in the [Implementation section](https://physcraper.readthedocs.io/en/clean-docs/methods_extended.html).
