+++ docs/mds/description.md
@@ -10,7 +10,7 @@ If you have access to a single gene or multilocus nucleotide alignment, and a ph
 adding nucleotide sequences of new lineage samples into your tree by using [Open Tree of Life](#the-open-tree-of-life) tools to reconcile Taxonomy, and the [BLAST](https://blast.ncbi.nlm.nih.gov/Blast.cgi) algorithm to search for loci in the [GenBank](https://www.ncbi.nlm.nih.gov/genbank/statistics/) genetic database that are likely to be
 locally similar to sequences in the initial DNA alignment.
 
-By using a starting alignment and tree, Physcraper takes advantage of DNA loci alignments as  homology hypotheses (ideally orthology, see [FAQs](https://physcraper.readthedocs.io/en/stable/faq.html)) that
+By using a starting alignment and tree, Physcraper takes advantage of DNA loci alignments as  homology hypotheses (ideally orthology, see [FAQs](https://physcraper.readthedocs.io/en/latest/faq.html)) that
 previous researchers have assessed, curated, and deemed appropriate for the phylogenetic scope.
 The sequences added during a BLAST search are limited either to a user specified taxon or
 monophyletic group, or within the taxonomic scope of the ingroup of the starting tree.
