+++ docs/mds/intro.md
@@ -3,5 +3,3 @@
 Use a phylogenetic tree and a DNA alignment to automatically find and add nucleotide sequences from a genetic database, to reproducibly improve and advance phylogenetic knowledge within a biological group.
 
 Physcraper relies on [taxonomic](https://tree.opentreeoflife.org/about/taxonomy-version/ott3.3) and [phylogenetic](https://github.com/OpenTreeOfLife/phylesystem-1) resources and [programmatic tools](https://github.com/OpenTreeOfLife/germinator/wiki/Open-Tree-of-Life-Web-APIs) from the [Open Tree of Life](https://tree.opentreeoflife.org/opentree/argus/opentree12.3@ott93302) project.
-
-Physcraper also leverages on programmatic tools from the [TreeBASE](https://treebase.org/treebase-web/urlAPI.html) project and [NCBI](https://www.ncbi.nlm.nih.gov/home/develop/api/), as well as multiple software projects listed as [requirements](#Requirements) below, to create an automatic and reproducible workflow for phylogenetics.
