+++ docs/source/introduction.rst
@@ -1,6 +1,0 @@
-Introduction to Physcraper
-==========================
-
-.. mdinclude:: ../mds/description.md
-
-.. mdinclude:: ../mds/quick-start.md
