+++ README.Rmd
@@ -30,6 +30,8 @@ IS_README <- TRUE
 
 ```
 
+Physcraper also leverages on programmatic tools from the [TreeBASE](https://treebase.org/treebase-web/urlAPI.html) project and [NCBI](https://www.ncbi.nlm.nih.gov/home/develop/api/), as well as multiple software projects listed as [requirements](#requirements) below, to create an automatic and reproducible workflow for phylogenetics.
+
 You are now on the code repository. Please refer to Physcraper’s
 [documentation website](https://physcraper.readthedocs.io/en/latest/)
 for more details on:
