+++ README.md
@@ -34,7 +34,7 @@ project.
 Physcraper also leverages on programmatic tools from the
 [TreeBASE](https://treebase.org/treebase-web/urlAPI.html) project and
 [NCBI](https://www.ncbi.nlm.nih.gov/home/develop/api/), as well as
-multiple software projects listed as [requirements](#Requirements)
+multiple software projects listed as [requirements](#requirements)
 below, to create an automatic and reproducible workflow for
 phylogenetics.
 
