+++ docs/mds/intro-part2.md
@@ -1,0 +1,5 @@
+Physcraper also leverages on programmatic tools from the
+[TreeBASE](https://treebase.org/treebase-web/urlAPI.html) project and
+[NCBI](https://www.ncbi.nlm.nih.gov/home/develop/api/), as well as multiple software
+projects listed as [requirements](#id1)
+below, to create an automatic and reproducible workflow for phylogenetics.
