+++ physcraper/configobj.py
@@ -247,13 +247,13 @@ max_length = {maxlen}
             sys.stderr.write(
                 "Local Blast DB not found at {},\
                 please use a remote search, or update as described\
-                in 'https://physcraper.readthedocs.io/en/latest/install.html#local-databases'\n".format(self.blastdb))
+                in 'https://physcraper.readthedocs.io/en/main/install.html#local-databases'\n".format(self.blastdb))
             sys.exit()
         if not os.path.exists("{}/nt.23.nhr".format(self.blastdb)):
             sys.stderr.write("Errors with local Blast DB at {}, \
                 may be incomplete. please use a remote search, \
                 or update as described in \
-                'https://physcraper.readthedocs.io/en/latest/install.html#local-databases'\n".format(self.blastdb))
+                'https://physcraper.readthedocs.io/en/main/install.html#local-databases'\n".format(self.blastdb))
             sys.exit()
         else:
             download_date = os.path.getmtime("{}/nt.23.nhr".format(self.blastdb))
@@ -265,12 +265,12 @@ max_length = {maxlen}
                     You downloaded them {} days ago. \
                     Continuing, but perhaps use a remote search, \
                     or update as decribed in \
-                    'https://physcraper.readthedocs.io/en/latest/install.html#local-databases'\n".format(time_passed))
+                    'https://physcraper.readthedocs.io/en/main/install.html#local-databases'\n".format(time_passed))
         if not os.path.exists(self.ncbi_nodes):
             sys.stderr.write("NCBI taxonomy not found at {} - \
              To perform a local blast search, please update nodes and names.dmp, \
                 as described in \
-                'https://physcraper.readthedocs.io/en/latest/install.html#local-databases'\n".format(self.ncbi_nodes))
+                'https://physcraper.readthedocs.io/en/main/install.html#local-databases'\n".format(self.ncbi_nodes))
 
             sys.exit()
         else:
@@ -282,6 +282,6 @@ max_length = {maxlen}
                 sys.stderr.write("Your taxonomy databases from NCBI were dowloaded {} days ago. \
                     To perform a local blast search, please update nodes and names.dmp, \
                     as described in \
-                    'https://physcraper.readthedocs.io/en/latest/install.html#local-databases'\n".format(time_passed))
+                    'https://physcraper.readthedocs.io/en/main/install.html#local-databases'\n".format(time_passed))
         assert(shutil.which("blastn")), "blastn  not found in path"
         self.url_base = None
