### 2021-08-11 #26 @bpucker
Hi all,

Sorry for my late reply. @LunaSare and @snacktavish addressed all my points very well. I can understand that the integration of a progress bar is challenging and was postponed to a later release. Congratulations on this nice work!
