### 2021-09-13 #26 @leomrtns
Hello! Thanks for the detailed response @LunaSare, I believe all my comments have been successfully addressed. I particularly like the FAQ and the "use case examples". I am looking forward the comparison between tools, but I understand this is will be a major endeavor.
Congratulations!
