### 2021-09-14 #26 @lwasser
yahoo! pinging @NickleDave our fearless editor for this package! it looks from the initial submission like this is not slated to move on to JOSS at this point  so we may be able to wrap things up!
