### 2021-09-15 #26 @NickleDave
Thanks @lwasser !!! Yes looks like we can wrap things up.

And a big thank you @leomrtns and @bpucker for your reviews and very valuable input.

*And* thanks to @LunaSare and @snacktavish for responding so clearly that even a non-expert like me can follow what's going on.

Seems like the pyOpenSci review has been helpful for `physcraper`'s development, I hope you'll agree.
