### 2021-09-15 #26 @NickleDave
🎉 `physcraper` has been approved by pyOpenSci! Thank you @LunaSare @snacktavish for submitting `physcraper` and many thanks to @bpucker @leomrtns for reviewing! 😸

(I know I just said thanks in the last comment but it's the language in our template and I like the smiling cat emoji)

- [ ] [Activate Zenodo watching the repo if you haven't already done so.](https://guides.github.com/activities/citable-code/)
- [ ] Add the badge for pyOpenSci peer-review to the README.md of `physcraper`. In markdown, the badge would be:  `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/26)`
- [ ] Add `physcraper` to the pyOpenSci website. @LunaSare or @snacktavish , please open a pull request to update [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/packages.yml): to add your package and name to the list of contributors
- [ ] @snacktavish @LunaSare @bpucker @leomrtns if you have time and are open to being listed on our website, please add yourselves to [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/contributors.yml) via a pr so we can list you on our website as contributors!

All -- if you have any feedback for us about the review process please feel free to share it here. We are always looking to improve our process and our documentation in the [contributing-guide](https://www.pyopensci.org/contributing-guide). We have also been updating our documentation to improve the process so all feedback is appreciated!
