### 2021-09-15 #26 @NickleDave
Adding as a separate comment so I don't overwhelm anyone trying to read that checklist:

@LunaSare looks like you already did a `pyOpenSci review` release? https://github.com/McTavishLab/physcraper/releases/tag/v0.6.1  -- I think you could do a ["post-release"](https://snarky.ca/what-to-do-when-you-botch-a-release-on-pypi/) where you just add the badge to the README (sorry if I'm telling you something you already know).  I am pretty sure that a Zenodo DOI is still the mechanism we use to track completed reviews though -- @lwasser please correct me if I'm wrong. You could just do a post release with the 5 more commits you have now, assuming you're okay with that.  **Or** if you want to also have the badge show up on the README, e.g. on PyPI, you could first add it, and *then* do the release on GitHub, which will trigger the creation of a DOI through Zenodo. You'd then publish to PyPI and the `post` release would show the badge there as well. (Again sorry if I'm explaining something that's already obvious to you--I had to learn this stuff the hard way 😢 so I'm just trying to save you potential packaging pain (AKA PPP)).
