### 2021-09-15 #26 @lwasser
yay @NickleDave this is all awesome! and @LunaSare @snacktavish thank you for all of your work on this review and reviewers @bpucker @leomrtns thank you for your time in reviewing for us!

@NickleDave yes great catch. one post release even if it is a patch with the badge in the readme is appreciated. then we can use that specific zenodo archive in the records here as the version we "approved".
