### 2021-09-15 #26 @NickleDave
Ah right.

I might have made things extra confusing by bringing up PyPI.
Definitely no need from our end to have it on PyPI, as long as you create a DOI, e.g. with Zenodo by doing a GitHub release.

but, as long as I brought it up:
re: publishing to PyPI, I always go back to this tutorial
https://packaging.python.org/tutorials/packaging-projects/
which you will notice has moved away from the setup.py / requirements.txt approach it looks like you're using now
I use `poetry` that handles a lot of the grunt work.
Nice tutorial on packaging in general here that uses 'poetry': https://py-pkgs.org/
An alternative would be `flit`: https://github.com/takluyver/flit

But again I didn't mean to imply you need to deal with any of that to get the DOI so we can finish up review.
