### 2021-09-15 #26 @lwasser
@LunaSare are you interested in pushing to pypi? i think it is ideal and we could likely help you through it and take notes on any pain points. i have a setup for our packages where it gets pushed when there is a new release using github actions.
we could potentially help you get there as right now [pypi's version of your package ](https://pypi.org/project/Physcraper/) is behind what we accepted.

But first we would need to know who originally pushed this to pypi so we can get credentials figured out and such and help from there to ensure there is only ONE version of it on pypi!
