### 2021-09-16 #26 @snacktavish
We created a v1.0.0 release updated the PyPi  version!
https://pypi.org/project/Physcraper/

We also did these:
- [x]  Activate Zenodo watching the repo if you haven't already done so.
https://zenodo.org/record/5512990#.YUPLLntlBhE
- [x] Add the badge for pyOpenSci peer-review to the README.md of physcraper. In markdown, the badge would be: [![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/26)
