### 2021-09-16 #26 @LunaSare
Thanks so much @lwasser and @NickleDave for the helpful guidance!
@snacktavish did the previous release of Physcraper to pypi. So we gave it a go and we were able to successfully submit the latest version of [Physcraper to pypi](https://pypi.org/project/Physcraper/) today, thanks to @snacktavish's  amazing coding skills!

@NickleDave, we ended up releasing [v1.0.0 of Physcraper](https://zenodo.org/badge/latestdoi/41294748). Thanks for teaching us about [post-releases](https://snarky.ca/what-to-do-when-you-botch-a-release-on-pypi/), I think it will save us some headaches in the future! Also, the [guide to make code citable](https://guides.github.com/activities/citable-code/) was super useful!
