+++ setup.py
@@ -2,8 +2,8 @@
 
 from setuptools import setup
 
-setup(name='Physcraper',
-      version='1.0.0',
+setup(name='Physcraper'
+      version='1.0.1',
       description='Physcraper',
       author='Emily Jane McTavish',
       author_email='ejmctavish@gmail.com',
