### 2022-09-28 #26 @lwasser
hi 👋  @snacktavish  and @bpucker ! just a friendly reminder to take 5-10 minutes to fill out our survey . We really appreciate it. Emily, if you can have all of the maintainers of your package fill it out that would be great. We will stay in touch with package authors - checking in 1-2 times a year to ensure the package continues to be maintained and see better understand challenges that you face. It's important for us to be able to stay in touch with the package maintenance crew! We also want to list everyone on our website as maintainers.  Thank you in advance for helping us by filling out the survey!! 🙌

[🔗 HERE IS THE SURVEY LINK 🔗 ](https://forms.gle/ZZXcD1hr18tYmY2p6)
