### 2022-10-19 #26 @lwasser
hey @snacktavish ! i'm just checking in. we do need to ensure that packages are still being maintained that have been accepted to pyopensci. I see that hte last commit to physcraper was in May of this year which is great! Is this package still being maintained?

And can you kindly also fill out our maintainer survey. we need to be able to keep in touch with our maintainers to ensure packages remain supported in our ecosystem.
Many thanks for doing this or if you have questions, feel free to email me at leah at pyopensci.org

All the best!
