### 2022-10-19 #26 @lwasser
Awesome. Thanks so much for the speedy replies today!

Leah

Leah A Wasser
Executive Director, pyOpenSci
@leahawasser
@pyopensci
***@***.***

> On Oct 19, 2022, at 2:48 PM, Emily Jane McTavish ***@***.***> wrote:
>
>
> We are!
>
