### 2022-10-24 #26 @lwasser
hey there @snacktavish @LunaSare i have another followup question. we are trying to grasp who "owns" or runs the maintainer team here vs who has contributed etc

I have another name on my list [Martha Kandziora](https://github.com/mkandziora) and martha seemed to contribute in 2019 but i don't see any recent commits from them. Is it safe to say that Martha is no longer a maintainer? I am just trying to ensure our data are up to date. Many thanks
